% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-addins.R
\name{addin_copyModule}
\alias{addin_copyModule}
\title{RStudio addin to copy an existing module}
\usage{
addin_copyModule()
}
\description{
RStudio addin to copy an existing module
}
\author{
Alex Chubaty
}
