% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module-dependencies.R
\docType{methods}
\name{addSimDep}
\alias{addSimDep}
\alias{addSimDep,ANY,moduleDeps-method}
\title{Add simulation dependencies}
\usage{
addSimDep(sim, x)

\S4method{addSimDep}{ANY,moduleDeps}(sim, x)
}
\arguments{
\item{sim}{A \code{\link{simList}} object.}

\item{x}{A \code{\link{moduleDeps}} object.}
}
\description{
Adds a \code{moduleDeps} object to the simulation dependency list stored in \code{.spadesEnv}.
}
\examples{
\dontrun{
  addSimDep(sim, x)
}
}
\author{
Alex Chubaty
}
\seealso{
moduleDeps
}

