% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{.addSimDepends}
\alias{.addSimDepends}
\alias{.addSimDepends,simEnv,.moduleDeps-method}
\title{Add simulation dependencies}
\usage{
.addSimDepends(sim, x)

\S4method{.addSimDepends}{simEnv,.moduleDeps}(sim, x)
}
\arguments{
\item{sim}{A \code{simEnv} object.}

\item{x}{A named list containing the parameters used to construct a new
           \code{.moduleDeps} object.}
}
\value{
A \code{simEnv} object.
}
\description{
Internal function.
Adds a \code{.moduleDeps} object to the simulation dependency list.
}
\author{
Alex Chubaty
}

