% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{methods}
\name{.addTagsToOutput,simList-method}
\alias{.addTagsToOutput,simList-method}
\title{addTagsToOutput for simList class objects}
\usage{
\S4method{.addTagsToOutput}{simList}(object, outputObjects, FUN)
}
\arguments{
\item{object}{Any R object.}

\item{outputObjects}{Optional character vector indicating which objects to
return. This is only relevant for \code{simList} objects}

\item{FUN}{A function}
}
\description{
See \code{\link[reproducible]{.addTagsToOutput}}.
}
\seealso{
\code{\link[reproducible]{.addTagsToOutput}}
}
