% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{arrangement-class}
\alias{arrangement-class}
\title{The \code{arrangement} class}
\description{
This class contains the plotting arrangement information.
}
\section{Slots}{

\describe{
\item{\code{rows}}{numeric. Number of rows in the arrangement.}

\item{\code{columns}}{numeric. Number of columns in the arragnement.}

\item{\code{actual.ratio}}{numeric. Ratio of columns to rows}

\item{\code{ds.dimensionRatio}}{numeric. Ratio of the device size to the ratio of the
extents}

\item{\code{ds}}{numeric of length 2. The dimensions of the plotting window in inches}

\item{\code{objects}}{list of length number of spatial objects. Each list has a character vector
of the layer names in each of those}

\item{\code{isRaster}}{logical vector, indicating whether each object is a Raster* object}

\item{\code{names}}{character vector. The names of the layers in the plot}

\item{\code{extents}}{list of class Extent objects. These are needed to calculate the
\code{ds.dimensionRatio}, which is used to scale the Spatial objects correctly}

\item{\code{layout}}{list of length 2, with width and height measurements for layout.}
}}
\author{
Eliot McIntire
}

