% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/environment.R
\docType{methods}
\name{assignSpaDES}
\alias{assignSpaDES}
\alias{assignSpaDES,character-method}
\title{Assign to the internal SpaDES environment.}
\usage{
assignSpaDES(x, value, ...)

\S4method{assignSpaDES}{character}(x, value, ...)
}
\arguments{
\item{x}{a variable name, given as a character string.
           No coercion is done, and the first element of a character vector
           of length greater than one will be used, with a warning.}

\item{value}{The object to assign.}

\item{...}{Additional arguments to pass to \code{assign}.}
}
\description{
Internal function. Simple wrapper for \code{\link{assign}}.
}
\author{
Alex Chubaty
}

