% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\docType{methods}
\name{.assignSpaDES}
\alias{.assignSpaDES}
\alias{.assignSpaDES,character,ANY-method}
\alias{.assignSpaDES,character,missing-method}
\title{Assign to the internal SpaDES environment.}
\usage{
.assignSpaDES(x, value, ...)

\S4method{.assignSpaDES}{character,ANY}(x, value, ...)

\S4method{.assignSpaDES}{character,missing}(x, value, ...)
}
\description{
Internal function. Simple wrapper for \code{\link{assign}}.
}
\author{
Alex Chubaty
}

