% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{methods}
\name{.cacheMessage,simList-method}
\alias{.cacheMessage,simList-method}
\title{cacheMessage for simList class objects}
\usage{
\S4method{.cacheMessage}{simList}(object, functionName)
}
\arguments{
\item{object}{Any R object.}

\item{functionName}{A character string indicating the function name}
}
\description{
See \code{\link[reproducible]{.cacheMessage}}.
}
\seealso{
\code{\link[reproducible]{.cacheMessage}}
}
