% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{methods}
\name{.checkCacheRepo,list-method}
\alias{.checkCacheRepo,list-method}
\title{checkCacheRepo for simList class objects}
\usage{
\S4method{.checkCacheRepo}{list}(object)
}
\arguments{
\item{object}{A list of all elements in the call to Cache}
}
\description{
See \code{\link[reproducible]{.checkCacheRepo}}.
}
\seealso{
\code{\link[reproducible]{.checkCacheRepo}}
}
