% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-repository.R
\docType{methods}
\name{checkModule}
\alias{checkModule}
\alias{checkModule,character,character-method}
\alias{checkModule,character,missing-method}
\title{Check for the existence of a remote module}
\usage{
checkModule(name, repo)

\S4method{checkModule}{character,character}(name, repo)

\S4method{checkModule}{character,missing}(name)
}
\arguments{
\item{name}{Character string giving the module name.}

\item{repo}{GitHub repository name.
Default is \code{"PredictiveEcology/SpaDES-modules"}, which is
specified by the global option \code{spades.moduleRepo}.}
}
\description{
Looks in the remote \code{repo} for a module named \code{name}.
}
\author{
Eliot McIntire and Alex Chubaty
}
