% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{checkParams}
\alias{checkParams}
\alias{checkParams,simList,list,list,character-method}
\title{Check use and existence of params passed to simulation.}
\usage{
checkParams(sim, defaultModules, defaultParams, path, ...)

\S4method{checkParams}{simList,list,list,character}(sim, defaultModules,
  defaultParams, path, ...)
}
\arguments{
\item{sim}{A simList simulation object.}

\item{defaultModules}{List of default modules.}

\item{defaultParams}{List of default parameters.}

\item{path}{The location of the modules' source files.}

\item{...}{Additional arguments. Not implemented.}
}
\value{
Invisibly return \code{TRUE} indicating object exists; code{FALSE} if not.
         Sensible warning messages are be produced identifying missing params.
}
\description{
Checks that all parameters passed are used in a module,
and that all parameters used in a module are passed.
}
\author{
Alex Chubaty
}

