% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{checkPath}
\alias{checkPath}
\alias{checkPath,NULL,ANY-method}
\alias{checkPath,character,logical-method}
\alias{checkPath,character,missing-method}
\title{Check filepath.}
\usage{
checkPath(path, create)

\S4method{checkPath}{character,logical}(path, create)

\S4method{checkPath}{character,missing}(path)

\S4method{checkPath}{`NULL`,ANY}(path)
}
\arguments{
\item{path}{A character string corresponding to a filepath.}

\item{create}{A logical indicating whether the path should
be created if it doesn't exist. Default is \code{FALSE}.}
}
\value{
Character string denoting the cleaned up filepath.
}
\description{
Checks the specified filepath for formatting consistencies,
such as trailing slashes, etc.
}
\seealso{
\code{\link{file.exists}}, \code{\link{dir.create}}.
}

