% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{do.event.checkpoint}
\alias{checkpoint.load}
\alias{checkpoint.save}
\alias{do.event.checkpoint}
\title{Simulation checkpoints.}
\usage{
do.event.checkpoint(sim, event.time, event.type, debug = FALSE)

checkpoint.load(file = "checkpoint.RData")

checkpoint.save(file = "checkpoint.RData")
}
\arguments{
\item{sim}{A \code{SimList} simulation object.}

\item{event.time}{A numeric specifying the time of the next event.}

\item{event.type}{A character string specifying the type of event:
                     one of either \code{"init"}, \code{"load"}, or \code{"save"}.}

\item{debug}{Optional logical flag determines whether sim debug info
                     will be printed (default \code{debug=FALSE}.}
}
\value{
Returns the modified \code{SimList} object.
}
\description{
Save and reload the current state of the simulation,
including the state of the random number generator,
by scheduling checkpoint events.
}
\details{
\code{\link{checkpoint.load}} and \code{\link{checkpoint.save}} code from:
https://raw.githubusercontent.com/achubaty/r-tools/master/checkpoint.R

RNG save code adapted from:
http://www.cookbook-r.com/Numbers/Saving_the_state_of_the_random_number_generator/
https://stackoverflow.com/questions/13997444/
}
\seealso{
\code{\link{.Random.seed}}.
}

