% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cir}
\alias{cir}
\title{Identify pixels in a circle around a SpatialPoints*Named object.}
\usage{
cir(spatialPoints, radii, raster)
}
\arguments{
\item{spatialPoints}{SpatialPoints*Named object around which to make circles .}

\item{radii}{vector of radii that has same length as spatialPoints}

\item{raster}{Raster on which the circles are built.}

\item{scaleRaster}{Description of this.}
}
\value{
A list of data.frames with x and y coordinates of each
unique pixel of the circle around each individual.
}
\description{
identify the pixels and coordinates that are at
 a (set of) buffer distance(s) of the SpatialPoints*Named objects. This can be used
 for agents.
}

