% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc-methods.R
\docType{methods}
\name{classFilter}
\alias{classFilter}
\alias{classFilter,character,character,character-method}
\alias{classFilter,character,character,missing-method}
\title{Filter objects by class}
\usage{
classFilter(x, include, exclude)

\S4method{classFilter}{character,character,character}(x, include, exclude)

\S4method{classFilter}{character,character,missing}(x, include)
}
\arguments{
\item{x}{Character vector of object names to filter, possibly from \code{ls}.}

\item{include}{Class(es) to include, as a character vector.}

\item{exclude}{Optional class(es) to exclude, as a character vector.}
}
\value{
Vector of object names matching the class filter.
}
\description{
Based on \url{http://stackoverflow.com/a/5158978/1380598}.
}
\note{
Currently only searches for objects in the global enviroment.

\code{\link{inherits}} is used internally to check the object class,
which can, in some cases, return results inconsistent with \code{is}.
See \url{http://stackoverflow.com/a/27923346/1380598}.
These (known) cases are checked manually and corrected.
}
\examples{
a <- list(1:10)     # class `list`
b <- letters        # class `character`
d <- runif(10)      # class `numeric`
f <- sample(1L:10L) # class `numeric`, `integer`
g <- lm( jitter(d) ~ d ) # class `lm`
h <- glm( jitter(d) ~ d ) # class `lm`, `glm`
classFilter(ls(), include=c("character", "list"))
classFilter(ls(), include="numeric")
classFilter(ls(), include="numeric", exclude="integer")
classFilter(ls(), include="lm")
classFilter(ls(), include="lm", exclude="glm")
rm(a, b, d, f, g, h)
}
\author{
Alex Chubaty
}

