% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{methods}
\name{clearStubArtifacts}
\alias{clearStubArtifacts}
\alias{clearStubArtifacts,ANY-method}
\title{Clear erroneous archivist artifacts}
\usage{
clearStubArtifacts(repoDir = NULL)

\S4method{clearStubArtifacts}{ANY}(repoDir = NULL)
}
\arguments{
\item{repoDir}{A character denoting an existing directory of the Repository for
which metadata will be returned. If it is set to NULL (by default), it
will use the repoDir specified in \code{archivist::setLocalRepo}.}
}
\value{
Done for its side effect on the repoDir
}
\description{
When an archive object is being saved, if this is occurring at the same time
as another process doing the same thing, a stub of a artifact occurs. This
function will clear those stubs.
}
\author{
Eliot McIntire
}
