% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{clickExtent}
\alias{clickExtent}
\title{Click to draw an Extent}
\usage{
clickExtent(devNum = NULL, plot.it = TRUE)
}
\arguments{
\item{devNum}{The device number for the new plot to be plotted on}

\item{plot.it}{The device number for the new plot to be plotted on}
}
\description{
Currently this does not work since update to spades v0.1.0.9000
}

