% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{.cols}
\alias{.cols}
\title{A short selection of colour palettes that can be use}
\format{\preformatted{List of 10
 $ transparentGrey: chr [1:9] "#00000000" "#25252566" "#52525266" "#73737366" ...
 $ grey           : chr [1:9] "#FFFFFF" "#F0F0F0" "#D9D9D9" "#BDBDBD" ...
 $ spectral       : chr [1:8] "#D53E4F" "#F46D43" "#FDAE61" "#FEE08B" ...
 $ terrain        : chr [1:100] "#00A600FF" "#03A700FF" "#07A800FF" "#0AAA00FF" ...
 $ heat           : chr [1:10] "#FF0000FF" "#FF2400FF" "#FF4900FF" "#FF6D00FF" ...
 $ topo           : chr [1:10] "#4C00FFFF" "#0019FFFF" "#0080FFFF" "#00E5FFFF" ...
 $ blueGreen      : chr [1:9] "#F7FCFD" "#E5F5F9" "#CCECE6" "#99D8C9" ...
 $ greens         : chr [1:9] "#F7FCF5" "#E5F5E0" "#C7E9C0" "#A1D99B" ...
 $ yellowBrown    : chr [1:9] "#FFFFE5" "#FFF7BC" "#FEE391" "#FEC44F" ...
 $ discrete1      : chr [1:8] "#8C510A" "#BF812D" "#DFC27D" "#F6E8C3" ...
}}
\usage{
.cols
}
\description{
Colour number 1 shows use of transparency
}
\keyword{datasets}

