% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\docType{methods}
\name{.convertSpatialToPlotGrob}
\alias{.convertSpatialToPlotGrob}
\alias{.convertSpatialToPlotGrob,spatialObjects-method}
\title{Internal function to convert a SpatialObject to a Graphical Object}
\usage{
.convertSpatialToPlotGrob(grobToPlot, sGrob, takeFromPlotObj, arr, newArr,
  spadesGrobCounter, subPlots, cols)

\S4method{.convertSpatialToPlotGrob}{spatialObjects}(grobToPlot, sGrob,
  takeFromPlotObj, arr, newArr, spadesGrobCounter, subPlots, cols)
}
\arguments{
\item{grobToPlot}{Graphical object to plot}

\item{sGrob}{spadesGrob object}

\item{takeFromPlotObj}{Logical indicating whether data for grobToPlot should be found in
current call to Plot or from disk}

\item{arr}{An arrangement object}

\item{newArr}{Logical, whether there is a new arrangement happening}

\item{spadesGrobCounter}{Numeric. A counter. No meaning outside Plot function.}

\item{subPlots}{Character. Name of plot area.}

\item{cols}{Color vector.}
}
\description{
Not meant to be used by user.
}
\author{
Eliot McIntire
}

