% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{coordinates,pointAgent-method}
\alias{coordinates,pointAgent-method}
\title{get coordinates of a pointAgent}
\usage{
\S4method{coordinates}{pointAgent}(obj, ...)
}
\arguments{
\item{obj}{ object deriving from class "Spatial" }

\item{...}{additional arguments that may be used by particular methods}
}
\description{
get coordinates of a pointAgent
}

