% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-template.R
\docType{methods}
\name{copyModule}
\alias{copyModule}
\alias{copyModule,character,character,character-method}
\alias{copyModule,character,character,missing-method}
\title{Create a copy of an existing module}
\usage{
copyModule(from, to, path, ...)

\S4method{copyModule}{character,character,character}(from, to, path, ...)

\S4method{copyModule}{character,character,missing}(from, to, path, ...)
}
\arguments{
\item{from}{The name of the module to copy.}

\item{to}{The name of the copy.}

\item{path}{The path to a local module directory. Defaults to the path set by
the \code{spades.modulePath} option. See \code{\link{setPaths}}.}

\item{...}{Additional arguments to \code{file.copy}, e.g., \code{overwrite = TRUE}.}
}
\value{
Invisible logical indicating success (\code{TRUE}) or failure (\code{FALSE}).
}
\description{
Create a copy of an existing module
}
\examples{
\dontrun{copyModule(from, to)}

}
\author{
Alex Chubaty
}
