% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{methods}
\name{.coreModules}
\alias{.coreModules}
\title{Named list of core \code{SpaDES} modules}
\usage{
.coreModules()
}
\value{
Returns a named list of the core modules.
}
\description{
Internal function.
}
\author{
Alex Chubaty
}
\keyword{internal}
