% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{createsOutput}
\alias{createsOutput}
\alias{createsOutput,character,character,character-method}
\title{This function is to define the output object that the module wants to export}
\usage{
createsOutput(objectName, objectClass, desc, ...)

\S4method{createsOutput}{character,character,character}(objectName, objectClass,
  desc, ...)
}
\arguments{
\item{objectName}{character vector, specify the object name}

\item{objectClass}{character, specify the object class}

\item{desc, }{character, the description for this output object}

\item{...}{other attibutes the module expects for this output object}
}
\value{
a data table
}
\description{
This function is to define the output object that the module wants to export
}
\author{
Yong Luo
}

