\docType{methods}
\name{crw}
\alias{crw}
\title{Correlated Random Walk}
\usage{
crw(agent, step.len, dir.sd, hab = NULL)
}
\arguments{
  \item{agent}{An object of class mobileAgent.}

  \item{step.len}{Description of this.}

  \item{dir.sd}{Description of this.}

  \item{hab}{Description of this.}
}
\value{
An agent object with updated spatial position.

#
}
\description{
Some additional details about this S4 generic and its
methods. The extra blank line between this section and the
title is critical for roxygen2 to differentiate the title
from the description section.
}
\seealso{
\code{\link{print}} and \code{\link{cat}}
}

