% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDefines.R
\docType{methods}
\name{defineEvent}
\alias{defineEvent}
\alias{defineEvent,simList,character,character-method}
\title{Define an event within a module}
\usage{
defineEvent(sim, moduleName, eventName)

\S4method{defineEvent}{simList,character,character}(sim, moduleName, eventName)
}
\arguments{
\item{sim}{DESCRIPTION NEEDED}

\item{moduleName}{DESCRIPTION NEEDED}

\item{eventName}{DESCRIPTION NEEDED}
}
\description{
Define an event within a module
}
\examples{
\dontrun{
## NOTHING YET
}

}
\author{
Alex Chubaty
}

