% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineIOObject.R
\docType{methods}
\name{defineIOObject}
\alias{defineIOObject}
\alias{defineIOObject,character,character,character,character,character-method}
\alias{defineIOObject,character,character,character,character,missing-method}
\alias{defineIOObject,character,character,character,missing,character-method}
\alias{defineIOObject,character,character,character,missing,missing-method}
\alias{defineIOObject,character,character,missing,character,character-method}
\alias{defineIOObject,character,character,missing,missing,character-method}
\alias{defineIOObject,character,character,missing,missing,missing-method}
\title{This function is to define the input and output object}
\usage{
defineIOObject(IO, objectName, objectClass, sourceURL, other)


  \S4method{defineIOObject}{character,character,character,character,character}(IO,
  objectName, objectClass, sourceURL, other)

\S4method{defineIOObject}{character,character,missing,character,character}(IO,
  objectName, sourceURL, other)

\S4method{defineIOObject}{character,character,character,missing,character}(IO,
  objectName, objectClass, other)

\S4method{defineIOObject}{character,character,character,character,missing}(IO,
  objectName, objectClass, sourceURL)

\S4method{defineIOObject}{character,character,missing,missing,character}(IO,
  objectName, other)

\S4method{defineIOObject}{character,character,character,missing,missing}(IO,
  objectName, objectClass)

\S4method{defineIOObject}{character,character,missing,missing,missing}(IO,
  objectName)
}
\arguments{
\item{IO}{character, specify whether the objects are inputs or outputs,
two options "input" or "output"}

\item{objectName}{character vector, specify the object name}

\item{objectClass}{character, specify the object class, default is NA}

\item{sourceURL}{character, specify the URL to reach the object, default is NA}

\item{other}{character,  notes or descriptions something like those.
default is NA}
}
\value{
a data frame
}
\description{
This function is to define the input and output object
}
\author{
Yong Luo
}

