% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module-dependencies.R
\docType{methods}
\name{defineModule}
\alias{defineModule}
\alias{defineModule,list-method}
\title{Define a new module.}
\usage{
defineModule(x)

\S4method{defineModule}{list}(x)
}
\arguments{
\item{x}{A named list containing the parameters used to construct a new
           \code{moduleDeps} object.}
}
\value{
This is a closure that serves as a wrapper for adding simulation
         dependencies via \code{simDepends(sim, add=TRUE)<-}.
}
\description{
Specify a new module's metadata as well as object and package dependecies.
Packages are loaded during this call.
}
\examples{
\dontrun{
  moduleInfo <- list(...)
  defineModule(moduleInfo)
}
}
\author{
Alex Chubaty
}

