% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{defineParameter}
\alias{defineParameter}
\alias{defineParameter,character,character-method}
\title{Define a parameter used in a module}
\usage{
defineParameter(name, class, default)

\S4method{defineParameter}{character,character}(name, class, default)
}
\arguments{
\item{name}{Character string giving the parameter name.}

\item{class}{Character string giving the parameter class.}

\item{default}{The default value to use when none is specified by the user.
                 Non-standard evaluation is used for the expression.}
}
\value{
data.frame
}
\description{
Used to specify a parameter's name, value, and set a default.
}
\examples{
\dontrun{
  rbind(
   defineParameter("lambda", "numeric", 1.23),
   defineParameter("mu", "numeric", 1e-3),
   defineParameter("nu", "numeric", 1e-2)
  )
}
}
\author{
Alex Chubaty
}

