% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{class}
\name{dependsList-class}
\alias{dependsList-class}
\title{The \code{dependsList} class}
\description{
List of SpaDES module object dependecies.
}
\section{Slots}{

\describe{
\item{\code{dependencies}}{List of \code{\link{dependsObj}} dependency objects.}
}}
\author{
Alex Chubaty
}

