% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{class}
\name{dependsObj-class}
\alias{dependsObj-class}
\title{The \code{dependsObj} class}
\description{
Contains minimum components of SpaDES module object dependecies.
}
\details{
Each \code{dependsObj} describes a module's dependency on another object for a simulation.
Multiple \code{dependsObj} objects (i.e., describing multiple dependencies) are typically supplied using a \code{\link{dependsList}}.
}
\section{Slots}{

\describe{
\item{\code{module}}{Character string naming the module for which object dependencies are being specified.}

\item{\code{objType}}{Character. One of "input" or "output", describing the relationship of the object to the module.}

\item{\code{objName}}{Character string naming the object.}

\item{\code{objClass}}{Character vector specifying the object's class(es).}

\item{\code{extent}}{\code{extent} object specifying the object's spatial extent.}

\item{\code{proj}}{Character string specifying the spatial projection of the object.
Usually the output of \code{\link{proj4string}} or \code{\link{crs}}.}

\item{\code{translators}}{Character}
}}
\author{
Alex Chubaty
}
\seealso{
dependsList
}

