% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module-dependencies-methods.R
\docType{methods}
\name{depsDetectCycles}
\alias{depsDetectCycles}
\alias{depsDetectCycles,igraph-method}
\title{Detect cycles in module dependencies}
\usage{
depsDetectCycles(simGraph)

\S4method{depsDetectCycles}{igraph}(simGraph)
}
\value{
List of paths corresponding to cycles.
}
\description{
Detect cycles in module dependencies
}
\author{
Alex Chubaty
}

