% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module-dependencies-methods.R
\docType{methods}
\name{depsEdgeList}
\alias{depsEdgeList}
\alias{depsEdgeList,simEnv,logical-method}
\alias{depsEdgeList,simEnv,missing-method}
\title{Build edge list for module dependency graph}
\usage{
depsEdgeList(sim, plot)

\S4method{depsEdgeList}{simEnv,logical}(sim, plot)

\S4method{depsEdgeList}{simEnv,missing}(sim, plot)
}
\arguments{
\item{sim}{A \code{simList} object.}

\item{plot}{Logical indicating whether the edgelist (and subsequent graph)
             will be used for plotting. If \code{TRUE}, duplicated rows
             (i.e., multiple object dependencies between modules) are removed
             so that only a single arrow is drawn connecting the modules.
             Default is \code{FALSE}.}
}
\value{
A \code{data.table} whose first two columns give a list of edges
         and remaining columns the attributes of the dependency objects
         (object name, class, etc.).
}
\description{
Build edge list for module dependency graph
}
\author{
Alex Chubaty
}

