% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module-dependencies-methods.R
\docType{methods}
\name{depsGraph}
\alias{depsGraph}
\alias{depsGraph,simList-method}
\title{Build a module dependency graph}
\usage{
depsGraph(sim)

\S4method{depsGraph}{simList}(sim)
}
\arguments{
\item{sim}{A \code{simList} object.}
}
\value{
An \code{\link{igraph}} object.
}
\description{
Build a module dependency graph
}
\author{
Alex Chubaty
}

