% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-dependencies-methods.R
\docType{methods}
\name{depsGraph}
\alias{depsGraph}
\alias{depsGraph,simList,logical-method}
\alias{depsGraph,simList,missing-method}
\title{Build a module dependency graph}
\usage{
depsGraph(sim, plot)

\S4method{depsGraph}{simList,logical}(sim, plot)

\S4method{depsGraph}{simList,missing}(sim)
}
\arguments{
\item{sim}{A \code{simList} object.}

\item{plot}{Logical indicating whether the edgelist (and subsequent graph)
will be used for plotting. If \code{TRUE}, duplicated rows
(i.e., multiple object dependencies between modules) are removed
so that only a single arrow is drawn connecting the modules.
Default is \code{FALSE}.}
}
\value{
An \code{\link{igraph}} object.
}
\description{
Build a module dependency graph
}
\author{
Alex Chubaty
}
