% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module-dependencies-methods.R
\docType{methods}
\name{depsPruneCycles}
\alias{depsPruneCycles}
\alias{depsPruneCycles,data.frame-method}
\title{Detect cycles in module dependencies}
\usage{
depsPruneCycles(edgeList)

\S4method{depsPruneCycles}{data.frame}(edgeList)
}
\arguments{
\item{edgeList}{A \code{data.frame} corresponding to an edge list describing
                 the dependency relationships among simulation modules.
                 Usually supplied via a call to \code{\link{simEdgeList}}.}
}
\value{
Modified edge list, with "offending" edge removed where possible.
}
\description{
Identify cycles and if posible, prune edges that are "safe" to remove
in order to produce an acyclic graph. Notifies the user if unable to
remove all cycles, and identifies what thoes cycles are.
}
\author{
Alex Chubaty
}

