% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module-dependencies-methods.R
\docType{methods}
\name{depsPruneEdges}
\alias{depsPruneEdges}
\alias{depsPruneEdges,data.frame-method}
\title{Prune edges to remove cycles in module dependencies}
\usage{
depsPruneEdges(simEdgeList)

\S4method{depsPruneEdges}{data.frame}(simEdgeList)
}
\arguments{
\item{simEdgeList}{An edge list produced by \code{\link{depsEdgeList}}.}
}
\value{
An updated edge list object.
}
\description{
Attempts to identify cycles in the dependency graph and remove edges representing
object dependencies which are provided by other modules in the simulation.
}
\author{
Alex Chubaty
}

