% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-other.R
\docType{methods}
\name{dev}
\alias{dev}
\title{Specify where to plot}
\usage{
dev(x, ...)
}
\arguments{
\item{x}{The number of a plot device. If missing, will open a new
non-RStudio plotting device}

\item{...}{Additional arguments passed to \code{\link{newPlot}}.}
}
\value{
Opens a new plot device on the screen. Invisibly returns the
device number selected.
}
\description{
Switch to an existing plot device, or if not already open,
launch a new graphics device based on operating system used. On Windows and
Mac, if no x is provided, then this will open or switch to the first non
R Studio device, which is much faster than the png-based R Studio Plot device.
Currently, this will not open anything new
}
\details{
For example, \code{dev(6)} switches the active plot device to device #6.
If it doesn't exist, it opens it. NOTE: if devices 1-5 don't exist
they will be opened too.
}
\author{
Eliot McIntire and Alex Chubaty
}
