% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-repository.R
\docType{methods}
\name{.digest}
\alias{.digest}
\alias{.digest,character-method}
\title{Calculate the hashes of multiple files}
\usage{
.digest(file, ...)

\S4method{.digest}{character}(file, algo = "xxhash64", ...)
}
\arguments{
\item{file}{Character vector of file paths.}

\item{...}{Additional arguments to \code{digest::digest}.}
}
\value{
A character vector of hashes.
}
\description{
Internal function. Wrapper for \code{\link[digest]{digest}} using md5sum.
}
\author{
Alex Chubaty
}
\keyword{internal}
