% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{doEvent,simList,logical-method}
\alias{doEvent,simList,logical-method}
\alias{doEvent,simList,missing-method}
\title{doEvent}
\usage{
\S4method{doEvent}{simList,logical}(sim, debug)

\S4method{doEvent}{simList,missing}(sim)
}
\description{
doEvent

doEvent
}

