% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{drawArrows}
\alias{drawArrows}
\alias{drawArrows,SpatialPoints,SpatialPoints-method}
\title{plot arrows showing direction of mobileAgent movement}
\usage{
drawArrows(current, previous, ...)

\S4method{drawArrows}{SpatialPoints,SpatialPoints}(current, previous, ...,
  length = 0.1)
}
\arguments{
\item{agent}{A \code{mobileAgent} object.}

\item{...}{Additional plotting parameters.}

\item{length}{The length of the arrows to draw (defaults to 0.1).}
}
\value{
Returns the modified \code{SimList} object.
}
\description{
Plots arrows showing direction of mobileAgent movement.

Plot arrows showing direction of mobileAgent movement
}
\examples{
current <- SpatialPoints(cbind(x=rnorm(10),y=rnorm(10)))
previous <- SpatialPoints(cbind(x=rnorm(10),y=rnorm(10)))
plot(current)
points(previous)
drawArrows(current, previous)
}

