% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{drawArrows}
\alias{drawArrows}
\title{Plots arrows showing direction of agent movement}
\usage{
drawArrows(from, to, addTo, title = TRUE, axes = TRUE, ...)

\S4method{drawArrows}{SpatialPoints,SpatialPoints,character}(from, to, addTo,
  title, axes, ..., length = 0.1)

\S4method{drawArrows}{SpatialPoints,SpatialPoints,missing}(from, to, addTo,
  title, axes, ..., length = 0.1)
}
\arguments{
\item{from}{Starting spatial coordinates (\code{SpatialPointsDataFrame}).}

\item{to}{Ending spatial coordinates (\code{SpatialPointsDataFrame}).}

\item{addTo}{Optional character string. The name of a map layer on which to draw the arrows.}

\item{title}{logical. Add title to plot. Defaults to \code{TRUE}.
Since this is also the viewport name, it is a good idea to plot it so the
viewport can be called for later plotting.}

\item{axes}{logical Add axes to plot. Defaults to \code{TRUE}.}

\item{...}{Additional plotting parameters passed to \code{\link{grid.polyline}}.
Currently does not appear to pass anything.}

\item{length}{numeric. Optional length, in inches, of arrow head on a SpatialLines object.}
}
\value{
Plots the vectors representing agent movement on the specified map. If plotted
on a blank plot, the plot object will be given the name of the "from" object, with
the word "Arrows" concatenated on the end.
}
\description{
Plots arrows showing direction of agent movement
}
\examples{
# Make 2 objects
caribou1 <- SpatialPoints(cbind(x=runif(10, -50, 50), y=runif(10, -50, 50)))
caribou2 <- SpatialPoints(cbind(x=runif(10, -50, 50), y=runif(10, -50, 50)))

drawArrows(caribou1, caribou2)

# or  to a previous Plot
fileList <-
   data.frame(files =
     dir(file.path(
                   find.package("SpaDES",
                                lib.loc=getOption("devtools.path"),
                                quiet=FALSE),
                  "maps"),
        full.names=TRUE, pattern= "tif"),
     functions="rasterToMemory",
     packages="SpaDES",
     stringsAsFactors=FALSE)

# Load files to memory (using rasterToMemory)
loadFiles(fileList=fileList)

Plot(DEM)
drawArrows(caribou1, caribou2, addTo="DEM")
}
\author{
Eliot McIntire
}

