% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{duplicatedInt}
\alias{duplicatedInt}
\title{Rcpp duplicated on integers using Rcpp Sugar}
\usage{
duplicatedInt(x)
}
\arguments{
\item{x}{Integer Vector}
}
\value{
A logical vector, as per \code{duplicated}
}
\description{
\code{.duplicatedInt} does same as \code{duplicated} in R, but only on integers, and faster.
It uses Rcpp sugar
}
