% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwdData.R
\docType{methods}
\name{dwdData}
\alias{dwdData}
\title{Download, unzip, untar data from an url}
\usage{
dwdData(urltble, datasetName, dfile = "all", outdir, sim, module,
  checkhash = TRUE, dbHash = "dbHash.sqlite", cascade = FALSE,
  quick = FALSE)
}
\arguments{
\item{urltble}{A data.table that stores available dataset name, url, password and filenames found within each dataset.
urltble is provided within the package as urls object.}

\item{datasetName}{A character string. Represent the dataset of interest for download. datasetName allow to derived
url and password from the urltble.}

\item{dfile}{A vector, given as a character string. Represent filename of interest to download. The default value is
to download everything using the argument "all".}

\item{outdir}{A character string. Indicate output directory where the files will be stored. Default will use the
modulePath from the sim object, if supplied with a module name, or a temporary location based on the
url of the file(s).}

\item{sim}{A simList simulation object, generally produced by simInit.}

\item{module}{A character string representing the names of the module to be loaded for the simulation.}

\item{checkhash}{A logical argument indicating if checksum values from files found locally should be cross-check with
checksum values logged from previous download. For first download, checksum values are compiled only. If checksum
values logged from previous download do not match checksum from files found locally, download will occur and
checksum value will be compiled. If FALSE, download will occur even if the files already exist locally.
Default value is TRUE.}

\item{dbHash}{A character string. Indicate the name of the database where checksum values are stored. If the database
does not exists, the function will create it and checksum will be compiled. Default is "dbHash.sqlite".}

\item{cascade}{A logical argument indicating if files should be untar and/or unzip.}

\item{quick}{A logical argument indicating if checksum should be compiled on the object (FALSE) or on character string
representing the combination of the filename its size (TRUE). Default is FALSE.}
}
\value{
Data will be stored in the subfolder using the basename of the url in a data folder.
}
\description{
This function download and untar and/or unzip dataset using url link from available dataset. Prior to download, the function
can check if the files are already locally by using the checksum value available in dbHash.sqlite database.
}
\examples{
dt<- data.table(dataset = c("NFDB"),
               url = c("http://cwfis.cfs.nrcan.gc.ca/downloads/nfdb/fire_poly/current_version/"),
               password= c(NA))
outdir<-tempdir()
mySim <- simInit(times=list(start=0.0, end=5.0),
                 objects=list(dt)
                 params=list(),
                 modules=list(),
                 paths=list(outputPath=outdir))
sim <- SpaDES::simInit(times = times, params = parameters, modules = modules, paths = paths)

dt<- data.table(dataset = c("NFDB"),
               url = c("http://cwfis.cfs.nrcan.gc.ca/downloads/nfdb/fire_poly/current_version/"),
               password= c(NA))
dwdData(dt, datasetName= 'NFDB', dfile = "all", outdir = tempdir(), sim, module, checkhash = FALSE, cascade = TRUE )
}
\author{
Melina Houle
}
