% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dwrpnorm2}
\alias{dwrpnorm2}
\title{Vectorized wrapped normal density function}
\usage{
dwrpnorm2(theta, mu, rho, sd = 1, acc = 1e-05, tol = acc)
}
\arguments{
\item{theta}{xxx}

\item{mu}{xxx}

\item{rho}{xxx}

\item{sd}{xxx}

\item{acc}{xxx}

\item{tol}{xxx}
}
\description{
This is a modified version of \code{\link{dwrpnorm}} found in \code{CircStats}
 to allow for multiple angles at once (i.e., this wersion is vectorized).
}
\author{
Eliot McIntire
}

