% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{methods}
\name{.emptyMetadata}
\alias{.emptyMetadata}
\alias{.emptyMetadata,missing-method}
\title{Default (empty) metadata}
\usage{
.emptyMetadata(x)

\S4method{.emptyMetadata}{missing}()
}
\arguments{
\item{x}{Not used. Should be missing.}
}
\description{
Internal use only.
Default values to use for metadata elements when not otherwise supplied.
}
\author{
Alex Chubaty
}
\keyword{internal}
