% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/numerical-comparisons.R
\docType{methods}
\name{setTolerance}
\alias{getTolerance}
\alias{getTolerance,missing-method}
\alias{setTolerance}
\alias{setTolerance,missing-method}
\alias{setTolerance,numeric-method}
\title{Tolerance for floating number comparisons}
\usage{
setTolerance(epsilon)

\S4method{setTolerance}{missing}(epsilon)

\S4method{setTolerance}{numeric}(epsilon)

getTolerance(epsilon)

\S4method{getTolerance}{missing}(epsilon)
}
\arguments{
\item{epsilon}{The numerical tolerance value (default \code{.Machine$double.eps^0.5}).}
}
\value{
Invoked for its side effect of setting the tolerance value in the package environment.
        This value will be used for relational operations defined in this package.
}
\description{
See \code{\link{all.equal}} for a good idea on default tolerance values for various classes.
}
\seealso{
relational-operators
}

