% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\docType{methods}
\name{equalExtent}
\alias{equalExtent}
\alias{equalExtent,list-method}
\title{Assess whether a list of extents are all equal}
\usage{
equalExtent(extents)

\S4method{equalExtent}{list}(extents)
}
\arguments{
\item{extents}{list of extents objects}
}
\description{
Assess whether a list of extents are all equal
}
\author{
Eliot McIntire
}

