% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting-diagrams.R
\docType{methods}
\name{eventDiagram}
\alias{eventDiagram}
\alias{eventDiagram,simList,character-method}
\title{Simulation event diagram}
\usage{
eventDiagram(sim, startDate)

\S4method{eventDiagram}{simList,character}(sim, startDate)
}
\arguments{
\item{sim}{A \code{simList} object (typically corresponding to a
            completed simulation).}

\item{startDate}{A character representation of date in YYYY-MM-DD format.}
}
\value{
Plots an event diagram as Gantt Chart.
}
\description{
Create a Gantt Chart representing the events in a completed simulation.
This event diagram is constructed using the completed event list, which by
default only stores the 10 most recently completed events (unless
\code{spades(debug=TRUE)} is used, in which case all events are retained).
To change the number of events stored, users may override this option using
\code{options(spades.nCompleted = value)}.
}
\details{
Simulation time is presented on the x-axis, starting at date 'startDate'.
Each module appears in a color-coded row, within which each event for that
module is displayed corresponding to the sequence of events for that module.
Note that only the start time of the event is meaningful is these figures:
the width of the bar associated with a particular module's event corresponds
to the module's timestepUnit, not the event's "duration".

Based on this StackOverflow answer: \url{http://stackoverflow.com/a/29999300/1380598}.
}
\note{
A red vertical line corresponding to the current date may appear on the figure.
This is useful for Gantt Charts generally but can be considered a 'bug' here.
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{mermaid}}.
}

