% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/environment.R
\docType{methods}
\name{existsGlobal}
\alias{existsGlobal}
\title{Is an object defined in the global environment?}
\usage{
existsGlobal(x, ...)

\S4method{existsGlobal}{ANY}(x, ...)
}
\arguments{
\item{x}{An object name, given as a character string.
           No coercion is done, and the first element of a character vector
           of length greater than one will be used, with a warning.}

\item{...}{Additional arguments passed to \code{\link{exists}}}
}
\description{
Simple wrapper for \code{\link{exists}}.
}
\author{
Alex Chubaty
}

