% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{expectsInput}
\alias{expectsInput}
\alias{expectsInput,character,character,character,character-method}
\alias{expectsInput,character,character,character,missing-method}
\title{This function is to define the input that the module expects}
\usage{
expectsInput(name, class, desc, sourceURL, ...)

\S4method{expectsInput}{character,character,character,character}(name, class,
  desc, sourceURL, ...)

\S4method{expectsInput}{character,character,character,missing}(name, class,
  desc, sourceURL, ...)
}
\arguments{
\item{name}{character vector, specify the object name}

\item{class}{character, specify the object class}

\item{desc, }{character, the description for this input object}

\item{sourceURL}{character, specify the URL to reach the object, default is NA}

\item{...}{other attibutes the module expects for this input object}
}
\value{
a data table
}
\description{
This function is to define the input that the module expects
}
\author{
Yong Luo
}

