% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{F}
\alias{F}
\alias{simList-accessors-function}
\alias{F,.simList-method}
\title{\code{F} allows the user to access functions from their module. It should
be unnecessary to use this as namespacing should take care of it. But, if
the user wants to specificy a function this way, it will work too.
To access a function from within a module, you can use \code{F(sim)$functionName}}
\usage{
F(sim, module = NULL, functionName = NULL)

\S4method{F}{.simList}(sim, module = NULL, functionName = NULL)
}
\arguments{
\item{sim}{A \code{simList} object from which to extract element(s) or
in which to replace element(s).}

\item{module}{Optional character string indicating which module params should come from.}

\item{functionName}{Optional character string indicating which function is desired.}
}
\description{
\code{F} allows the user to access functions from their module. It should
be unnecessary to use this as namespacing should take care of it. But, if
the user wants to specificy a function this way, it will work too.
To access a function from within a module, you can use \code{F(sim)$functionName}
}
