% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{fileExt}
\alias{fileExt}
\title{Extract the file extension of a file}
\usage{
fileExt(x)
}
\arguments{
\item{x}{List or character vector of file names.}
}
\value{
A character vector of file extensions.
}
\description{
Extract the file extension of a file
}
\author{
Eliot McIntire and Alex Chubaty
}
