\docType{data}
\name{.fileExtensions}
\alias{.fileExtensions}
\title{File extensions map}
\format{\preformatted{'data.frame':	6 obs. of  3 variables:
 $ exts     : Factor w/ 6 levels "asc","csv","png",..: 5 3 2 4 6 1
 $ functions: Factor w/ 4 levels "raster","read.csv",..: 1 1 2 4 3 1
 $ package  : Factor w/ 3 levels "raster","rgdal",..: 1 1 3 2 3 1
}}
\usage{
.fileExtensions
}
\description{
How to load various types of files in R.
}
\keyword{datasets}

