% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{fileName}
\alias{fileName}
\title{Extract filename (without extension) of a file}
\usage{
fileName(x)
}
\arguments{
\item{x}{List or character vector}
}
\value{
A character vector.
}
\description{
Extract filename (without extension) of a file
}
\author{
Eliot McIntire
}
