% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{.fileTableIn}
\alias{.fileTableIn}
\alias{.fileTableIn,missing-method}
\alias{.fileTableOut}
\alias{.fileTableOut,missing-method}
\title{Create empty fileTable for inputs and outputs}
\usage{
.fileTableIn(x)

\S4method{.fileTableIn}{missing}()

.fileTableOut(x)

\S4method{.fileTableOut}{missing}()
}
\arguments{
\item{x}{Not used (should be missing)}
}
\value{
An empty data.frame with structure needed for input/output fileTable.
}
\description{
Internal functions.
Returns an empty fileTable to be used with inputs and outputs.
}

