% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\name{.fillOutputRows}
\alias{.fillOutputRows}
\title{An internal function for coercing a data.frame to inputs()}
\usage{
.fillOutputRows(outputDF, endTime)
}
\arguments{
\item{endTime}{Numeric time. The end(sim).}

\item{inputDF}{A data.frame with partial columns to pass to inputs( ) <-}
}
\description{
An internal function for coercing a data.frame to inputs()
}

