% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-simInit.R
\docType{methods}
\name{.findAllObjNames}
\alias{.findAllObjNames}
\alias{.findAllObjNames,simList-method}
\title{Find all input and output object names, and update the internal list}
\usage{
.findAllObjNames(sim)

\S4method{.findAllObjNames}{simList}(sim)
}
\arguments{
\item{sim}{A \code{simList} object.}
}
\description{
Used to track the names of shared data dependencies in a simulation.
}
\details{
This is faster than accessing the \code{.moduleDeps} object on demand.
}
\author{
Alex Chubaty
}
\keyword{internal}

