% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/initialize.R
\name{findFactors}
\alias{findFactors}
\title{Find Factors}
\usage{
findFactors(x)
}
\arguments{
\item{x}{An integer to factorize}
}
\value{
A vector of integer factors
}
\description{
Finds the integer factors of an integer
}
\details{
Used internally in GaussMap
}
\seealso{
\code{\link{GaussMap}}
}

