% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{methods}
\name{findObjects}
\alias{findObjects}
\alias{.findObjects}
\title{Find objects if passed as character strings}
\usage{
.findObjects(objects, functionCall = "simInit")
}
\arguments{
\item{objects}{A character vector of object names}

\item{functionCall}{A character string identifying the function name to be
searched in the call stack. Default is "simInit"}
}
\description{
Objects are passed into simList via \code{simInit} call or \code{objects(simList)}
assignment. This function is an internal helper to find those objects from their
environments by searching the call stack.
}
\author{
Eliot McIntire
}
\keyword{internal}
