% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{GaussMap}
\alias{GaussMap}
\title{GaussMap}
\usage{
GaussMap(x, scale = 10, var = 1, speedup = 10, inMemory = FALSE, ...)
}
\arguments{
\item{x}{An spatial object (e.g., a raster).}

\item{scale}{The spatial scale in map units of the Gaussian pattern.}

\item{var}{Spatial variance.}

\item{speedup}{An index of how much faster than normal to generate maps.}

\item{inMemory}{Should the RasterLayer be forced to be in memory? Default \code{FALSE}.}

\item{...}{Additional arguments to \code{raster}.}
}
\value{
A map of extent \code{ext} with a Gaussian random pattern.
}
\description{
Produces a raster of a random gaussian process.
}
\details{
This is a wrapper for the \code{RFsimulate} function in the RandomFields
package. The main addition is the \code{speedup} argument which allows
for faster map generation. A \code{speedup} of 1 is normal and will get
progressively faster as the number increases, at the expense of coarser pixel
resolution of the pattern generated
}
\seealso{
\code{\link{RFsimulate}} and \code{\link{extent}}
}

