\docType{methods}
\name{GaussMap}
\alias{GaussMap}
\title{GaussMap}
\usage{
GaussMap(ext, scale = 10, var = 1, speedup = 10)
}
\arguments{
  \item{ext}{Description of \code{ext}.}

  \item{scale}{Description of \code{scale}.}

  \item{var}{Description of \code{var}.}

  \item{speedup}{Description of \code{speedup}.}
}
\value{
A GaussMap thing, which needs an explanation.

#
}
\description{
Some additional details about this S4 generic and its
methods. The extra blank line between this section and the
title is critical for roxygen2 to differentiate the title
from the description section.
}
\seealso{
\code{\link{print}} and \code{\link{cat}}
}

