% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-colours.R
\docType{methods}
\name{getColors}
\alias{getColors}
\alias{getColours}
\alias{getColors,Raster-method}
\alias{getColors,ANY-method}
\alias{getColors,SpatialPoints-method}
\title{Get colours for plotting Raster* objects.}
\usage{
getColors(object)

\S4method{getColors}{Raster}(object)

\S4method{getColors}{ANY}(object)

\S4method{getColors}{SpatialPoints}(object)
}
\arguments{
\item{object}{A \code{Raster*} object.}
}
\value{
Returns a named list of colors.
}
\description{
Get colours for plotting Raster* objects.
}
\seealso{
\code{\link{setColors<-}},
         \code{\link[RColorBrewer]{brewer.pal}}
}
\author{
Alex Chubaty

#@importClassesFrom NetLogoRClasses agentMatrix
}
