% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc-methods.R
\docType{methods}
\name{getFileName}
\alias{getFileName}
\alias{getFileName,logical-method}
\title{Get the name of a source file}
\usage{
getFileName(fullname)

\S4method{getFileName}{logical}(fullname)
}
\arguments{
\item{fullname}{Logical (default \code{FALSE}) indicating whether the full path should be returned.}
}
\value{
Character string representing the filename.
}
\description{
This will only work for files that are \code{source}d.
Based on this: \url{http://stackoverflow.com/a/1816487/1380598}.
}
\author{
Alex Chubaty
}

