% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-methods.R
\docType{methods}
\name{getFileName}
\alias{getFileName}
\alias{getFileName,logical-method}
\title{Get the name of a \code{source}-ed file}
\usage{
getFileName(fullname)

\S4method{getFileName}{logical}(fullname)
}
\arguments{
\item{fullname}{Logical (default \code{FALSE}) indicating whether the full
path should be returned.}
}
\value{
Character string representing the filename.
}
\description{
Use \code{getFileName} in a file that is \code{source}-ed.
Based on \url{http://stackoverflow.com/a/1816487/1380598}.
}
\author{
Alex Chubaty
}

