% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/environment.R
\docType{methods}
\name{getGlobal}
\alias{getGlobal}
\alias{getGlobal,ANY-method}
\title{Get objects from the global environment}
\usage{
getGlobal(x, ...)

\S4method{getGlobal}{ANY}(x, ...)
}
\arguments{
\item{x}{a variable name, given as a character string.
           No coercion is done, and the first element of a character vector
           of length greater than one will be used, with a warning.}

\item{x}{For \code{getGlobal}, an object name (given as a character string).
           For \code{mgetGlobal}, a character vector of object names.}
}
\description{
Simple wrapper for \code{\link{get}}.
}
\author{
Alex Chubaty
}

