% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-repository.R
\docType{methods}
\name{getModuleVersion}
\alias{getModuleVersion}
\alias{getModuleVersion,character,character-method}
\alias{getModuleVersion,character,missing-method}
\title{Find the latest module version from a SpaDES module repository}
\usage{
getModuleVersion(name, repo)

\S4method{getModuleVersion}{character,character}(name, repo)

\S4method{getModuleVersion}{character,missing}(name)
}
\arguments{
\item{name}{Character string giving the module name.}

\item{repo}{GitHub repository name.
Default is \code{"PredictiveEcology/SpaDES-modules"}, which is
specified by the global option \code{spades.moduleRepo}.}
}
\description{
Modified from \url{http://stackoverflow.com/a/25485782/1380598}.
}
\author{
Alex Chubaty
}
