% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module-dependencies.R
\docType{methods}
\name{getSimDeps}
\alias{getSimDeps}
\alias{getSimDeps,missing-method}
\title{Get simulation dependencies}
\usage{
getSimDeps(x)

\S4method{getSimDeps}{missing}()
}
\arguments{
\item{x}{placeholder only. not used.}
}
\description{
Get the simulation dependency list stored in \code{.spadesEnv}.
}
\examples{
\dontrun{
  getSimDeps()
}
}
\author{
Alex Chubaty
}
\seealso{
moduleDeps
}

