% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\docType{methods}
\name{.getSpaDES}
\alias{.getSpaDES}
\alias{.getSpaDES,ANY-method}
\title{Get objects from the internal SpaDES environment}
\usage{
.getSpaDES(x, ...)

\S4method{.getSpaDES}{ANY}(x, ...)
}
\arguments{
\item{x}{an object name (given as a character string).}

\item{...}{Additional arguments to pass to \code{get}.}
}
\description{
Internal function. Simple wrapper for \code{\link{get}}.
}
\author{
Alex Chubaty
}
\keyword{internal}
