% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\docType{methods}
\name{gpar}
\alias{gpar}
\alias{gpar,ANY-method}
\title{Importing some grid functions}
\usage{
gpar(...)

\S4method{gpar}{ANY}(...)
}
\arguments{
\item{...}{ Any number of named arguments. }
}
\description{
Currently only the gpar function is imported. This is a convenience so that users
can change \code{Plot} arguments without having to load the entire grid package.
}
\seealso{
\code{\link[grid]{gpar}}
}
