% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{head,mobileAgent-method}
\alias{head,mobileAgent-method}
\title{head method for mobileAgent class}
\usage{
\S4method{head}{mobileAgent}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
head method for mobileAgent class
}

