% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{heading}
\alias{heading}
\alias{heading,SpatialPoints,SpatialPoints-method}
\title{Heading between spatial points.}
\usage{
heading(from, to, ...)

\S4method{heading}{SpatialPoints,SpatialPoints}(from, to, ...)
}
\arguments{
\item{from}{The starting position; an object of class SpatialPoints.}

\item{to}{The ending position;  an object of class SpatialPoints.}

\item{...}{Additional arguments.}
}
\value{
The heading between the points, in degrees.
}
\description{
Determines the heading between spatial points.
}

