% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hList.R
\docType{methods}
\name{hList}
\alias{hList}
\title{Extract checksum value}
\usage{
hList(flist, dir, quick, csalgorithm = "xxhash64")
}
\arguments{
\item{flist}{A vector listing object from which we want to extract chekcsum value.}

\item{dir}{A character string representing the location of the file.}

\item{quick}{logical. Set to TRUE if checksum value should be computed using the combination of filename and size.
Set to FALSE to return checksum value from the object.}

\item{csalgorithm}{The algorithms to be used in the digest function. Currently the default in this function is xxhash64.}
}
\value{
The hList function returns a dataframe where filename, checksum value from the object (checksumFile),
        checksum value from the combination of filename and file size and the algorithm used to compute
        checksum values are stored.
}
\description{
The function hList extract checksum value from a list of file and return the value its respective filename in an
R object. The function is for now using the algorithm 'xxhash64'.
}
\examples{
file.list<- list.files(tempdir())
hfile <-hList(file.list, dir = tempdir(), quick = TRUE)
}
\author{
Melina Houle
}
