\docType{methods}
\name{initialize,agent-method}
\alias{initialize,agent-method}
\title{initialize agent}
\usage{
\S4method{initialize}{agent}(.Object, ..., numagents = NULL)
}
\arguments{
  \item{.Object}{ An object: see the Details section.}

  \item{...}{data to include in the new object.  Named
  arguments correspond to slots in the class definition.
  Unnamed arguments must be objects from classes that this
  class extends.}
}
\description{
initialize agent
}

