% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-class.R
\docType{methods}
\name{initialize,simList-method}
\alias{initialize,simList-method}
\title{Generate a \code{simList} object}
\usage{
\S4method{initialize}{simList}(.Object)
}
\arguments{
\item{.Object}{A \code{simList} object.}
}
\description{
Given the name or the definition of a class, plus optionally data to be
included in the object, \code{new} returns an object from that class.
}
