\docType{methods}
\name{initialize,mobileAgent-method}
\alias{initialize,mobileAgent-method}
\title{initialize mobileAgent}
\usage{
\S4method{initialize}{mobileAgent}(.Object, ..., agentlocation = NULL,
  numagents = NULL, probinit = NULL)
}
\arguments{
  \item{.Object}{ An object: see the Details section.}

  \item{...}{data to include in the new object.  Named
  arguments correspond to slots in the class definition.
  Unnamed arguments must be objects from classes that this
  class extends.}
}
\description{
initialize mobileAgent
}

