\docType{methods}
\name{initialize,pointAgent-method}
\alias{initialize,pointAgent-method}
\title{initialize pointAgent}
\usage{
\S4method{initialize}{pointAgent}(.Object, ..., numagents = NULL)
}
\arguments{
  \item{numagents}{The number of agents to initialize.}

  \item{.Object}{ An object: see the Details section.}

  \item{...}{data to include in the new object.  Named
  arguments correspond to slots in the class definition.
  Unnamed arguments must be objects from classes that this
  class extends.}
}
\description{
initialize pointAgent
}

