% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-dependencies-class.R
\docType{methods}
\name{.inputObjects}
\alias{.inputObjects}
\alias{.inputObjects,missing-method}
\alias{.outputObjects}
\alias{.outputObjects,missing-method}
\title{Create an empty (template) inputObjects and outputObjects data.frames}
\usage{
.inputObjects(x)

\S4method{.inputObjects}{missing}()

.outputObjects(x)

\S4method{.outputObjects}{missing}()
}
\arguments{
\item{x}{Not used. Should be missing.}
}
\value{
An empty inputObjects or outputObjects data.frame.
}
\description{
Internal function.
}
\author{
Alex Chubaty
}
\keyword{internal}
