% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\name{.fillInputRows}
\alias{.fillInputRows}
\title{.fillInputRows is internal}
\usage{
.fillInputRows(inputDF, startTime)
}
\arguments{
\item{inputDF}{A data.frame with partial columns to pass to inputs( ) <-}

\item{startTime}{Numeric time. The start(sim).}
}
\description{
.fillInputRows is internal
}

