% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{inRasterStack}
\alias{inRasterStack}
\title{determine which of the layers are provided within a stack}
\usage{
inRasterStack(object)

\S4method{inRasterStack}{list}(object)
}
\description{
determine which of the layers are provided within a stack
}

