% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{makeExtsToPlot}
\alias{.makeExtsToPlot,list-method}
\alias{makeExtsToPlot}
\title{Get dimensions of x and y axes}
\usage{
.makeExtsToPlot(toPlot = NULL, zoomExtent = NULL, numLayers = NULL, lN)

\S4method{.makeExtsToPlot}{list}(toPlot = NULL, zoomExtent = NULL,
  numLayers = NULL, lN)
}
\arguments{
\item{toPlot}{list of objects to plot}

\item{zoomExtent}{an extent object}

\item{numLayers}{numeric indicating number of layers in the list \code{toPlot}}

\item{lN}{character vector of names of these layers}
}
\description{
For spatial objects, these are the extents. Used internally in Plot methods
}

