% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{smallestTimestepUnit}
\alias{smallestTimestepUnit}
\alias{smallestTimestepUnit,simList-method}
\title{Determine what the the largest timestep units}
\usage{
smallestTimestepUnit(sim)

\S4method{smallestTimestepUnit}{simList}(sim)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}
}
\value{
The timestep unit as a character string
}
\description{
When modules have different timestep units, SpaDES automatically takes the
largest (e.g., "year") as the unit for a simulation. This function determines which
is the largest unit
}
\author{
Eliot McIntire
}

