% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\docType{methods}
\name{layerNames}
\alias{layerNames}
\alias{layerNames,list-method}
\alias{layerNames,ANY-method}
\alias{layerNames,Raster-method}
\alias{layerNames,.spadesPlot-method}
\alias{layerNames,igraph-method}
\title{Extract the layer names of Spatial Objects}
\usage{
layerNames(object)

\S4method{layerNames}{list}(object)

\S4method{layerNames}{ANY}(object)

\S4method{layerNames}{Raster}(object)

\S4method{layerNames}{.spadesPlot}(object)

\S4method{layerNames}{igraph}(object)
}
\arguments{
\item{object}{A \code{Raster*}, \code{SpatialPoints*}, \code{SpatialLines*},
or \code{SpatialPolygons*} object; or list of these.}
}
\description{
There are already methods for \code{Raster*} objects. This adds methods for
\code{SpatialPoints*}, \code{SpatialLines*}, and \code{SpatialPolygons*},
returning an empty character vector of length 1.
This function was created to give consistent, meaningful results for all
classes of objects plotted by \code{Plot}.
}
\author{
Eliot McIntire
}
