% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length,agent-method}
\alias{length,agent-method}
\title{length agent (this is really \code{num.agents()})}
\usage{
\S4method{length}{agent}(x)
}
\arguments{
\item{x}{an \R object.  For replacement, a vector or factor.}
}
\description{
length agent (this is really \code{num.agents()})
}

