% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module-repository.R
\docType{methods}
\name{listModules}
\alias{listModules}
\alias{listModules,character,character-method}
\alias{listModules,character,missing-method}
\alias{listModules,missing,character-method}
\alias{listModules,missing,missing-method}
\title{List modules from local or on github repository}
\usage{
listModules(repo, path)

\S4method{listModules}{character,character}(repo, path)

\S4method{listModules}{missing,missing}()

\S4method{listModules}{character,missing}(repo)

\S4method{listModules}{missing,character}(path)
}
\arguments{
\item{repo}{GitHub repository name.
               Default is \code{"PredictiveEcology/SpaDES-modules"},
               which is specified by the global option \code{spades.modulesRepo}.}

\item{path}{Character string of length 1. The base directory (relative)
 within which there are only module subdirectories. Defaults to "modules".}
}
\description{
Modified from
\url{http://stackoverflow.com/questions/25485216/how-to-get-list-files-from-a-github-repository-folder-using-r}.
If repo is unspecified, then modulesDir
}
\author{
Eliot McIntire
}

