% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{load.packages}
\alias{load.packages}
\alias{load.packages,list,logical,logical-method}
\alias{load.packages,list,logical,missing-method}
\alias{load.packages,list,missing,logical-method}
\alias{load.packages,list,missing,missing-method}
\title{Load packages.}
\usage{
load.packages(package.list, install, quiet)

\S4method{load.packages}{list,logical,logical}(package.list, install, quiet)

\S4method{load.packages}{list,missing,missing}(package.list)

\S4method{load.packages}{list,missing,logical}(package.list, quiet)

\S4method{load.packages}{list,logical,missing}(package.list, install)
}
\arguments{
\item{package.list}{A list of character strings specifying
                       the names of packages to be loaded.}

\item{install}{Logical flag. If required packages are not
                       already installed, should they be installed?}

\item{quiet}{Logical flag. Should the final "packages loaded"
                       message be suppressed?}
}
\value{
Nothing is returned. Specified packages are loaded and attached using \code{library()}.
}
\description{
Load and optionally install additional packages.
}
\examples{
\dontrun{pkgs <- list("ggplot2", "lme4")}
\dontrun{load.packages(pkgs) # loads packages if installed}
\dontrun{load.packages(pkgs, install=TRUE) # loads packages after installation (if needed)}
}
\seealso{
\code{\link{library}}.
}

