% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logChecksum.R
\docType{methods}
\name{logChecksum}
\alias{logChecksum}
\title{List available dataset}
\usage{
logChecksum(hashdata, dbHash)
}
\arguments{
\item{hashdata}{A dataframe containing filename, checksumFile, checksumSize and algorithm for each file are stored.}

\item{dbHash}{A character string indicating the name of the database where checksum values are stored. If the database
does not exists, the function will create it and checksum will be compiled. Default is "dbHash.sqlite".}
}
\value{

}
\description{
This function store in a list available data from url website.
}
\examples{

dt<- data.table(dataset = c("NFDB"),
               url = c("http://cwfis.cfs.nrcan.gc.ca/downloads/nfdb/fire_poly/current_version/"),
               password= c(NA))

setwd(tempdir())
logChecksum(hashdata)
}
\author{
Melina Houle
}
