% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{ls.simList}
\alias{ls,simList-method}
\alias{ls.simList}
\title{List simulation objects}
\usage{
ls.simList(name)

\S4method{ls}{simList}(name)
}
\arguments{
\item{name}{A \code{simList} object.}
}
\description{
Return a vector of character strings giving the names of the objects in the
specified simulation environment.
Can be used with a \code{simList} object, because the method for this class
is simply a wrapper for calling \code{ls} on the simulation environment
stored in the \code{simList} object.
}

