% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{makeColorMatrix}
\alias{makeColorMatrix}
\alias{makeColorMatrix,Raster,Extent,numeric-method}
\title{Convert Raster to color matrix useable by raster function for plotting}
\usage{
makeColorMatrix(grobToPlot, zoomExtent, maxpixels, legendRange, cols = NULL,
  na.color = "#FFFFFF00", zero.color = "#FFFFFF00", skipSample = TRUE)

\S4method{makeColorMatrix}{Raster,Extent,numeric}(grobToPlot, zoomExtent,
  maxpixels, legendRange, cols = NULL, na.color = "#FFFFFF00",
  zero.color = "#FFFFFF00", skipSample = TRUE)
}
\arguments{
\item{grobToPlot}{a SpatialObject}

\item{zoomExtent}{an extent object for zooming to. Defaults to whole extent of grobToPlot}

\item{maxpixels}{numeric. Number of cells to subsample the complete grobToPlot}

\item{legendRange}{numeric vector of length 2, representing the lower and upper bounds of
a legend that will override the data bounds contained within the grobToPlot}

\item{cols}{colours specified in a way that can be understood directly or by
 colorRampPalette}

\item{na.color}{string indicating the color for NA values. Default transparent.}

\item{zero.color}{string indicating the color for zero values, when zero is
the minimum value, otherwise, it is treated as any other color. Default transparent.}

\item{skipSample}{logical. If no downsampling is necessary, skip. Default TRUE.}
}
\description{
Convert Raster to color matrix useable by raster function for plotting
}

