% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeLayout}
\alias{makeLayout}
\title{Make an optimal layout of plots}
\usage{
makeLayout(arr, visualSqueeze, legend = TRUE, axes = TRUE, title = TRUE)
}
\arguments{
\item{arr}{an object of class \code{arrangement}}

\item{visualSqueeze}{numeric. The proportion of the white space to be used for
plots. Default is 0.75.}

\item{legend}{logical. Whether legend should be included as part of layout
calculation. Default is TRUE.}

\item{axes}{Logical. Whether the axes should be included as part of layout
calculation. Default is TRUE.}

\item{title}{Logical. Whether the names of each plot should be written above
plots and should be included as part of layout calculation.  Default is TRUE.}
}
\description{
Using the size of the current device, and number and dimension ratios of the plots,
this function will place them optimally in the plotting region.
}

