% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\docType{methods}
\name{.makeList}
\alias{.makeList}
\alias{.makeList,list-method}
\title{Convert \code{plotArgs} to list of lists}
\usage{
.makeList(plotArgs, numSpadesPlotObjects)

\S4method{.makeList}{list}(plotArgs, numSpadesPlotObjects)
}
\arguments{
\item{plotArgs}{The arguments passed to \code{Plot} as a \code{list}.}

\item{numSpadesPlotObjects}{Numeric. The number of \code{.spadesPlotObjects}.
                This can't easily be deduced from the \code{plotArgs} because
                of the RasterStacks. So passed manually.}
}
\description{
Internal function. Take the inputs as plotArgs to the Plot function, and make
them a list of length \code{numSpadesPlotObjects} entries of lists.
}
\author{
Eliot McIntire
}

