\docType{methods}
\name{makeViewport}
\alias{makeViewport}
\alias{makeViewport,Raster-method}
\title{makeViewport}
\usage{
makeViewport(obj, layout.pos.col, layout.pos.row, visualSqueeze = 0.75)

\S4method{makeViewport}{Raster}(obj, layout.pos.col = 1, layout.pos.row = 1)
}
\description{
makeViewport
}

