% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{makeViewports}
\alias{makeViewports}
\title{Make viewports}
\usage{
makeViewports(extents, layout, arr, newArr = FALSE)
}
\arguments{
\item{extents}{a list of extents objects.}

\item{layout}{an object with layouts described, normally created by \code{makeLayouts}.}

\item{arr}{an object of class \code{arrangement}.}

\item{newArr}{logical. Whether this function will create a completely new viewport. Default \code{FALSE}.}
}
\description{
Given a set of extents, and a layout for these extents, this function will output
a viewport tree to allow plotting.
}
\details{
This function will either create a totally new set of viewports, or simply add
some nested viewports to an existing arrangement, i.e., is there still white
space availabe to plot.
}

