% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread-process.R
\name{\%fin\%}
\alias{\%fin\%}
\alias{match}
\title{A faster '\%in\%' based on fastmatch package}
\usage{
x \%fin\% table
}
\arguments{
\item{x}{See \code{\link[fastmatch]{fmatch}}.}

\item{table}{See \code{\link[fastmatch]{fmatch}}.}
}
\description{
A faster '\%in\%', directly pulled from \code{fastmatch::match}, based on
\url{http://stackoverflow.com/questions/32934933/faster-in-operator}.
}

