% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread-process.R
\name{.matchedPointDirection}
\alias{.matchedPointDirection}
\alias{matchedPointDirection}
\title{Calculate matched point directions}
\usage{
.matchedPointDirection(to, from)
}
\description{
Internal function
}
\keyword{internal}

