% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\docType{methods}
\name{maxTimeunit}
\alias{maxTimeunit}
\alias{maxTimeunit,simList-method}
\title{Determine the largest timestep unit in a simulation}
\usage{
maxTimeunit(sim)

\S4method{maxTimeunit}{simList}(sim)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}
}
\value{
The timeunit as a character string. This defaults to \code{NA} if
none of the modules has explicit units.
}
\description{
Determine the largest timestep unit in a simulation
}
\author{
Eliot McIntire and Alex Chubaty
}
