% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc-methods.R
\docType{methods}
\name{mergeLists}
\alias{mergeLists}
\alias{mergeLists,NULL,NULL-method}
\alias{mergeLists,NULL,list-method}
\alias{mergeLists,list,NULL-method}
\alias{mergeLists,list,list-method}
\title{Merge named lists}
\usage{
mergeLists(x, y)

\S4method{mergeLists}{list,list}(x, y)

\S4method{mergeLists}{`NULL`,list}(x, y)

\S4method{mergeLists}{list,`NULL`}(x, y)

\S4method{mergeLists}{`NULL`,`NULL`}(x, y)
}
\arguments{
\item{x}{a named list}

\item{y}{a named list}
}
\value{
A named list, with elements sorted by name.
         The values of matching elements in list \code{y}
         replace the values in list \code{x}.
}
\description{
Merge two named list based on their named entries. Where
any element matches in both lists, the value from the
second list is used in the updated list.
Subelements are not examined and are simply replaced.
}
\examples{
L1 <- list(a="hst", b=NA_character_, c=43)
L2 <- list(a="gst", c=42, d=list(letters))
mergeLists(L, L2)
}
\author{
Alex Chubaty
}

