% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moduleMetadata.R
\docType{methods}
\name{keywords}
\alias{extents}
\alias{extents,list-method}
\alias{extents,simList-method}
\alias{keywords}
\alias{keywords,list-method}
\alias{keywords,simList-method}
\alias{timeframes}
\alias{timeframes,list-method}
\alias{timeframes,simList-method}
\title{Parse and extract specific elements of module metadata}
\usage{
keywords(x)

\S4method{keywords}{list}(x)

\S4method{keywords}{simList}(x)

extents(x)

\S4method{extents}{list}(x)

\S4method{extents}{simList}(x)

timeframes(x)

\S4method{timeframes}{list}(x)

\S4method{timeframes}{simList}(x)
}
\arguments{
\item{x}{Either a named list of module metadata (i.e., the result of a call to
\code{moduleMetadata}) or a \code{simList}.}
}
\value{
A list of various named elements in the metadata
}
\description{
These functions will return specific elements of the metadata, either via
passing a \code{simList} object, or by passing the module metadata list
output from the \code{metadata} function.
}
\examples{
library(igraph)
 path <- system.file(package="SpaDES", "sampleModules")
 myMetadata <- listModules(path=path) \%>\%
   moduleMetadata(., path=path)
 keywords(myMetadata)

 ## use a simList object
 times <- list(start=0, end=1, timeunit="second")
 parameters <- list(.globals=list(stackName="landscape"),
                    randomLandscapes=list(.plotInitialTime=NA))
 modules <- list("randomLandscapes", "caribouMovement")
 paths <- list(modulePath=system.file("sampleModules", package="SpaDES"),
               outputPath=tempdir())
 sim1 <- simInit(times=times, params=parameters, modules=modules, paths=paths)
 keywords(sim1)
library(igraph)
 path <- system.file(package="SpaDES", "sampleModules")
 myMetadata <- listModules(path=path) \%>\%
   moduleMetadata(., path=path)
 extents(myMetadata)
library(igraph)
 path <- system.file(package="SpaDES", "sampleModules")
 myMetadata <- listModules(path=path) \%>\%
   moduleMetadata(., path=path)
 timeframes(myMetadata)
}

