% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moduleMetadata.R
\docType{methods}
\name{metadata}
\alias{metadata}
\alias{metadata,character,character,missing-method}
\alias{metadata,character,missing,missing-method}
\alias{metadata,character,missing,simList-method}
\alias{metadata,missing,missing,simList-method}
\title{Parse and extract module metadata}
\usage{
metadata(module, path, simList)

\S4method{metadata}{character,character,missing}(module, path)

\S4method{metadata}{character,missing,missing}(module)

\S4method{metadata}{character,missing,simList}(module, simList)

\S4method{metadata}{missing,missing,simList}(simList)
}
\arguments{
\item{module}{Character vector. The names of modules from which metadata should
be returned.}

\item{path}{Character string specifying the file path to modules directory.
              Default is the current working directory.}

\item{simList}{a \code{simList} object. If this is provided, module metadata
will be returned from modules contained within the \code{simList}, not the \code{path}}
}
\value{
A list of module metadata, matching the structure in
        \code{\link{defineModule}}.
}
\description{
This function will return the metadata from either a module file or a
\code{simList} object. If a \code{simList} object is provided, then the
\code{path} is ignored.
}
\examples{
library(igraph)
 path <- system.file(package="SpaDES", "sampleModules")
 sampleModules <- dir(path)
 (x <- metadata(sampleModules, path))

 listModules(path=path) \%>\%
   metadata(module=., path=path)
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{defineModule}}
}

