% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.modifySearchOrder}
\alias{.modifySearchOrder}
\title{Modify package order in search path}
\usage{
.modifySearchOrder(pkgs, curSearch)
}
\arguments{
\item{pkgs}{The packages that are to be placed at the beginning of the seach path}

\item{curSearch}{The current search path, i.e., a call to \code{search()}}
}
\value{
Nothing. This is used for its side effects, which are "severe".
}
\description{
Intended for internal use only. It modifies the search path (i.e., \code{search()})
such that the packages required by the current module are placed first in the
search path.
}
\keyword{internal}

