% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modules-dependecies.R
\docType{class}
\name{moduleDeps-class}
\alias{moduleDeps-class}
\title{The \code{moduleDeps} class}
\description{
Descriptor object for specifying SpaDES module dependecies.
}
\details{
Each \code{moduleDeps} describes a module's dependency on another object for a simulation.
Multiple \code{moduleDeps} objects (i.e., describing multiple dependencies) are typically supplied using a \code{\link{dependsList}}.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the module as a character string.}

\item{\code{description}}{Description of the module as a character string.}

\item{\code{keywords}}{Character vector containing a module's keywords.}

\item{\code{authors}}{The author(s) of the module as a \code{\link{person}} object.}

\item{\code{spatialExtent}}{Specifies the module's spatial extent as an \code{\link{Extent}} object.}

\item{\code{timeframe}}{Specifies the valid timeframe for which the module was designed to simulate.
                   Must be a \code{\link{POSIXlt}} object of length 2, specifying the start and end times
                   (e.g., \code{as.POSIXlt(c("1990-01-01 00:00:00", "2100-12-31 11:59:59"))}).}

\item{\code{translators}}{Character vector describing the "translators" available for this module.}

\item{\code{citation}}{A citation for the module, as a character string.}

\item{\code{objectDeps}}{A \code{data.frame} specifying the object dependecies of the module:
                   each row specifies a directional dependency, with columns
                   \code{objectName}, \code{objectClass}, and \code{objectType}.
                   These first two are self-explanatory, and \code{objectType}
                   should be one of \code{"input"} or \code{"output"}.
                   For objects that are used within the module as both an input and an output,
                   each of these should be added as separate entries in the data.frame.}
}}
\author{
Alex Chubaty
}
\seealso{
dependsList
}

