% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting-diagrams.R
\docType{methods}
\name{moduleDiagram}
\alias{moduleDiagram}
\alias{moduleDiagram,simList-method}
\title{Simulation module dependency diagram}
\usage{
moduleDiagram(sim)

\S4method{moduleDiagram}{simList}(sim)
}
\arguments{
\item{sim}{A \code{simList} object (typically corresponding to a
            completed simulation).}
}
\value{
Plots module dependency diagram.
}
\description{
Create a network diagram illustrating the simplified module dependencies of a
simulation. Offers a less detailed view of specific objects than does
plotting the \code{depsEdgeList} directly with \code{\link{objectDiagram}}.
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{igraph}}.
}

