% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-diagrams.R
\docType{methods}
\name{moduleGraph}
\alias{moduleGraph}
\alias{moduleGraph,simList,logical-method}
\alias{moduleGraph,simList,missing-method}
\title{Build a module dependency graph}
\usage{
moduleGraph(sim, plot, ...)

\S4method{moduleGraph}{simList,logical}(sim, plot, ...)

\S4method{moduleGraph}{simList,missing}(sim, plot, ...)
}
\arguments{
\item{sim}{A \code{simList} object.}

\item{plot}{Logical indicating whether the edgelist (and subsequent graph)
will be used for plotting. If \code{TRUE}, duplicated rows
(i.e., multiple object dependencies between modules) are removed
so that only a single arrow is drawn connecting the modules.
Default is \code{FALSE}.}

\item{...}{Arguments passed to \code{Plot}}
}
\value{
A list with 2 elements, an \code{\link{igraph}} object and an \code{igraph}
communities object.

# @importFrom igraph graph_from_data_frame cluster_optimal edges # already with import igraph
}
\description{
This is still experimental, but this will show the hierarchical structure of
parent and children modules and return a list with an igraph object
and an igraph communities object, showing the groups.
Currently only tested with relatively simple structures.
}
\seealso{
moduleDiagram
}
\author{
Eliot McIntire
}
