\docType{methods}
\name{name,SpatialPointsNamed-method}
\alias{layerNames,list-method}
\alias{name,SpatialPointsNamed-method}
\alias{name<-}
\title{set list of simulation modules}
\usage{
\S4method{name}{SpatialPointsNamed}(object)

name(object) <- value

\S4method{layerNames}{list}(object)
}
\description{
set list of simulation modules
}

