% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{name}
\alias{name}
\title{Get and set the name of an object.}
\usage{
name(object)
}
\arguments{
\item{...}{Additional arguments.}
}
\description{
Using the simulation visualization features in \code{Plot} requires
objects to be plotted to have a name set. Currently, required of:
\code{SpatialPoints}, \code{SpatialPointsDataFrame}, and \code{RasterStack}
objects. Adding a name to an object of these classes converts these objects
to a \code{*Named} class.
}
\author{
Eliot McIntire

Alex Chubaty
}
\seealso{
\code{\link{SpatialPointsNamed}},
         \code{\link{SpatialPointsDataFrameNamed}},
         \code{\link{RasterStackNamed}}.
}

