% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{newModule}
\alias{newModule}
\alias{newModule,character,character,logical-method}
\alias{newModule,character,character,missing-method}
\alias{newModule,character,missing,logical-method}
\alias{newModule,character,missing,missing-method}
\title{Create new module from template.}
\usage{
newModule(name, path, open)

\S4method{newModule}{character,character,logical}(name, path, open)

\S4method{newModule}{character,missing,logical}(name, open)

\S4method{newModule}{character,character,missing}(name, path)

\S4method{newModule}{character,missing,missing}(name)
}
\arguments{
\item{name}{Character string. Your module's name.}

\item{path}{Character string. Subdirectory in which to place the new module code file.
             The default is the current working directory.}

\item{open}{Logical. Should the new module file be opened after creation?
             Default \code{TRUE}.}
}
\value{
Nothing is returned. The new module is created at \code{path/name.R}.
}
\description{
Autogenerate a skeleton for a new SpaDES module.
}
\examples{
\dontrun{## create a "fastfood" module in the "modules" subdirectory.}
\dontrun{newModule("fastfood", "modules")}
}
\author{
Alex Chubaty
}

