% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-template.R
\docType{methods}
\name{newModuleDocumentation}
\alias{newModuleDocumentation}
\alias{newModuleDocumentation,character,character,logical,character,character-method}
\alias{newModuleDocumentation,character,missing,logical,ANY,ANY-method}
\alias{newModuleDocumentation,character,character,missing,ANY,ANY-method}
\alias{newModuleDocumentation,character,missing,missing,ANY,ANY-method}
\title{Create new module documentation}
\usage{
newModuleDocumentation(name, path, open, type, children)


  \S4method{newModuleDocumentation}{character,character,logical,character,character}(name,
  path, open, type, children)

\S4method{newModuleDocumentation}{character,missing,logical,ANY,ANY}(name, open)

\S4method{newModuleDocumentation}{character,character,missing,ANY,ANY}(name,
  path)

\S4method{newModuleDocumentation}{character,missing,missing,ANY,ANY}(name)
}
\arguments{
\item{name}{Character string specifying the name of the new module.}

\item{path}{Character string. Subdirectory in which to place the new module code file.
The default is the current working directory.}

\item{open}{Logical. Should the new module file be opened after creation?
Default \code{TRUE}.}

\item{type}{Character string specifying one of \code{"child"} (default),
or \code{"parent"}.}

\item{children}{Required when \code{type = "parent"}. A character vector
specifying the names of child modules.}
}
\description{
Create new module documentation
}
\seealso{
Other module creation helpers: \code{\link{newModuleCode}},
  \code{\link{newModuleTests}}, \code{\link{newModule}}
}
\author{
Eliot McIntire and Alex Chubaty
}
