% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-template.R
\docType{methods}
\name{newModuleTests}
\alias{newModuleTests}
\alias{newModuleTests,character,character,logical-method}
\title{Create template testing structures for new modules}
\usage{
newModuleTests(name, path, open)

\S4method{newModuleTests}{character,character,logical}(name, path, open)
}
\arguments{
\item{name}{Character string specifying the name of the new module.}

\item{path}{Character string. Subdirectory in which to place the new module code file.
The default is the current working directory.}

\item{open}{Logical. Should the new module file be opened after creation?
Default \code{TRUE}.}
}
\description{
Create template testing structures for new modules
}
\seealso{
Other module creation helpers: \code{\link{newModuleCode}},
  \code{\link{newModuleDocumentation}},
  \code{\link{newModule}}
}
\author{
Eliot McIntire and Alex Chubaty
}
