% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{newPlot}
\alias{newPlot}
\title{Open a new plotting window}
\usage{
newPlot(noRStudioGD = TRUE, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{noRStudioGD}{logical Passed to dev.new. Default is TRUE to avoid using
RStudio graphics device, which is slow.}
}
\description{
Open a new plotting window
}
\note{
\code{\link{dev.new}} is supposed to be the correct way to open a new
window in a platform-generic way, however, this doesn't work in RStudio.
}
\seealso{
\code{\link{dev}}.
}

