% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\docType{methods}
\name{nlayers,list-method}
\alias{nlayers,.spadesPlot-method}
\alias{nlayers,ANY-method}
\alias{nlayers,list-method}
\title{Find the number of layers in a Spatial Object}
\usage{
\S4method{nlayers}{list}(x)

\S4method{nlayers}{.spadesPlot}(x)

\S4method{nlayers}{ANY}(x)
}
\arguments{
\item{x}{A \code{.spadesPlotObjects} object or list of these.}
}
\value{
The number of layers in the object.
}
\description{
There are already methods for \code{Raster*} in the raster package.
Adding methods for \code{list}, \code{SpatialPolygons}, \code{SpatialLines},
and \code{SpatialPoints}, \code{gg}, \code{histogram}, \code{igraph}.
These latter classes return \code{1}.
}
\author{
Eliot McIntire
}

