% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{.objectNames}
\alias{.objectNames}
\title{Extracts the object name inside a call to \code{Plot}}
\usage{
.objectNames(calledFrom = "Plot")
}
\arguments{
\item{calledFrom}{character vector, length 1, indicating which function call is
desired. Defaults to \code{Plot}}
}
\description{
Extracts the object name inside a call to \code{Plot}
}

