% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{o}
\alias{o}
\alias{o,.simList-method}
\title{\code{o} is a stub. It may be used in a future release as part of namespacing of objects in a module.}
\usage{
o(object, name = NULL, module = NULL)

\S4method{o}{.simList}(object, name = NULL, module = NULL)
}
\description{
\code{o} is a stub. It may be used in a future release as part of namespacing of objects in a module.
}

