% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{methods}
\name{.objSizeInclEnviros,simList-method}
\alias{.objSizeInclEnviros,simList-method}
\title{objSizeInclEnviros for simList class objects}
\usage{
\S4method{.objSizeInclEnviros}{simList}(object)
}
\arguments{
\item{object}{Any R object.}
}
\description{
See \code{\link[reproducible]{.objSizeInclEnviros}}.
}
\seealso{
\code{\link[reproducible]{.objSizeInclEnviros}}
}
