% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module-template.R
\docType{methods}
\name{openAllModules}
\alias{openAllModules}
\alias{openAllModules,character-method}
\alias{openAllModules,missing-method}
\title{Open all modules nested within a base directory}
\usage{
openAllModules(basedir)

\S4method{openAllModules}{character}(basedir)

\S4method{openAllModules}{missing}()
}
\arguments{
\item{basedir}{Character string. The base directory within which there are only
module subdirectories}
}
\value{
Nothing is returned. All file are open via \code{file.edit}.
}
\description{
This is just a convenience wrapper for openning several modules at once, recursively.
}
\examples{
\dontrun{openAllModules("~\\SpaDESModules")}
}
\author{
Eliot McIntire
}

