% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{packages}
\alias{packages}
\alias{packages,simList-method}
\title{Get simulation package dependencies}
\usage{
packages(sim)

\S4method{packages}{simList}(sim)
}
\arguments{
\item{sim}{A \code{simList} object.}
}
\value{
A sorted character vector of package names.
}
\description{
Get simulation package dependencies
}
\author{
Alex Chubaty
}

