% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{params}
\alias{p}
\alias{p,.simList-method}
\alias{params}
\alias{params,.simList-method}
\alias{params<-}
\alias{params<-,.simList-method}
\alias{simList-accessors-params}
\title{Get and set simulation parameters.}
\usage{
params(object)

\S4method{params}{.simList}(object)

params(object) <- value

\S4method{params}{.simList}(object) <- value

p(object, module)

\S4method{p}{.simList}(object, module)
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}

\item{module}{Optional character string indicating which module params should come from.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
\code{params} and \code{p} access the parameter slot in the \code{simList}.
\code{params} has a replace method, so can be used to update a parameter value.

\code{p} is a concise way to access parameters within a module. It works more like
a namespaced function in the sense that the module from which it is called is the
default place it will look for the parameter. To access a parameter from within
a module, you can use \code{p(sim)$paramName} instead of
\code{params(sim)$moduleName$paramName}
}
\details{
Additonal methods are provided to access core module and global parameters:
Commonly used
\tabular{ll}{
   \code{globals} \tab List of global simulation parameters.\cr
   \code{params} \tab Nested list of all simulation parameters.\cr
}
Advanced use
\tabular{lll}{
   Accessor method \tab Module \tab Description \cr
   \code{checkpointFile} \tab \code{.checkpoint} \tab Name of the checkpoint file. (advanced)\cr
   \code{checkpointInterval} \tab \code{.checkpoint} \tab The simulation checkpoint interval. (advanced)\cr
   \code{progressType} \tab \code{.progress} \tab Type of graphical progress bar used. (advanced)\cr
   \code{progressInterval} \tab \code{.progress} \tab Interval for the progress bar. (advanced)\cr
}

Currently, only get and set methods are defined. Subset methods are not.
}
\seealso{
\code{\link{SpaDES}}, specifically the section 1.2 on \code{simList} methods.
         \code{\link{simList-class}},
         \code{\link{simList-accessors-envir}},
         \code{\link{simList-accessors-events}},
         \code{\link{simList-accessors-inout}},
         \code{\link{simList-accessors-modules}},
         \code{\link{simList-accessors-objects}},
         \code{\link{simList-accessors-paths}},
         \code{\link{simList-accessors-times}}.
}

