% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{.parseElems,simList-method}
\alias{.parseElems,simList-method}
\title{.parseElems for simList class objects}
\usage{
\S4method{.parseElems}{simList}(tmp, elems, envir)
}
\arguments{
\item{tmp}{A evaluated object}

\item{elems}{A character string to be parsed}

\item{envir}{An environment}
}
\description{
See \code{\link[quickPlot]{.parseElems}}.
}
\seealso{
\code{\link[quickPlot]{.parseElems}}
}
