% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{.parseTimeunit}
\alias{.parseModule}
\alias{.parseModule,simList,list-method}
\alias{.parseModulePartial}
\alias{.parseModulePartial,character,character-method}
\alias{.parseTimeunit}
\alias{.parseTimeunit,simList,list-method}
\title{Parse and initialize a module}
\usage{
.parseTimeunit(sim, modules)

\S4method{.parseTimeunit}{simList,list}(sim, modules)

.parseModulePartial(filename, defineModuleElement)

\S4method{.parseModulePartial}{character,character}(filename,
  defineModuleElement)

.parseModule(sim, modules)

\S4method{.parseModule}{simList,list}(sim, modules)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}

\item{modules}{A list of modules with a logical attribute "parsed".}

\item{filename}{The filename of the module to be parsed.}

\item{defineModuleElement}{Character string indicating which of the list
elements in defineModule should be extracted}
}
\value{
.parseTimeunit just returns the time units of each module

\code{.parseModulePartial} extracts just the individual element
requested from the module. This can be useful if parsing the whole module
would cause an error.#'

A \code{simList} simulation object.
}
\description{
Internal function, used during \code{\link{simInit}}.
}
\author{
Eliot McIntire

Eliot McIntire

Alex Chubaty
}

