% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{.paths}
\alias{.paths}
\title{Default paths to use for a simulation}
\usage{
.paths()
}
\value{
A named list of paths used in the simulation.
}
\description{
Internal function, used during \code{\link{simInit}}.
}
\author{
Alex Chubaty
}
\keyword{internal}

