% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modules-dependecies.R
\docType{class}
\name{person4-class}
\alias{person4-class}
\title{Persons (S4)}
\description{
An S4 implementation of the \code{\link{person}} class.
}
\author{
Alex Chubaty
}
\seealso{
person
}

