% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.pkgEnv}
\alias{.pkgEnv}
\title{The \code{SpaDES} package environment}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.pkgEnv
}
\description{
Environment used internally to store internal package objects and methods.
}
\keyword{internal}
