% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Plot}
\alias{Plot}
\alias{Plot,spatialObjects-method}
\title{Plot: Fast, optimally arranged, multipanel plotting function with spades}
\usage{
Plot(..., new = FALSE, addTo = NULL, gp = gpar(), axes = "L",
  speedup = 1, size = 5, cols = NULL, zoomExtent = NULL,
  visualSqueeze = 0.75, legend = TRUE, legendRange = NULL, draw = TRUE,
  pch = 19, title = TRUE, na.color = "#FFFFFF00",
  zero.color = "#FFFFFF00")

\S4method{Plot}{spatialObjects}(..., new = FALSE, addTo = NULL,
  gp = gpar(), axes = "L", speedup = 1, size = 5, cols = NULL,
  zoomExtent = NULL, visualSqueeze = 0.75, legend = TRUE,
  legendRange = NULL, draw = TRUE, pch = 19, title = TRUE,
  na.color = "#FFFFFF00", zero.color = "#FFFFFF00")
}
\arguments{
\item{...}{Raster* object(s) and or SpatialPoints* objects. See details for
naming.}

\item{new}{Logical. If TRUE, then the previous plot is wiped and a new one made; if FALSE, then the ... plots
will be added to the current device, adding or rearranging the plot layout as necessary. Default is FALSE.}

\item{addTo}{String vector, with same length as ...  This is for overplotting, when the overplot is not to occur on
the plot with the same name, such as plotting a SpatialPoints* object on a RasterLayer.}

\item{gp}{A gpar object, created by gpar() function, to change plotting parameters (see \code{\link{grid}} package)}

\item{axes}{Logical or "L", representing the left and bottom axes, overall plots}

\item{speedup}{Numeric. The factor by which the number of pixels is divided by to plot rasters. See Details.}

\item{size}{Numeric. The size, in points, for SpatialPoints symbols, if using a scalable symbol.}

\item{cols}{character vector. Hex codes for colors. See Details.}

\item{zoomExtent}{extent object. Supplying a single extent object that is smaller than the
rasters will call a crop statement before plotting. Defaults to NULL. This occurs after any
downsampling of rasters, so it may produce very pixelated maps.}

\item{visualSqueeze}{numeric. The proportion of the white space to be used for plots. Default is 0.75.}

\item{legend}{logical. Whether legend should be drawn next to plots. Default is TRUE.}

\item{legendRange}{numeric vector of length 2, representing the lower and upper bounds of
a legend that will override the data bounds contained within the grobToPlot}

\item{draw}{logical, whether to actually draw the plots. Currently, there is no reason for this
to be FALSE. Default is TRUE}

\item{pch}{see ?par}

\item{title}{Logical. Whether the names of each plot should be written above plots}

\item{na.color}{string indicating the color for NA values. Default transparent.}

\item{zero.color}{string indicating the color for zero values, when zero is
the minimum value, otherwise, zero is treated as any other color. Default transparent.}
}
\description{
The main plotting function accompanying spades. This can take objects of type Raster* or SpatialPoints*,
and any combination of those.
}
\details{
If new=TRUE, then a new plot will be generated. When new=FALSE, then any plot that
already exists will be overplotted, while plots that have not already been plotted will be added. This function
rearrange the plotting device to maximize the size of all the plots, minimizing white space. If using RStudio,
it is recommended to make and use a new device because the built in device is not made for rapid redrawing.
The function is based on the grid package.

Each panel in the multipanel plot must have a name. This name is used to overplot,
rearrange the plots, or overlay using \code{addTo} when necessary. If the \code{...} are named
spatialObjects, then \code{Plot} will use these names. If not, then \code{Plot}
will use the object name and the layer name (in the case of RasterLayers or
RasterStacks). If plotting a RasterLayer and the layer name is "layer" or the
same as the object name, then only the object name will be used for simplicity. In
other words, only enough information is used to uniquely identify the plot.

Silently, one hidden object is made, \code{.spadesArr}, which is used for arranging plots in the
device window, and identifying the objects to be replotted if rearranging is required,
subsequent to an new=FALSE additional plot.

\code{speedup} is not a precise number because it is faster to plot an un-resampled
raster if the new resampling is close to the original number of pixels. At the moment,
this is set to 1/3 of the original pixels. In other words, \code{speedup} may not do anything
if the factor for speeding up is not high enough (i.e., >3).

\code{col} can be used to set the colors of Raster* objects, but it is preferable to use
setColors to give each layer its own color table. See examples.
}
\examples{
\dontrun{
library(raster)
library(rgdal)
#  Make list of maps from package database to load, and what functions to use to load them
fileList <-
   data.frame(files =
     dir(file.path(
                   find.package("SpaDES",
                                lib.loc=getOption("devtools.path"),
                                quiet=FALSE),
                  "maps"),
        full.names=TRUE, pattern= "tif"),
     functions="rasterToMemory",
     .stackName="landscape",
     packages="SpaDES",
     stringsAsFactors=FALSE)

# Load files to memory (using rasterToMemory) and stack them (because .stackName is provided above)
loadFiles(fileList=fileList)

# extract a single one of these rasters
DEM <- landscape$DEM

# can change color palette
setColors(landscape, n = 50)<-list(DEM=topo.colors(50),
                           forestCover = RColorBrewer::brewer.pal(9,"Set1"),
                           forestAge = RColorBrewer::brewer.pal("Blues",n=8),
                           habitatQuality = RColorBrewer::brewer.pal(9,"Spectral"),
                           percentPine = RColorBrewer::brewer.pal("GnBu",n=8))

#Make a new raster derived from a previous one; must give it a unique name
habitatQuality2 <- landscape$habitatQuality ^ 0.3
names(habitatQuality2) <- "habitatQuality2"

# make a SpatialPoints object
caribou <- SpatialPoints(coords=cbind(x=runif(1e2,-50,50),y=runif(1e2,-50,50)))


#Plot all maps on a new plot windows - Do not use RStudio window
\\notrun{
if(is.null(dev.list())) {
  dev(2)
} else {
  if(any(names(dev.list())=="RStudioGD")) {
    dev(which(names(dev.list())=="RStudioGD")+3)
  } else {
    dev(max(dev.list()))
  }
}
}

Plot(landscape)

# Can overplot, using addTo
Plot(caribou, addTo="landscape.forestAge", size=4, axes=FALSE)

# can add a plot to the plotting window
Plot(caribou, new=FALSE)

# Can add two maps with same name, if one is in a stack; they are given
#  unique names based on object name
Plot(landscape, caribou, DEM)

# can mix stacks, rasters, SpatialPoint*
Plot(landscape, habitatQuality2, caribou)

# can mix stacks, rasters, SpatialPoint*, and SpatialPolygons*
Plot(landscape, caribou)
Plot(habitatQuality2, new=FALSE)
Sr1 = Polygon(cbind(c(2,4,4,1,2),c(2,3,5,4,2))*20-50)
Sr2 = Polygon(cbind(c(5,4,2,5),c(2,3,2,2))*20-50)
Srs1 = Polygons(list(Sr1), "s1")
Srs2 = Polygons(list(Sr2), "s2")
SpP = SpatialPolygons(list(Srs1,Srs2), 1:2)
Plot(SpP)
Plot(SpP, addTo="landscape.Fires", gp=gpar(lwd=2))

}
}

