% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{points,pointAgent-method}
\alias{points,pointAgent-method}
\title{add agents to plot}
\usage{
\S4method{points}{pointAgent}(x, ..., which.to.plot = NULL)
}
\arguments{
\item{which.to.plot}{Optional subset of agent IDs to plot.}

\item{x}{coordinate vectors of points to plot.}

\item{...}{Further \link{graphical parameters} may also be supplied as
    arguments.  See \sQuote{Details}.}
}
\description{
add agents to plot
}

