\docType{class}
\name{polygonAgent-class}
\alias{polygonAgent-class}
\title{The \code{polygonAgent} class.}
\description{
Extends the \code{vectorAgent} class. Describes spatial
agents whose spatial components are characterized by
polygons (i.e., in contrast to points or a raster).
}
\details{
\describe{ \item{ID}{A character}

\item{spatial}{A \code{SpatialPolygons} object.}

\item{other}{A list for storing other (user-specific) agent
attributes.} }
}

