% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POM.R
\docType{methods}
\name{POM}
\alias{POM}
\alias{POM,simList-method}
\title{Use Pattern Oriented Modeling to fit unknown parameters}
\usage{
POM(sim, params, objects, objFn, cl, optimizer = "DEoptim", ...)

\S4method{POM}{simList}(sim, params, objects, objFn, cl,
  optimizer = "DEoptim", ...)
}
\arguments{
\item{sim}{A \code{simList} simulation object, generally produced by \code{simInit}.}

\item{objFn}{An objective function to be passed into
\code{optimizer}}

\item{cl}{A cluster object. Optional. This would generally be created using
parallel::makeCluster or equivalent. This is an alternative way, instead
of beginCluster(), to use parallelism for this function, allowing for
more control over cluster use.}

\item{optimizer}{The function to use to optimize. Default is
DEoptim}

\item{...}{All objects needed in objFn}
}
\value{
The values for parameters used in objFn that minimize
the objFn.
}
\description{
Use Pattern Oriented Modeling to fit unknown parameters
}
\examples{
\dontrun{
 set.seed(52461)
 mySim <- simInit(
   times = list(start = 0.0, end = 2.0, timeunit = "year"),
   params = list(
     .globals = list(stackName = "landscape", burnStats = "nPixelsBurned"),
     fireSpread = list(nFires = 1),
     randomLandscapes = list(nx = 200, ny = 200)
   ),
   modules = list("randomLandscapes", "fireSpread", "caribouMovement"),
   paths = list(modulePath = system.file("sampleModules", package = "SpaDES"))
 )
 out <- spades(copy(mySim), .plotInitialTime = NA)
 fireData <- out$landscape$Fires
 Plot(fireData, new=TRUE)


 library(parallel)
 library(DEoptim)
 objFn <- function(spreadProb, fireMap, simList) {
    params(mySim)$fireSpread$spreadprob = spreadProb
    out <- spades(copy(simList), .plotInitialTime = NA)
    abs(cellStats(fireMap - out$landscape$Fires, "sum"))
 }
 cl <- makeCluster(6)
 clusterExport(cl, c("mySim", "fireData"))
 clusterEvalQ(cl, {
 library(SpaDES)
 library(RColorBrewer)
 library(raster)
 })
 output <- DEoptim(objFn, lower = 0.10, upper = 0.26, simList = mySim, fireMap = fireData,
     control=list(parallelType=3, itermax = 10), cl = cl)
 stopCluster(cl)

 )
 }
}
\author{
Eliot McIntire
}
\references{
Matloff, N. (2011). The Art of R Programming (ch. 7.8.3). San Fransisco, CA: No Starch Press, Inc.. Retrieved from \url{https://www.nostarch.com/artofr.htm}
}
\seealso{
\code{\link{spades}}, \code{\link[parallel]{makeCluster}},
\code{\link{simInit}}
}

