% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{methods}
\name{prepareFileBackedRaster}
\alias{prepareFileBackedRaster}
\title{Alternative to \code{archivist::saveToRepo} for rasters}
\usage{
prepareFileBackedRaster(obj, repoDir = NULL,
  compareRasterFileLength = 1e+06, ...)
}
\arguments{
\item{obj}{The raster object to save to the repository.}

\item{repoDir}{Character denoting an existing directory in which an artifact will be saved.}

\item{compareRasterFileLength}{Numeric. Optional. When there are Rasters, that
have file-backed storage, this is passed to the length arg in \code{digest}
when determining if the Raster file is already in the database.
Note: uses \code{\link[digest]{digest}} for file-backed Raster.
Default 1e6. Passed to \code{prepareFileBackedRaster}.}

\item{...}{passed to \code{archivist::saveToRepo}}
}
\value{
A raster object and its file backing will be passed to the archivist repository.
}
\description{
Rasters are sometimes file-based, so the normal save mechanism doesn't work.
This function creates an explicit save of the file that is backing the raster,
in addition to saving the object metadata in the \code{archivist} repository database.
}
\author{
Eliot McIntire
}
