% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{methods}
\name{.prepareOutput,simList-method}
\alias{.prepareOutput,simList-method}
\title{prepareOutput for simList class objects}
\usage{
\S4method{.prepareOutput}{simList}(object, cacheRepo, ...)
}
\arguments{
\item{object}{Any R object}

\item{cacheRepo}{A repository used for storing cached objects.
This is optional if \code{Cache} is used inside a SpaDES module.}

\item{...}{Arguments of \code{FUN} function .}
}
\description{
See \code{\link[reproducible]{.prepareOutput}}.
}
\seealso{
\code{\link[reproducible]{.prepareOutput}}
}
