% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{.prepareRaster}
\alias{.prepareRaster}
\title{Prepare raster for plotting}
\usage{
.prepareRaster(grobToPlot, zoomExtent, legendRange, takeFromPlotObj, arr,
  speedup, newArr)
}
\arguments{
\item{zoomExtent}{}

\item{legendRange}{}

\item{takeFromPlotObj}{}

\item{arr}{}

\item{speedup}{}

\item{newArr}{}

\item{gropToPlot}{}
}
\description{
Prepare raster for plotting
}
\author{
Eliot McIntire
}

