% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/environment.R
\docType{methods}
\name{prevArgs}
\alias{prevArgs}
\alias{prevArgs,character-method}
\title{Get the arguments from a previously made function call}
\usage{
prevArgs(calledFrom)

\S4method{prevArgs}{character}(calledFrom)
}
\arguments{
\item{calledFrom}{Name of previounly called function.}
}
\value{
List contaning the frame called from and arguments previously passed
        to the function \code{calledFrom}.
}
\description{
Get the arguments from a previously made function call
}
\author{
Eliot Mcintire and Alex Chubaty
}

