% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priority.R
\docType{methods}
\name{priority}
\alias{.first}
\alias{.highest}
\alias{.last}
\alias{.lowest}
\alias{.normal}
\alias{priority}
\title{Event priority}
\usage{
.first()

.highest()

.last()

.lowest()

.normal()
}
\value{
A numeric.
}
\description{
Preset envent priorities: 1 = first (highest); 5 = normal; 10 = last (lowest).
}
\author{
Alex Chubaty
}

