% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{probInit}
\alias{probInit}
\title{probInit}
\usage{
probInit(map, p = NULL, absolute = FALSE)
}
\arguments{
\item{map}{Description of this.}

\item{p}{Description of this.}

\item{absolute}{Description of this.}
}
\value{
Decribe what it returns: fromto.
}
\description{
Details here.
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{print}} and \code{\link{cat}}
}

