% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{progressInterval}
\alias{progressInterval}
\alias{progressInterval,.simList-method}
\alias{progressInterval<-}
\alias{progressInterval<-,.simList-method}
\alias{progressType}
\alias{progressType,.simList-method}
\alias{progressType<-}
\alias{progressType<-,.simList-method}
\title{Get and set simulation progress bar details}
\usage{
progressInterval(object)

\S4method{progressInterval}{.simList}(object)

progressInterval(object) <- value

\S4method{progressInterval}{.simList}(object) <- value

progressType(object)

\S4method{progressType}{.simList}(object)

progressType(object) <- value

\S4method{progressType}{.simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\description{
The progress bar can be set in two ways in SpaDES. First, by setting values
in the .progress list element in the params list element passed to \code{\link{simInit}}.
Second, at the \code{\link{spades}} call itself, which can be simpler. See examples.
}
\details{
Progress Bar:
Progress type can be "text", "graphical" or "shiny".
Progress interval can be a numeric.
These both can get set by passing a
\code{.progress=list(type="graphical", interval=1)} into the
simInit call. See examples
}
\examples{
\dontrun{
mySim <- simInit(times=list(start=0.0, end=100.0),
                 params=list(.globals=list(stackName="landscape"),
                             .progress=list(type="text", interval=10),
                             .checkpoint = list(interval = 10, file = "chkpnt.RData")),
                 modules=list("randomLandscapes"),
                 paths=list(modulePath=system.file("sampleModules", package="SpaDES")))

# progress bar
progressType(mySim) # "text"
progressInterval(mySim) # 10

# parameters
params(mySim) # returns all parameters in all modules
              # including .global, .progress, .checkpoint
globals(mySim) # returns only global parameters

# checkpoint
checkpointFile(mySim) # returns the name of the checkpoint file
                      # In this example, "chkpnt.RData"
checkpointInterval(mySim) # 10
}
}

