\docType{class}
\name{rasterAgent-class}
\alias{rasterAgent-class}
\title{The \code{rasterAgent} class.}
\description{
Extends the \code{agent} class by making it spatial.
Describes spatial agents whose spatial components are
characterized by a raster (i.e., in contrast to points or a
polygon). This is a virtual class and not intended to be
used directly.
}
\details{
\describe{ \item{ID}{A character}

\item{other}{A list for storing other (user-specific) agent
attributes.} }
}

