% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapReduce.R
\docType{methods}
\name{rasterizeReduced}
\alias{rasterizeReduced}
\title{rasterizeReduced - convert reduced representation to full raster}
\usage{
rasterizeReduced(reduced, fullRaster, plotCol, mapcode = names(fullRaster),
  ...)
}
\arguments{
\item{reduced}{\code{data.frame} or \code{data.table} that has at least one column
of codes that are represented in the \code{fullRaster}}

\item{fullRaster}{\code{RasterLayer} of codes used in \code{reduced} that represents
a spatial representation of the data}

\item{plotCol}{a character, length 1, with the name of the column in \code{reduced} that
whose value will be plotted}

\item{mapcode}{a character, length 1, with the name of the column in \code{reduced} that
is represented in \code{fullRaster}}
}
\value{
A \code{RasterLayer} of with same dimensions as \code{fullRaster} representing
\code{plotCol} spatially, according to the join between the \code{mapcodeAll} contained within
\code{reduced} and \code{fullRaster}
}
\description{
rasterizeReduced - convert reduced representation to full raster
}
\examples{
\dontrun{
Ras <- raster(extent(0,100,0,100), res=1)
fullRas <- randomPolygons(Ras, numTypes=5, speedup=2, minpatch=5)
names(fullRas) <- "mapcodeAll"
reducedDT <- data.table(mapcodeAll=1:10, communities=sample(1:1000,10), biomass=rnbinom(10,mu=4000,0.3))
biomass <- rasterizeReduced(reducedDT, fullRas, "mapcodeAll", "biomass")
communities <- rasterizeReduced(reducedDT, fullRas, "mapcodeAll", "communities")
setColors(communities) <- c("blue","orange","red")
Plot(biomass, communities, new=T)
}
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{raster}}
}

