% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{RasterStackNamed-class}
\alias{RasterStackNamed-class}
\title{RasterStackNamed class}
\description{
This new class extends \code{RasterStack} by adding a slot
called \code{name}.
}
\details{
Using the simulation visualization features in \code{Plot} requires
objects to be plotted to have a name set.
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the object.}
}}
\seealso{
\code{\link{RasterStack}}
}

