% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{RasterStackNamed}
\alias{RasterStackNamed}
\alias{RasterStackNamed,RasterStack-method}
\title{Create a new \code{RasterStackNamed} object.}
\usage{
RasterStackNamed(..., name)

\S4method{RasterStackNamed}{RasterStack}(..., name)
}
\arguments{
\item{...}{Additional arguments to \code{RasterStack}}

\item{name}{The name of the object.}
}
\value{
Returns an object of class \code{RasterStackNamed}.
}
\description{
Create a new \code{RasterStackNamed} object.
}
\author{
Eliot McIntire

Alex Chubaty
}
\seealso{
\code{\link{RasterStackNamed}}
}

