% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{rasterToMemory}
\alias{rasterToMemory}
\title{Read raster to memory}
\usage{
rasterToMemory(x, ...)
}
\arguments{
\item{x}{An object passed directly to the function raster (e.g., character string of a filename).}
}
\value{
A raster object whose values are stored in memory.
}
\description{
Wrapper to the \code{raster} function, that creates the raster object in
memory, even if it was read in from file.
}
\author{
Eliot McIntire

Alex Chubaty
}
\seealso{
\link{\code{raster}}
}

