% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\docType{methods}
\name{rasterToMemory}
\alias{rasterToMemory}
\alias{rasterToMemory,ANY-method}
\title{Read raster to memory}
\usage{
rasterToMemory(x, ...)

\S4method{rasterToMemory}{ANY}(x, ...)
}
\arguments{
\item{x}{An object passed directly to the function raster (e.g., character string of a filename).}

\item{...}{Additional arguments to \code{raster}.}
}
\value{
A raster object whose values are stored in memory.
}
\description{
Wrapper to the \code{raster} function, that creates the raster object in
memory, even if it was read in from file.
}
\seealso{
\code{\link{raster}}.
}
\author{
Eliot McIntire and Alex Chubaty
}
