% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-methods.R
\name{rbindlistDtDtpot}
\alias{rbindlistDtDtpot}
\title{Internal helper}
\usage{
rbindlistDtDtpot(dt, dtPotential, returnFrom)
}
\arguments{
\item{dt}{Data.table}

\item{dtPotential}{Data.table}

\item{returnFrom}{Logical}
}
\description{
Not for users. A function to setnames and rbindlist that is used 3 places in spread2.
}
\keyword{internal}
