% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runifC}
\alias{runifC}
\title{Rcpp Sugar version of runif}
\usage{
runifC(N)
}
\arguments{
\item{N}{Integer Vector}
}
\value{
A vector of uniform random numbers as per \code{runif}
}
\description{
Slightly faster than runif, and used a lot
}
