% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/numerical-comparisons.R
\docType{methods}
\name{\%>=\%}
\alias{\%!=\%}
\alias{\%<=\%}
\alias{\%==\%}
\alias{\%>=\%}
\title{Relational operators}
\usage{
x \%>=\% y

x \%<=\% y

x \%==\% y

x \%!=\% y
}
\arguments{
\item{x}{An object for which methods have been written.}

\item{y}{An object for which methods have been written.}
}
\value{
A logical vector indicating the result of the element by element comparison.
        The elements of shorter vectors are recycled as necessary.
}
\description{
Binary operators which allow the comparison of values in atomic vectors.
}
\details{
These are similar to their counterparts in \code{base}, except a tolerance
\code{epsilon} can be specified to account for floating point errors.
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{epsilon}}
}

