\docType{methods}
\name{ProbInit}
\alias{ProbInit}
\title{ProbInit}
\usage{
ProbInit(map, p = NULL, absolute = FALSE)
}
\arguments{
  \item{map}{Description of this.}

  \item{p}{Description of this.}

  \item{absolute}{Description of this.}
}
\value{
Decribe what it returns: fromto.

#
}
\description{
Details here.
}
\seealso{
\code{\link{print}} and \code{\link{cat}}
}

