% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc-methods.R
\docType{methods}
\name{rndstr}
\alias{rndstr}
\title{Generate random strings}
\usage{
rndstr(n = 1, len = 8, characterFirst = TRUE)
}
\arguments{
\item{n}{Number of strings to generate (default 1).}

\item{len}{Length of strings to generate (default 8).}

\item{characterFirst}{Logical, if TRUE, then a letter will be the first character of the string.
This is useful if this random string is being used for object ames.}
}
\value{
Character vector of random strings.
}
\description{
Generate a vector of random alphanumeric strings each of an arbitrary length.
}
\author{
Alex Chubaty
}

