% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{methods}
\name{robustDigest,simList-method}
\alias{robustDigest,simList-method}
\title{robustDigest for simList class objects}
\usage{
\S4method{robustDigest}{simList}(object, objects,
  compareRasterFileLength = 1e+06, algo = "xxhash64")
}
\description{
See \code{\link[reproducible]{robustDigest}}. This strips out stuff
from a simList class object that would make it otherwise not
reproducibly digestible between sessions, operating systems,
or machines. This will likely still not allow identical digest
results across R versions.
}
\seealso{
\code{\link[reproducible]{robustDigest}}
}
