% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SELES.R
\docType{methods}
\name{agentLocation}
\alias{agentLocation}
\title{SELES - Agent Location at initiation}
\usage{
agentLocation(map)
}
\arguments{
\item{map}{SpatialPoints*, SpatialPolygons* or Raster*}
}
\value{
Object of same class as provided as input. If a Raster*, then zeros are converted
to NAs
}
\description{
A SELES-like function to maintain conceptual backwards compatability with that simulation
tool. Sets the the location of the intiating agents.  THIS IS NOT FULLY IMPLEMENTED.
This is intended to ease transitinos from
\href{http://www.lfmi.uqam.ca/seles.htm}{SELES}.
You must know how to use SELES for these to be useful
}
\author{
Eliot McIntire
}

