% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SELES.R
\docType{methods}
\name{numAgents}
\alias{numAgents}
\title{Number of Agents to initiate}
\usage{
numAgents(N)
}
\arguments{
\item{N}{Number of agents to intitate}
}
\value{
A numeric, indicating number of agents to start
}
\description{
A SELES-like function to maintain conceptual backwards compatability with that simulation
tool. Sets the the number of agents to initiate. THIS IS NOT FULLY IMPLEMENTED.
}
\author{
Eliot McIntire
}

