% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SELES.R
\docType{methods}
\name{probInit}
\alias{probInit}
\title{Probability of Initiation}
\usage{
probInit(map, p = NULL, absolute = FALSE)
}
\arguments{
\item{map}{A \code{spatialObjects} object. Provides CRS and is related to probabilities.}

\item{p}{probability, provided as a numeric or raster}

\item{absolute}{logical. Is \code{p} absolute probabilities or relative?}
}
\value{
An RasterLayer with probabilities of initialization
}
\description{
A SELES-like function to maintain conceptual backwards compatability with that simulation
tool. Describes the probability of initiation of agents or events.  THIS IS NOT FULLY IMPLEMENTED.
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{print}} and \code{\link{cat}}
}

