% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SELES.R
\docType{methods}
\name{transitions}
\alias{transitions}
\title{SELES - Transitioning to next time step}
\usage{
transitions(p, agent)
}
\arguments{
\item{p}{realized probability of persisting (i.e., either 0 or 1)}

\item{agent}{SpatialPoints* object}
}
\value{
Returns new SpatialPoints* object with potentially fewer agents
}
\description{
A SELES-like function to maintain conceptual backwards compatability with that simulation
tool. Describes the probability of an agent successfully persisting until next
time step. THIS IS NOT FULLY IMPLEMENTED.
This is intended to ease transitinos from
\href{http://www.lfmi.uqam.ca/seles.htm}{SELES}.
You must know how to use SELES for these to be useful
}
\author{
Eliot McIntire
}

