% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{show,simList-method}
\alias{show,simList-method}
\title{Show an Object}
\usage{
\S4method{show}{simList}(object)
}
\arguments{
\item{object}{\code{simList}}
}
\description{
Show an Object
}

