\docType{methods}
\name{show}
\alias{show}
\alias{show,SimList}
\alias{show,agent}
\alias{show,mobileAgent}
\alias{show,mobileAgent-method}
\alias{show,pointAgent}
\alias{show,pointAgent-method}
\alias{show,spreadAgent}
\alias{show,spreadAgent-method}
\title{Show an Object}
\usage{
\S4method{show}{agent}(object)

\S4method{show}{spreadAgent}(object)

\S4method{show}{pointAgent}(object)

\S4method{show}{mobileAgent}(object)

\S4method{show}{SimList}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
Show an Object

Show an Object
}

