% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,RasterStackNamed-method}
\alias{show,RasterStackNamed-method}
\title{show RasterStackNamed}
\usage{
\S4method{show}{RasterStackNamed}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
show RasterStackNamed
}

