% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{show,simList-method}
\alias{show,simList-method}
\title{show simList}
\usage{
\S4method{show}{simList}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
show simList
}

