% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,SpatialPointsDataFrameNamed-method}
\alias{show,SpatialPointsDataFrameNamed-method}
\title{show SpatialPointsDataFrameNamed}
\usage{
\S4method{show}{SpatialPointsDataFrameNamed}(object)
}
\description{
show SpatialPointsDataFrameNamed
}

