% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,SpatialPointsNamed-method}
\alias{show,SpatialPointsNamed-method}
\title{show SpatialPointsNamed}
\usage{
\S4method{show}{SpatialPointsNamed}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
show SpatialPointsNamed
}

