% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,SpatialPolygonsDataFrameNamed-method}
\alias{show,SpatialPolygonsDataFrameNamed-method}
\title{show SpatialPolygonsDataFrameNamed}
\usage{
\S4method{show}{SpatialPolygonsDataFrameNamed}(object)
}
\description{
show SpatialPolygonsDataFrameNamed
}

