% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,SpatialPolygonsNamed-method}
\alias{show,SpatialPolygonsNamed-method}
\title{show SpatialPolygonsNamed}
\usage{
\S4method{show}{SpatialPolygonsNamed}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
show SpatialPolygonsNamed
}

