% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{sim.modules,SimList-method}
\alias{sim.modules,SimList-method}
\title{get list of simulation modules}
\usage{
\S4method{sim.modules}{SimList}(object)
}
\arguments{
\item{object}{A \code{SimList} simulation object.}
}
\description{
get list of simulation modules
}

