% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting-diagrams.R
\docType{methods}
\name{sim2gantt}
\alias{sim2gantt}
\alias{sim2gantt,simList,character-method}
\title{sim2gantt}
\usage{
sim2gantt(sim, startDate)

\S4method{sim2gantt}{simList,character}(sim, startDate)
}
\arguments{
\item{sim}{A \code{simList} object (typically corresponding to a
            completed simulation).}

\item{startDate}{A character representation of date in YYYY-MM-DD format.}
}
\value{
A list of data.frames
}
\description{
Internal function to convert the completed events list of a \code{simList}
object to a list of \code{data.frame}s suitable to pass to a call to
\code{\link[DiagrammeR]{mermaid}} to make a Gannt chart representing the
events in a completed simulation.
}
\author{
Alex Chubaty
}

