\docType{class}
\name{SimData}
\alias{SimData}
\alias{SimData-class}
\title{The SimData class}
\description{
This class store data used by and generated in a
simulation.
}
\details{
An object of class SimData contains: agents: custom list of
agents used in the simulation; maps: custom list of maps
loaded or generated by the simulation; stats: custom list
of stats collected during the simulation.
}
\section{Slots}{

\describe{
\item{\code{agents}}{The list of agents, as an agent class (or subclass).}

\item{\code{maps}}{The list(s) of maps, as raster class (raster package) or Spatial class (sp package).}

\item{\code{stats}}{The list(s) of statistics.}
}}
\note{
add additional notes here.
}

