% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{simDepends}
\alias{simDepends}
\alias{simDepends,simList-method}
\alias{simDepends<-}
\alias{simDepends<-,simList-method}
\title{Get and set simulation dependencies}
\usage{
simDepends(object)

\S4method{simDepends}{simList}(object)

simDepends(object) <- value

\S4method{simDepends}{simList}(object) <- value
}
\arguments{
\item{sim}{A \code{\link{simList}} object.}
}
\description{
Get and set simulation dependencies
}
\examples{
\dontrun{
  simDepends(sim)
}
}
\author{
Alex Chubaty
}
\seealso{
moduleDeps
}

