% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module-dependencies.R
\docType{class}
\name{simDeps-class}
\alias{simDeps-class}
\title{The \code{simDeps} class}
\description{
Defines all simulation dependencies for all modules within a SpaDES simulation.
}
\section{Slots}{

\describe{
\item{\code{dependencies}}{List of \code{\link{moduleDeps}} dependency objects.}
}}
\author{
Alex Chubaty
}

