% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{simParams}
\alias{simCheckpointFile}
\alias{simCheckpointFile,simEnv-method}
\alias{simCheckpointFile<-}
\alias{simCheckpointFile<-,simEnv-method}
\alias{simCheckpointInterval}
\alias{simCheckpointInterval,simEnv-method}
\alias{simCheckpointInterval<-}
\alias{simCheckpointInterval<-,simEnv-method}
\alias{simEnv-accessors-params}
\alias{simFileList}
\alias{simFileList,simEnv-method}
\alias{simFileList<-}
\alias{simFileList<-,simEnv-method}
\alias{simGlobals}
\alias{simGlobals,simEnv-method}
\alias{simGlobals<-}
\alias{simGlobals<-,simEnv-method}
\alias{simGlobalsOutputPath}
\alias{simGlobalsOutputPath,simEnv-method}
\alias{simGlobalsOutputPath<-}
\alias{simGlobalsOutputPath<-,simEnv-method}
\alias{simParams}
\alias{simParams,simEnv-method}
\alias{simParams<-}
\alias{simParams<-,simEnv-method}
\alias{simProgressGraphical}
\alias{simProgressGraphical,simEnv-method}
\alias{simProgressGraphical<-}
\alias{simProgressGraphical<-,simEnv-method}
\alias{simProgressInterval}
\alias{simProgressInterval,simEnv-method}
\alias{simProgressInterval<-}
\alias{simProgressInterval<-,simEnv-method}
\title{Get and set simulation parameters.}
\usage{
simParams(object)

\S4method{simParams}{simEnv}(object)

simParams(object) <- value

\S4method{simParams}{simEnv}(object) <- value

simCheckpointFile(object)

\S4method{simCheckpointFile}{simEnv}(object)

simCheckpointFile(object) <- value

\S4method{simCheckpointFile}{simEnv}(object) <- value

simCheckpointInterval(object)

\S4method{simCheckpointInterval}{simEnv}(object)

simCheckpointInterval(object) <- value

\S4method{simCheckpointInterval}{simEnv}(object) <- value

simFileList(object)

\S4method{simFileList}{simEnv}(object)

simFileList(object) <- value

\S4method{simFileList}{simEnv}(object) <- value

simProgressGraphical(object)

\S4method{simProgressGraphical}{simEnv}(object)

simProgressGraphical(object) <- value

\S4method{simProgressGraphical}{simEnv}(object) <- value

simProgressInterval(object)

\S4method{simProgressInterval}{simEnv}(object)

simProgressInterval(object) <- value

\S4method{simProgressInterval}{simEnv}(object) <- value

simGlobals(object)

\S4method{simGlobals}{simEnv}(object)

simGlobals(object) <- value

\S4method{simGlobals}{simEnv}(object) <- value

simGlobalsOutputPath(object)

\S4method{simGlobalsOutputPath}{simEnv}(object)

simGlobalsOutputPath(object) <- value

\S4method{simGlobalsOutputPath}{simEnv}(object) <- value
}
\arguments{
\item{object}{A \code{simEnv} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Accessor functions for the \code{params} slot of a \code{simList} object
and its elements, within a \code{simEnv} object.
Additonal methods are provided to access core module and global parameters:
Commonly used
\tabular{ll}{
   \code{simGlobals} \tab List of global simulation parameters.\cr
   \code{simParams} \tab Nested list of all simulation parameter.\cr
}
Advanced use
\tabular{lll}{
   Accessor method \tab Module \tab Description \cr
   \code{simCheckpointFile} \tab \code{.checkpoint} \tab Name of the checkpoint file. (advanced)\cr
   \code{simCheckpointInterval} \tab \code{.checkpoint} \tab The simulation checkpoint interval. (advanced)\cr
   \code{simGlobalsOutputPath} \tab \code{NA} \tab Global simulation output path. (advanced)\cr
   \code{simFileList} \tab \code{.load} \tab List of files to load for the simulation. (advanced)\cr
   \code{simObjectsLoaded} \tab \code{.load} \tab List of loaded simulation objects. (advanced)\cr
   \code{simProgressGraphical} \tab \code{.progress} \tab Type of graphical progress bar used. (advanced)\cr
   \code{simProgressInterval} \tab \code{.progress} \tab Interval for the progress bar. (advanced)\cr
}
}
\details{
Currently, only get and set methods are defined. Subset methods are not.
}
\seealso{
\code{\link{simList-class}},
         \code{\link{simEnv-accessors-modules}},
         \code{\link{simEnv-accessors-events}},
         \code{\link{simEnv-accessors-times}}.
}

