% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\docType{class}
\name{simEnv-class}
\alias{simEnv}
\alias{simEnv-class}
\title{The \code{simEnv} class}
\description{
This class is simply an environment that contains a \code{simList} object.
}
\author{
Alex Chubaty
}

