% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{simEvents}
\alias{<-}
\alias{simEvents}
\alias{simEvents,simList-method}
\alias{simEvents<-}
\title{Accessor methods for \code{simList} object slots}
\usage{
simEvents(object)

\S4method{simEvents}{simList}(object)

simEvents(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Currently, only get and set methods are defined. Subset methods are not.

get the simulation event queue

set the simulation event queue

set the simulation event queue
}
\details{
Additonal methods are provided to access the current, start, and stop times of the
simulation: \code{simCurrentTime(sim)}, \code{simStartTime(sim)}, \code{simStopTime(sim)}.
}
\author{
Alex Chubaty
}

