% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{simEventsCompleted}
\alias{<-}
\alias{simEventsCompleted}
\alias{simEventsCompleted,simList-method}
\alias{simEventsCompleted<-}
\title{Accessor methods for \code{simList} object slots}
\usage{
simEventsCompleted(object)

\S4method{simEventsCompleted}{simList}(object)

simEventsCompleted(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Currently, only get and set methods are defined. Subset methods are not.

get the simulation completed events list

set the simulation completed events list

set the simulation completed events list
}
\details{
Additonal methods are provided to access the current, start, and stop times of the
simulation: \code{simCurrentTime(sim)}, \code{simStartTime(sim)}, \code{simStopTime(sim)}.
}
\author{
Eliot McIntire
}

