% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{simFileList}
\alias{<-}
\alias{simFileList}
\alias{simFileList,simList-method}
\alias{simFileList<-}
\title{Accessor methods for \code{simList} object slots}
\usage{
simFileList(object)

\S4method{simFileList}{simList}(object)

simFileList(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Currently, only get and set methods are defined. Subset methods are not.

get .loadFileList from simulation parameters

set .loadFileList in simulation parameters

set .loadFileList in simulation parameters
}
\details{
Additonal methods are provided to access the current, start, and stop times of the
simulation: \code{simCurrentTime(sim)}, \code{simStartTime(sim)}, \code{simStopTime(sim)}.
}
\author{
Alex Chubaty
}

