% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{simGlobalsOutputPath}
\alias{simGlobalsOutputPath}
\alias{simGlobalsOutputPath,simList-method}
\alias{simGlobalsOutputPath<-}
\alias{simGlobalsOutputPath<-,simList-method}
\title{Get and set the output path from global simulation parmeters.}
\usage{
simGlobalsOutputPath(object)

\S4method{simGlobalsOutputPath}{simList}(object)

simGlobalsOutputPath(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Get and set the output path from global simulation parmeters.

get .globals$.outputPath from simulation parameters

set .globals$.outputPath in simulation parameters

set .globals$.outputPath in simulation parameters
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{simCheckpointFile}} for accessing the name of the checkpoint file;
         \code{\link{simCheckpointInterval}} for accessing the simulation checkpoint interval;
         \code{\link{simGlobals}} for accessing the global simulation parameters;
         \code{\link{simModulesLoaded}} for accessing the list of loaded simulation modules;
         \code{\link{simObjectsLoaded}} for accessing the list of loaded simulation objects;
         \code{\link{simModules}} for accessing the list of simulation modules to be loaded;
         \code{\link{simParams}} for accessing the list of simulation parameters;
         \code{\link{simEvents}} for accessing the scheduled simulation event queue;
         \code{\link{simCompleted}} for accessing the list of completed simulation events;
         \code{\link{simTimes}} for accessing the list of simulation times
         (\code{simCurrentTime}, \code{simStartTime}, and \code{simStopTime}
         to access the simulation current, start, and stop times individually).
}

