% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{simInit}
\alias{simInit}
\alias{simInit,character,list,list,list,character-method}
\alias{simInit,character,list,list,list,missing-method}
\alias{simInit,missing,list,list,list,character-method}
\alias{simInit,missing,list,list,list,missing-method}
\title{Initialize a new simulation}
\usage{
simInit(simname, times, params, modules, path)

\S4method{simInit}{character,list,list,list,character}(simname, times, params,
  modules, path)

\S4method{simInit}{character,list,list,list,missing}(simname, times, params,
  modules)

\S4method{simInit}{missing,list,list,list,character}(times, params, modules,
  path)

\S4method{simInit}{missing,list,list,list,missing}(times, params, modules)
}
\arguments{
\item{times}{A named list of numeric simulation start and stop times
       (e.g., \code{times=list(start=0.0, stop=10.0)}).}

\item{params}{A named list of simulation parameters and their values.}

\item{modules}{A named list of character strings specfying the names
of modules to be loaded for the simulation. **Note:** the module name
should correspond to the R source file from which the module is loaded.
Example: a module named "caribou" will be sourced form the file \code{caribou.R},
located at the specified \code{path} (see below).}

\item{path}{An optional character string specifying the location of the module source files.
If no path is specified, it defaults to the current working directory.}
}
\value{
A \code{simList} simulation object, pre-initialized from values specified
in the arguments supplied.
}
\description{
Create a new simulation object, preloaded with parameters,
modules, times, etc.

simInit

simInit

simInit

simInit
}
\details{
Based on code from Matloff (2011) ch. 7.8.3 on discrete event simulation:
- implemented using S4 classes and methods;
- uses `data.table` instead of `data.frame` (which is much faster);
- implemented in a more modular fashion so it's easier
  to add submodules to the simulation.
}
\examples{
\dontrun{mySim <- simInit(times=list(start=0.0, stop=10.0), params=list(Ncaribou=100),
modules=list("habitat", "caribou"), path="/path/to/my/modules/")}
\dontrun{mySim}
}
\author{
Alex Chubaty
}
\references{
Matloff, N. (2011). The Art of R Programming (373 pp.). San Fransisco, CA: No Starch Press, Inc.. Retrieved from \url{http://www.nostarch.com/artofr.htm}
}
\seealso{
\code{\link{doSim}}.
}

