% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{simObjects}
\alias{simObjects}
\alias{simObjects,simList-method}
\alias{simObjects<-}
\alias{simObjects<-,simList-method}
\title{Show objects referenced in the simulation environment}
\usage{
simObjects(object, ...)

\S4method{simObjects}{simList}(object, ...)

simObjects(object) <- value

\S4method{simObjects}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{...}{arguments passed to \code{ls}, allowing, e.g. \code{all.names=TRUE}}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets a list of objects in the \code{simList} environment.
}
\description{
Show objects referenced in the simulation environment
}

