% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{simEnv}
\alias{simEnv}
\alias{simEnv,simList-method}
\alias{simEnv<-}
\alias{simEnv<-,simList-method}
\alias{simList-accessors-simEnv}
\title{Simulation environment}
\usage{
simEnv(object)

\S4method{simEnv}{simList}(object)

simEnv(object) <- value

\S4method{simEnv}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Accessor functions for the \code{.envir} slot in a \code{simList} object.
These are included for advanced users.
}
\details{
Currently, only get and set methods are defined. Subset methods are not.
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{simList-class}},
         \code{\link{simList-accessors-modules}},
         \code{\link{simList-accessors-params}},
         \code{\link{simList-accessors-events}},
         \code{\link{simList-accessors-times}}.
}

