% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{simTimes}
\alias{simCurrentTime}
\alias{simCurrentTime,simList-method}
\alias{simCurrentTime<-}
\alias{simCurrentTime<-,simList-method}
\alias{simList-accessors-times}
\alias{simStartTime}
\alias{simStartTime,simList-method}
\alias{simStartTime<-}
\alias{simStartTime<-,simList-method}
\alias{simStopTime}
\alias{simStopTime,simList-method}
\alias{simStopTime<-}
\alias{simStopTime<-,simList-method}
\alias{simTimes}
\alias{simTimes,simList-method}
\alias{simTimes<-}
\alias{simTimes<-,simList-method}
\title{Get and set simulation times.}
\usage{
simTimes(object)

\S4method{simTimes}{simList}(object)

simTimes(object) <- value

\S4method{simTimes}{simList}(object) <- value

simCurrentTime(object)

\S4method{simCurrentTime}{simList}(object)

simCurrentTime(object) <- value

\S4method{simCurrentTime}{simList}(object) <- value

simStartTime(object)

\S4method{simStartTime}{simList}(object)

simStartTime(object) <- value

\S4method{simStartTime}{simList}(object) <- value

simStopTime(object)

\S4method{simStopTime}{simList}(object)

simStopTime(object) <- value

\S4method{simStopTime}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Accessor functions for the \code{simtimes} slot of a \code{simList} object
and its elements.
Additonal methods are provided to access the current, start, and stop times
of the simulation.
\tabular{ll}{
   \code{simCurrentTime} \tab Current simulation time.\cr
   \code{simStartTime} \tab Simulation start time.\cr
   \code{simStopTime} \tab Simulation stop time.\cr
   \code{simTimes} \tab List of all simulation times (current, start, stop).\cr
}
}
\details{
Currently, only get and set methods are defined. Subset methods are not.
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{simList-class}},
         \code{\link{simList-accessors-params}},
         \code{\link{simList-accessors-modules}},
         \code{\link{simList-accessors-events}}.
}

