% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{simModules}
\alias{simCheckpointFile}
\alias{simCheckpointFile,simList-method}
\alias{simCheckpointFile<-}
\alias{simCheckpointFile<-,simList-method}
\alias{simCheckpointInterval}
\alias{simCheckpointInterval,simList-method}
\alias{simCheckpointInterval<-}
\alias{simCheckpointInterval<-,simList-method}
\alias{simCompleted}
\alias{simCompleted,simList-method}
\alias{simCompleted<-}
\alias{simCompleted<-,simList-method}
\alias{simDepends}
\alias{simDepends,simList-method}
\alias{simDepends<-}
\alias{simDepends<-,simList-method}
\alias{simEvents}
\alias{simEvents,simList-method}
\alias{simEvents<-}
\alias{simEvents<-,simList-method}
\alias{simFileList}
\alias{simFileList,simList-method}
\alias{simFileList<-}
\alias{simFileList<-,simList-method}
\alias{simGlobals}
\alias{simGlobals,simList-method}
\alias{simGlobals<-}
\alias{simGlobals<-,simList-method}
\alias{simGlobalsOutputPath}
\alias{simGlobalsOutputPath,simList-method}
\alias{simGlobalsOutputPath<-}
\alias{simGlobalsOutputPath<-,simList-method}
\alias{simModules}
\alias{simModules,simList-method}
\alias{simModules<-}
\alias{simModules<-,simList-method}
\alias{simModulesLoadOrder}
\alias{simModulesLoadOrder,simList-method}
\alias{simModulesLoadOrder<-}
\alias{simModulesLoadOrder<-,simList-method}
\alias{simModulesLoaded}
\alias{simModulesLoaded,simList-method}
\alias{simModulesLoaded<-}
\alias{simModulesLoaded<-,simList-method}
\alias{simObjectsLoaded}
\alias{simObjectsLoaded,simList-method}
\alias{simObjectsLoaded<-}
\alias{simObjectsLoaded<-,simList-method}
\alias{simParams}
\alias{simParams,simList-method}
\alias{simParams<-}
\alias{simParams<-,simList-method}
\alias{simProgressGraphical}
\alias{simProgressGraphical,simList-method}
\alias{simProgressGraphical<-}
\alias{simProgressGraphical<-,simList-method}
\alias{simProgressInterval}
\alias{simProgressInterval,simList-method}
\alias{simProgressInterval<-}
\alias{simProgressInterval<-,simList-method}
\title{Get and set slots and elements in a simList object}
\usage{
simModules(object)

\S4method{simModules}{simList}(object)

simModules(object) <- value

\S4method{simModules}{simList}(object) <- value

simModulesLoadOrder(object)

\S4method{simModulesLoadOrder}{simList}(object)

simModulesLoadOrder(object) <- value

\S4method{simModulesLoadOrder}{simList}(object) <- value

simModulesLoaded(object)

\S4method{simModulesLoaded}{simList}(object)

simModulesLoaded(object) <- value

\S4method{simModulesLoaded}{simList}(object) <- value

simObjectsLoaded(object)

\S4method{simObjectsLoaded}{simList}(object)

simObjectsLoaded(object) <- value

\S4method{simObjectsLoaded}{simList}(object) <- value

simParams(object)

\S4method{simParams}{simList}(object)

simParams(object) <- value

\S4method{simParams}{simList}(object) <- value

simCheckpointFile(object)

\S4method{simCheckpointFile}{simList}(object)

simCheckpointFile(object) <- value

\S4method{simCheckpointFile}{simList}(object) <- value

simCheckpointInterval(object)

\S4method{simCheckpointInterval}{simList}(object)

simCheckpointInterval(object) <- value

\S4method{simCheckpointInterval}{simList}(object) <- value

simFileList(object)

\S4method{simFileList}{simList}(object)

simFileList(object) <- value

\S4method{simFileList}{simList}(object) <- value

simProgressGraphical(object)

\S4method{simProgressGraphical}{simList}(object)

simProgressGraphical(object) <- value

\S4method{simProgressGraphical}{simList}(object) <- value

simProgressInterval(object)

\S4method{simProgressInterval}{simList}(object)

simProgressInterval(object) <- value

\S4method{simProgressInterval}{simList}(object) <- value

simGlobals(object)

\S4method{simGlobals}{simList}(object)

simGlobals(object) <- value

\S4method{simGlobals}{simList}(object) <- value

simGlobalsOutputPath(object)

\S4method{simGlobalsOutputPath}{simList}(object)

simGlobalsOutputPath(object) <- value

\S4method{simGlobalsOutputPath}{simList}(object) <- value

simEvents(object)

\S4method{simEvents}{simList}(object)

simEvents(object) <- value

\S4method{simEvents}{simList}(object) <- value

simCompleted(object)

\S4method{simCompleted}{simList}(object)

simCompleted(object) <- value

\S4method{simCompleted}{simList}(object) <- value

simDepends(object)

\S4method{simDepends}{simList}(object)

simDepends(object) <- value

\S4method{simDepends}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Currently, only get and set methods are defined. Subset methods are not.
}
\details{
\code{\link{simCheckpointFile}} for accessing the name of the checkpoint file;
\code{\link{simCheckpointInterval}} for accessing the simulation checkpoint interval;
\code{\link{simGlobals}} for accessing the global simulation parameters;
\code{\link{simGlobalsOutputPath}} for accessing the global simulation output path;
\code{\link{simModules}} for accessing the list of simulation modules to be loaded;
\code{\link{simModulesLoaded}} for accessing the list of loaded simulation modules;
\code{\link{simObjectsLoaded}} for accessing the list of loaded simulation objects;
\code{\link{simParams}} for accessing the list of simulation parameters;
\code{\link{simProgressGraphical}} for accessing the graphical progress bar parameter;
\code{\link{simProgressInterval}} for accessing the interval for the progress bar;
\code{\link{simEvents}} for accessing the scheduled simulation event queue;
\code{\link{simCompleted}} for accessing the list of completed simulation events;
\code{\link{simTimes}} for accessing the list of simulation times
(\code{simCurrentTime}, \code{simStartTime}, and \code{simStopTime}
to access the simulation current, start, and stop times individually).
}
\author{
Alex Chubaty
}

