% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{SimList}
\alias{SimList}
\alias{SimList-class}
\title{The \code{SimList} class}
\description{
This class contains the minimum components of a simulation.
}
\section{Slots}{

\describe{
\item{\code{.loaded}}{List of character names specifying which modules are currently loaded.}

\item{\code{modules}}{List of character names specifying which modules to load.}

\item{\code{params}}{Named list of potentially other lists specifying simulation parameters.}

\item{\code{events}}{The list of scheduled events, as a data.table class. This is implemented
                 such that the data.table is always sorted (keyed) by time, making it easy
                 to insert new events into the table.}

\item{\code{simtimes}}{List of numerical values describing the simulation start and stop timos,
                 and the current simulation time.}

\item{\code{debug}}{Logical value specifying whether to run simulation in debugging mode.}
}}
\note{
Each event is represented by a data.table row consisting of:
         event.time: the time the event is to occur;
         module.name: the module from which the event is taken;
         event.type: a character string for the programmer-defined event type.
}
\seealso{
\code{\link{data.table}}
}

