% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{[[}
\alias{$}
\alias{$,simList-method}
\alias{[[}
\alias{[[,simList-method}
\title{Extract objects from the simulation environment}
\usage{
\S4method{[[}{simList,ANY,ANY}(x, i, j, ..., drop)

\S4method{$}{simList}(x, name)
}
\description{
Extract objects from the simulation environment
}

