% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{inputs}
\alias{inputArgs}
\alias{inputArgs,simList-method}
\alias{inputArgs<-}
\alias{inputArgs<-,simList-method}
\alias{inputs}
\alias{inputs,simList-method}
\alias{inputs<-}
\alias{inputs<-,simList-method}
\alias{outputArgs}
\alias{outputArgs,simList-method}
\alias{outputArgs<-}
\alias{outputArgs<-,simList-method}
\alias{outputs}
\alias{outputs,simList-method}
\alias{outputs<-}
\alias{outputs<-,simList-method}
\title{Inputs and outputs - adding file-based inputs and saving files}
\usage{
inputs(object)

\S4method{inputs}{simList}(object)

inputs(object) <- value

\S4method{inputs}{simList}(object) <- value

outputs(object)

\S4method{outputs}{simList}(object)

outputs(object) <- value

\S4method{outputs}{simList}(object) <- value

inputArgs(object)

\S4method{inputArgs}{simList}(object)

inputArgs(object) <- value

\S4method{inputArgs}{simList}(object) <- value

outputArgs(object)

\S4method{outputArgs}{simList}(object)

outputArgs(object) <- value

\S4method{outputArgs}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\description{
These functions are one of two mechanisms to add the information
about which input files to load into a \code{spades} and the information
about which output files to save.
The other way is to pass them as arguments to a simInit function call.

\code{inputArgs} and \code{outputArgs} are ways to specify any
arguments that are needed for file loading and file saving. This
is still somewhat experimental.
}
\details{
\code{inputs} accepts a data.frame, with 6 columns. Currently, only one is required.
Columns are \code{objectName} (required, character),
\code{file} (character), \code{fun} (character),
\code{package} (character),
\code{interval} (numeric), and \code{loadTime} (numeric). See ii-modules vignette for details on
these columns.

\code{outputs} accepts a data.frame, with 5 columns. Currently,
only one is required.
Columns are \code{objectName} (character, required),
\code{file} (character), \code{fun} (character),
\code{package} (character), and \code{saveTime} (numeric). See ii-modules vignette for details on
these columns.
}
\seealso{
ii-modules vignette
}

