% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{inputPath}
\alias{inputPath}
\alias{inputPath,simList-method}
\alias{inputPath<-}
\alias{inputPath<-,simList-method}
\alias{modulePath}
\alias{modulePath,simList-method}
\alias{modulePath<-}
\alias{modulePath<-,simList-method}
\alias{paths}
\alias{paths,simList-method}
\alias{paths<-}
\alias{paths<-,simList-method}
\title{Specify paths for modules, inputs and outputs}
\usage{
inputPath(object)

\S4method{inputPath}{simList}(object)

inputPath(object) <- value

\S4method{inputPath}{simList}(object) <- value

modulePath(object)

\S4method{modulePath}{simList}(object)

modulePath(object) <- value

\S4method{modulePath}{simList}(object) <- value

paths(object)

\S4method{paths}{simList}(object)

paths(object) <- value

\S4method{paths}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\description{
These are ways to add or access the file paths used by spades. There
are three file paths: modulePath, inputPath, outputPath. Each has a function
to get or set the value in a simList object.
}

