% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{[[}
\alias{$}
\alias{$<-,simList-method}
\alias{[[}
\alias{[[<-,simList-method}
\title{Replace objects referenced in the simulation environment}
\usage{
\S4method{[[}{simList,ANY,ANY}(x, i) <- value

\S4method{$}{simList}(x, name) <- value
}
\description{
Replace objects referenced in the simulation environment
}

