% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{class}
\name{simList_-class}
\alias{simList_}
\alias{simList_-class}
\title{The \code{simList_} class}
\description{
Internal use only. Used when saving/loading a \code{simList}.
}
\details{
This is identical to class \code{simList}, except that the \code{.envir} slot
is replaced by a \code{.list} containing a list to store the objects from the
environment contained within the \code{simList}.
Saving/loading a list behaves more reliably than saving/loading an environment.
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{simList}}
}

