% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{simModulesLoadOrder}
\alias{simModulesLoadOrder}
\alias{simModulesLoadOrder,simList-method}
\alias{simModulesLoadOrder<-}
\alias{simModulesLoadOrder<-,simList-method}
\title{Get and set the module load order for the simulation.}
\usage{
simModulesLoadOrder(object)

\S4method{simModulesLoadOrder}{simList}(object)

\S4method{simModulesLoadOrder}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Currently, only get and set methods are defined. Subset methods are not.
}
\author{
Alex Chubaty
}

