% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{.simReqdPkgs}
\alias{.simReqdPkgs}
\alias{.simReqdPkgs,simList-method}
\title{Get simulation package dependencies}
\usage{
.simReqdPkgs(sim)

\S4method{.simReqdPkgs}{simList}(sim)
}
\arguments{
\item{sim}{A \code{simList} object.}
}
\value{
A sorted character vector of package names.
}
\description{
Internal function.
}
\author{
Alex Chubaty
}

