% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{simStartTime}
\alias{<-}
\alias{simStartTime}
\alias{simStartTime,simList-method}
\alias{simStartTime<-}
\title{Accessor methods for \code{simList} object slots}
\usage{
simStartTime(object)

\S4method{simStartTime}{simList}(object)

simStartTime(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Currently, only get and set methods are defined. Subset methods are not.

get the simulation start time

set the simulation start time

set the simulation start time
}
\details{
Additonal methods are provided to access the current, start, and stop times of the
simulation: \code{simCurrentTime(sim)}, \code{simStartTime(sim)}, \code{simStopTime(sim)}.
}
\author{
Alex Chubaty
}

