\docType{methods}
\name{simTimes,simList-method}
\alias{<-}
\alias{simTimes,simList-method}
\alias{simTimes<-}
\title{get list of simulation times}
\usage{
\S4method{simTimes}{simList}(object)

simTimes(object) <- value
}
\arguments{
  \item{object}{A \code{simList} simulation object.}

  \item{value}{The object to be stored at the slot.}
}
\description{
get list of simulation times

set list of simulation times

set list of simulation times
}

