% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{simTimes}
\alias{simCurrentTime}
\alias{simCurrentTime,simList-method}
\alias{simCurrentTime<-}
\alias{simCurrentTime<-,simList-method}
\alias{simStartTime}
\alias{simStartTime,simList-method}
\alias{simStartTime<-}
\alias{simStartTime<-,simList-method}
\alias{simStopTime}
\alias{simStopTime,simList-method}
\alias{simStopTime<-}
\alias{simStopTime<-,simList-method}
\alias{simTimes}
\alias{simTimes,simList-method}
\alias{simTimes<-}
\alias{simTimes<-,simList-method}
\title{Get and set simulation times.}
\usage{
simTimes(object)

\S4method{simTimes}{simList}(object)

simTimes(object) <- value

\S4method{simTimes}{simList}(object) <- value

simCurrentTime(object)

\S4method{simCurrentTime}{simList}(object)

simCurrentTime(object) <- value

\S4method{simCurrentTime}{simList}(object) <- value

simStartTime(object)

\S4method{simStartTime}{simList}(object)

simStartTime(object) <- value

\S4method{simStartTime}{simList}(object) <- value

simStopTime(object)

\S4method{simStopTime}{simList}(object)

simStopTime(object) <- value

\S4method{simStopTime}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Currently, only get and set methods are defined. Subset methods are not.
}
\details{
Additonal methods are provided to access the current, start, and stop times of the
simulation: \code{simCurrentTime}, \code{simStartTime}, and \code{simStopTime}
to access the simulation current, start, and stop times individually.
}
\author{
Alex Chubaty
}

