% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{sim.modules}
\alias{<-}
\alias{currentTime}
\alias{currentTime,SimList-method}
\alias{currentTime<-}
\alias{sim.debug}
\alias{sim.debug,SimList-method}
\alias{sim.debug<-}
\alias{sim.events}
\alias{sim.events,SimList-method}
\alias{sim.events<-}
\alias{sim.loaded}
\alias{sim.loaded,SimList-method}
\alias{sim.loaded<-}
\alias{sim.modules}
\alias{sim.modules,SimList-method}
\alias{sim.modules<-}
\alias{sim.params}
\alias{sim.params,SimList-method}
\alias{sim.params<-}
\alias{sim.times}
\alias{sim.times,SimList-method}
\alias{sim.times<-}
\alias{startTime}
\alias{startTime,SimList-method}
\alias{startTime<-}
\alias{stopTime}
\alias{stopTime,SimList-method}
\alias{stopTime<-}
\title{Accessor methods for \code{SimList} object slots}
\usage{
sim.modules(object)

\S4method{sim.modules}{SimList}(object)

sim.modules(object) <- value

sim.loaded(object)

\S4method{sim.loaded}{SimList}(object)

sim.loaded(object) <- value

sim.params(object)

\S4method{sim.params}{SimList}(object)

sim.params(object) <- value

sim.times(object)

\S4method{sim.times}{SimList}(object)

sim.times(object) <- value

currentTime(object)

\S4method{currentTime}{SimList}(object)

currentTime(object) <- value

startTime(object)

\S4method{startTime}{SimList}(object)

startTime(object) <- value

stopTime(object)

\S4method{stopTime}{SimList}(object)

stopTime(object) <- value

sim.events(object)

\S4method{sim.events}{SimList}(object)

sim.events(object) <- value

sim.debug(object)

\S4method{sim.debug}{SimList}(object)

sim.debug(object) <- value
}
\arguments{
\item{object}{A \code{SimList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{SimList} object.
}
\description{
Currently, only get and set methods are defined. Subset methods are not.

get list of simulation modules

set list of simulation modules

set list of simulation modules

get list of loaded simulation modules

get list of loaded simulation modules

set list of loaded simulation modules

set list of loaded simulation modules

get list of simulation parameters

get list of simulation parameters

set list of simulation parameters

set list of simulation parameters

get list of simulation times

get list of simulation times

set list of simulation times

set list of simulation times

get the current simulation time

get the current simulation time

set the current simulation time

set the current simulation time

get the simulation start time

get the simulation start time

set the simulation start time

set the simulation start time

get the simulation stop time

get the simulation stop time

set the simulation stop time

set the simulation stop time

get the simulation event queue

get the simulation event queue

set the simulation event queue

set the simulation event queue

get the simulation debug toggle

get the simulation debug toggle

set the simulation debug toggle

set the simulation debug toggle
}
\details{
Additonal methods are provided to access the current, start, and stop times of the
simulation: \code{currentTime(sim)}, \code{startTime(sim)}, \code{stopTime(sim)}.
}

