\docType{methods}
\name{dosim,SimList,logical-method}
\alias{dosim,SimList,logical-method}
\title{dosim}
\usage{
\S4method{dosim}{SimList,logical}(sim, debug)
}
\arguments{
  \item{sim}{A \code{SimList} simulation object.}

  \item{debug}{Optional logical flag determines whether sim
  debug info will be printed. If not specified, the default
  is \code{debug=FALSE}.}
}
\description{
dosim
}

