% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-methods.R
\docType{methods}
\name{sortDotsFirst}
\alias{sortDotsFirst}
\title{Sort a any named object with dotted names first}
\usage{
sortDotsFirst(obj)
}
\arguments{
\item{obj}{An arbitrary R object for which a \code{names} function
returns a character vector.}
}
\value{
The same object as \code{obj}, but sorted with .objects first.
}
\description{
Internal use only. This exists so Windows and Linux machines can have
the same order after a sort.
}
\author{
Eliot McIntire
}
\keyword{internal}

