% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spades-package.R
\name{SpaDES-maps}
\alias{SpaDES-maps}
\title{Dummy maps included with SpaDES}
\description{
All 5 maps included with SpaDES are randomly generated maps created by \code{GaussMap()}
}
\details{
DEM.tif: converted to a a small number of
discrete levels (in 100m hypothetical units)

habitatQuality.tif: made to look like a continuous habitat
surface, rescaled to 0 to 1

forestAge.tif: rescaled to possible
forest ages in a boreal forest setting

forestCover.tif: rescaled to possible
forest cover in a boreal forest setting

percentPine.tif: rescaled to percentages.
}

