% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spades-package.R
\docType{package}
\name{spades-package}
\alias{SpaDES}
\alias{spades-package}
\title{Overview of the functions in the SpaDES package}
\description{
This package allows implementation a variety of simulation-type models, with a focus on spatially
explicit models. The core simulation components are built upon
a discrete event simulation framework that facilitates modularity, and easily
enables the user to include additional functionality by running user-built
simulation modules. Included are numerous tools to visualize various spatial data formats, as well
as non-spatial data.

Bug reports: https://github.com/PredictiveEcology/SpaDES/issues
}
\section{Plotting}{

There are several user-accessible plotting functions that are optimized for modularity
and speed of plotting:

\tabular{ll}{
  \code{\link{Plot}} \tab The main function. Uses \code{\link[raster:Raster-class]{Raster*}},
    \code{\link[sp:SpatialPoints-class]{SpatialPoints*}}, \code{\link[sp:SpatialPolygons-class]{SpatialPolygons*}},
    \code{gg}, \code{\link[graphics]{hist}}, \code{igraph} objects\cr
  \code{\link{clickValues}} \tab Is for extracting values from a raster object at the mouse click location(s)\cr
  \code{\link{clickExtent}} \tab Used to zoom into a raster or polygon map that was plotted with \code{\link{Plot}}\cr
  --------------------------- \tab ------------------------------------------------------------------------------------------ \cr
}
}
\author{
Alex M. Chubaty \email{achubaty@nrcan.gc.ca}

Eliot J. B. McIntire \email{Eliot.McIntire@nrcan.gc.ca}
}
\keyword{package}

