% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.pkgs}
\alias{.pkgs}
\title{Core SpaDES packages}
\format{An object of class \code{character} of length 6.}
\usage{
.pkgs
}
\description{
Packages of the \code{SpaDES} ecosystem.
}
\keyword{internal}
