% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\docType{data}
\name{.spadesEnv}
\alias{.spadesEnv}
\title{The SpaDES environment}
\format{An object of class \code{environment} of length 1.}
\usage{
.spadesEnv
}
\description{
Environment used internally to store internal package objects and methods.
}
\keyword{datasets}

