% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{clickValues}
\alias{.clickCoord}
\alias{clickCoordinates}
\alias{clickValues}
\title{Mouse interactions with Plots}
\usage{
clickValues(n = 1)

clickCoordinates(n = 1)

.clickCoord(X, n = 1, gl = NULL)
}
\arguments{
\item{n}{The number of mouse clicks to do}

\item{X}{The raster object whose values will be returned where mouse clicks occur}

\item{gl}{An object created by a call to \code{grid.locator}}
}
\value{
\code{clickValues} returns the layer names and values at the clicked points.
\code{clickExtent} invisibly returns the extent object, and optionally plots it
in a new device window. \code{clickCoordinates} returns the xy coordinates in
the units of the plot clicked on.
}
\description{
These functions use \code{grid.locator}. The primary two user-level functions are
\code{clickValues} and \code{clickExtent}. These functions automatically select
the correct viewport (i.e., map) where the mouse clicks occured so the user
does not have to manually specify which map is being clicked on.
}
\details{
\code{clickValues} is equivalent to running \code{X[SpatialPoints(locator(n))]}, where
X is the raster being clicked on, in base graphics. This function determines which place in the
grid.layout was clicked and makes all appropriate calculations to determine the value
on the raster(s) at that or those location(s)

\code{clickExtent} is for drawing an extent with two mouse clicks on a given Plotted map.

\code{clickCoordinates} is the workhorse function that determines which plot has been
clicked on and passes this plot name and the clicked coordinates to \code{.clickCoord}.

\code{.clickCoord} is intended for internal use and is called by other functions here.
}

