% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{class}
\name{spadesPlot-class}
\alias{spadesPlot}
\alias{spadesPlot-class}
\title{The \code{spadesPlot} class}
\description{
This class contains the plotting spadesPlot information.
}
\details{
These \code{gp*} parameters will specify plot parameters that are
available with \code{gpar()}. \code{gp} will adjust plot parameters,
\code{gpText} will adjust title and legend text, \code{gpAxis} will
adjust the axes. \code{size} adjusts point size in a
\code{SpatialPoints} object. These will persist with the
original \code{Plot} call for each individual object. Multiple
entries can be used, but they must be named list elements
and they must match the \code{...} items to plot. This is true
for a RasterStack also, i.e., the list of named elements
must be the same length as the number of layers being
plotted. The naming convention used is

\code{RasterStackName.layerName}, i.e,   \code{landscape.DEM}.
}
\section{Slots}{

\describe{
\item{\code{arrangement}}{arrangement. An arrangement object}

\item{\code{spadesGrobList}}{list. A list of lists of spadesGrob objects}
}}
\author{
Eliot McIntire
}

