% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-classes.R
\docType{class}
\name{.spadesPlotObjects-class}
\alias{.spadesPlotObjects-class}
\alias{.spadesPlotObjects}
\title{The \code{.spadesPlotObjects} class}
\description{
This class contains the union of spatialObjects and several other plot-type objects.
These are the object classes that the \code{\link{Plot}} function can handle.
}
\section{Slots}{

\describe{
\item{\code{members}}{SpatialPoints*, SpatialPolygons*, SpatialLines*, RasterLayer, RasterStack}
}}

\seealso{
\code{\link{spadesClasses}}
}
\author{
Eliot McIntire
}
