% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{.spadesPlottables}
\alias{.spadesPlottables}
\title{The \code{.spadesPlottables} class}
\description{
This class is the union of all .spadesPlotObjects (e.g., RasterLayer*,
SpatialPoints*, SpatialPolygons*, ggplot, hist etc.) and \code{\link{.spadesPlot}}
class objects.  This allows replotting of a \code{\link{.spadesPlot}} object
}
\section{Slots}{

\describe{
\item{\code{members}}{\code{\link{.spadesPlotObjects}} and \code{\link{.spadesPlot}}}
}}
\author{
Eliot McIntire
}
\seealso{
\code{\link{spadesClasses}}
}

