% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{.spadesPlottables-class}
\alias{.spadesPlottables-class}
\title{The \code{.spadesPlottables} class}
\description{
This class is the union of all .spadesPlotObjects (e.g., RasterLayer*,
SpatialPoints*, SpatialPolygons*, ggplot, hist etc.) and \code{\link{.spadesPlot}}
class objects.
}
\section{Slots}{

\describe{
\item{\code{members}}{\code{\link{.spadesPlotObjects}} and \code{\link{.spadesPlot}}}
}}
\section{All non-exported object classes in SpaDES}{


\tabular{ll}{
  \code{\link{.spadesPlot-class}} \tab Main class for \code{\link{Plot}} - contains .spadesGrob and .arrangement objects\cr
  \code{\link{.spadesGrob-class}} \tab GRaphical OBject used by SpaDES - smallest unit\cr
  \code{\link{.arrangement-class}} \tab The layout or "arrangement" of plot objects\cr
  Unions of classes\tab \cr
  \code{\link{.spadesPlottables-class}} \tab The union of all object classes Plot can accept\cr
  \code{\link{.spadesPlotObjects-class}} \tab The union of .spatialObjects and several others\cr
  \code{\link{.spatialObjects-class}} \tab The union of several spatial classes\cr
  --------------------------- \tab ------------------------------------------------------------------------------------------ \cr
}
}
\author{
Eliot McIntire
}

