% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{timestepInSeconds}
\alias{timestepInSeconds}
\alias{timestepInSeconds,simList,character-method}
\title{Convert an scheduled event time to seconds}
\usage{
timestepInSeconds(sim, moduleName)

\S4method{timestepInSeconds}{simList,character}(sim, moduleName)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}

\item{moduleName}{Character string specifying the module from which to call the event.}
}
\value{
Returns the eventTime in seconds, based on the default \code{timeunit} of
the \code{moduleName}.
}
\description{
Ensure all events from modules are working in the same time units.
}
\author{
Eliot McIntire
}

