% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{dyears}
\alias{dmonths}
\alias{dmonths,numeric-method}
\alias{dweek}
\alias{dweeks}
\alias{dweeks,numeric-method}
\alias{dyears}
\alias{dyears,numeric-method}
\title{New time units}
\usage{
dyears(x)

\S4method{dyears}{numeric}(x)

dmonths(x)

\S4method{dmonths}{numeric}(x)

dweeks(x)

\S4method{dweeks}{numeric}(x)
}
\arguments{
\item{x}{numeric. Number of the desired units}
}
\value{
Number of seconds within each unit
}
\description{
SpaDES commonly needs generic durations, like "year" which will round neatly over
a century (i.e., leap years are integrated as 1/4 day within each year), months are
defined as year/12, weeks as year/52
}
\author{
Eliot McIntire
}

