% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-classes.R
\docType{class}
\name{spatialObjects-class}
\alias{spatialObjects-class}
\alias{spatialObjects}
\title{The \code{spatialObjects} class}
\description{
This class is the union of several spatial objects from raster and sp packages.
Notably missing is \code{RasterBrick}, for now.
}
\section{Slots}{

\describe{
\item{\code{members}}{SpatialPoints*, SpatialPolygons*, SpatialLines*,
RasterLayer, RasterStack}
}}

\seealso{
\code{\link{spadesClasses}}
}
\author{
Eliot McIntire
}
