% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{SpatialPointsDataFrameNamed}
\alias{SpatialPointsDataFrameNamed}
\alias{SpatialPointsDataFrameNamed,character-method}
\alias{name,RasterStackNamed-method}
\alias{name,SpatialPointsDataFrameNamed-method}
\alias{name,SpatialPointsNamed-method}
\alias{name<-}
\title{Create a new \code{SpatialPointsDataFrameNamed} object.}
\usage{
SpatialPointsDataFrameNamed(..., name)

\S4method{SpatialPointsDataFrameNamed}{character}(..., name)

\S4method{name}{SpatialPointsNamed}(object)

\S4method{name}{SpatialPointsDataFrameNamed}(object)

\S4method{name}{RasterStackNamed}(object)

name(object) <- value
}
\arguments{
\item{...}{Additional arguments to \code{SpatialPointsDataFrame}}

\item{name}{The name of the object.}
}
\value{
Returns an object of class \code{SpatialPointsDataFrameNamed}.
}
\description{
Create a new \code{SpatialPointsDataFrameNamed} object.
}
\author{
Eliot McIntire

Alex Chubaty
}
\seealso{
\code{\link{SpatialPointsDataFrame}}
}

