% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{SpatialPointsNamed}
\alias{SpatialPointsNamed}
\alias{SpatialPointsNamed,character-method}
\title{Create a new \code{SpatialPointsNamed} object.}
\usage{
SpatialPointsNamed(..., name)

\S4method{SpatialPointsNamed}{character}(..., name)
}
\arguments{
\item{...}{Additional arguments to \code{SpatialPoints}}

\item{name}{The name of the object.}
}
\value{
Returns an object of class \code{SpatialPointsNamed}.
}
\description{
Create a new \code{SpatialPointsNamed} object.
}
\author{
Eliot McIntire

Alex Chubaty
}
\seealso{
\code{\link{SpatialPoints}}
}

