% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{SpatialPolygonsDataFrameNamed}
\alias{SpatialPolygonsDataFrameNamed}
\alias{SpatialPolygonsDataFrameNamed,character-method}
\title{Create a new \code{SpatialPolygonsDataFrameNamed} object.}
\usage{
SpatialPolygonsDataFrameNamed(..., name)

\S4method{SpatialPolygonsDataFrameNamed}{character}(..., name)
}
\arguments{
\item{...}{Additional arguments to \code{SpatialPolygonsDataFrame}}

\item{name}{The name of the object.}
}
\value{
Returns an object of class \code{SpatialPolygonsDataFrameNamed}.
}
\description{
Create a new \code{SpatialPolygonsDataFrameNamed} object.
}
\author{
Eliot McIntire

Alex Chubaty
}
\seealso{
\code{\link{SpatialPolygonsDataFrame}}
}

