% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SpatialPolygonsNamed-class}
\alias{SpatialPolygonsNamed-class}
\title{SpatialPolygonsNamed class}
\description{
This new class extends \code{SpatialPolygons} by adding a slot
called \code{name}.
}
\details{
Using the simulation visualization features in \code{Plot} requires
objects to be plotted to have a name set.
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the object.}
}}
\seealso{
\code{\link{SpatialPolygons}}
}

