% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{SpatialPolygonsNamed}
\alias{SpatialPolygonsNamed}
\alias{SpatialPolygonsNamed,character-method}
\title{Create a new \code{SpatialPolygonsNamed} object.}
\usage{
SpatialPolygonsNamed(..., name)

\S4method{SpatialPolygonsNamed}{character}(..., name)
}
\arguments{
\item{...}{Additional arguments to \code{SpatialPolygons}}

\item{name}{The name of the object.}
}
\value{
Returns an object of class \code{SpatialPolygonsNamed}.
}
\description{
Create a new \code{SpatialPolygonsNamed} object.
}
\author{
Eliot McIntire

Alex Chubaty
}
\seealso{
\code{\link{SpatialPolygons}}
}

