% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splitRaster.R
\docType{methods}
\name{splitRaster}
\alias{splitRaster}
\title{Split a raster into multiple tiles}
\usage{
splitRaster(x, nx, ny, crop = TRUE, ...)
}
\arguments{
\item{x}{The raster to be split.}

\item{nx}{The number of tiles to make along the x-axis.}

\item{ny}{The number of tiles to make along the y-axis.}

\item{crop}{Logical. Should the raster be cropped into tiles,
             or simply masked? Default is \code{TRUE}.}

\item{...}{Additional arguments.}
}
\value{
Opens a new plot device on the screen.
}
\description{
Divides up a raster into an arbitrary number of pieces.
}
\examples{
# an example with dimensions:
# nrow = 77
# ncol = 101
# nlayers = 3
x <- brick(system.file("external/rlogo.grd", package="raster"))
x1 <- x[[1]] # use first layer only
nx=3; ny=4
y = splitRaster(x1, nx, ny)

layout(mat=matrix(seq_len(nx*ny), ncol=nx, nrow=ny))
lapply(y, plot)
}

