% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-methods.R
\name{rbindlistDtDtpot}
\alias{rbindlistDtDtpot}
\alias{reorderColsWDistance}
\alias{angleQuality}
\alias{asymmetryAdjust}
\title{Internal helper}
\usage{
rbindlistDtDtpot(dt, dtPotential, returnFrom, needDistance, dtPotentialColNames)

reorderColsWDistance(needDistance, dtPotential, dtPotentialColNames)

angleQuality(dtPotential, landscape, actualAsymmetryAngle)

asymmetryAdjust(angleQualities, quantity, actualAsymmetry)
}
\arguments{
\item{dt}{Data.table}

\item{dtPotential}{Data.table}

\item{returnFrom}{Logical}

\item{needDistance}{Logical}

\item{dtPotentialColNames}{Character Vector.}

\item{landscape}{RasterLayer passed from \code{spread2}}

\item{actualAsymmetryAngle}{Angle in degrees, either a vector length 1 or vector NROW(dtPotential)}

\item{angleQualities}{Matrix. The output from \code{angleQuality}}

\item{quantity}{Variable of interest to adjust, e.g., \code{spreadProb}}

\item{actualAsymmetry}{Asymmetry intensity. Derived from \code{asymmetry} arg in \code{spread2}}

\item{dtPotential}{Data.table of potential spread locations.}
}
\description{
Not for users. A function to setnames and rbindlist that is used 3 places in spread2.

Internal helper

Not for users. A function used in spread2.

Not for users. A function used in spread2.
}
\keyword{internal}
