\docType{class}
\name{spreadAgent-class}
\alias{spreadAgent-class}
\title{The \code{spreadAgent} class.}
\description{
Extends the \code{rasterAgent} class. Describes spatial
agents whose spatial component ("area") is characterized by
a raster and are capable of spreading (growing/shrinking).
}
\details{
\describe{ \item{ID}{A character}

\item{NumPix}{A numeric object describing the number of
raster pixels occupied.}

\item{other}{A list for storing other (user-specific) agent
attributes.} }
}

