% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{methods}
\name{.tagsByClass,simList-method}
\alias{.tagsByClass,simList-method}
\title{tagsByClass for simList class objects}
\usage{
\S4method{.tagsByClass}{simList}(object)
}
\arguments{
\item{object}{Any R object.}
}
\description{
See \code{\link[reproducible]{.tagsByClass}}. Adds current \code{moduleName},
\code{eventType}, \code{eventTime}, and \code{function:spades} as userTags
}
\seealso{
\code{\link[reproducible]{.tagsByClass}}
}
\author{
Eliot McIntire
}
