% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-other.R
\docType{methods}
\name{.unittrim}
\alias{.unittrim}
\title{Convert \code{grid.locator} units}
\usage{
.unittrim(grid.locator)
}
\arguments{
\item{grid.locator}{an object that was output by a call to \code{grid.locator}
and mouse click(s).}
}
\description{
Internal function from example in \code{?grid.locator}.
Converts \code{grid.locator} units to meaningful units.
Used within \code{.clickCoord}
}
\author{
Paul Murrell
}
\keyword{internal}

