% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-parseModule.R
\docType{methods}
\name{.unparsed}
\alias{.unparsed}
\alias{.unparsed,list-method}
\title{Determine which modules in a list are unparsed}
\usage{
.unparsed(modules)

\S4method{.unparsed}{list}(modules)
}
\arguments{
\item{modules}{A character vector specifying the modules to parse.}
}
\value{
The ids of the unparsed list elements.
}
\description{
Internal function, used during \code{\link{simInit}}.
}
\author{
Alex Chubaty
}
\keyword{internal}
