% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{.updateSpadesPlot}
\alias{.updateSpadesPlot}
\alias{.updateSpadesPlot,.spadesPlot,ANY-method}
\alias{.updateSpadesPlot,.spadesPlot,missing-method}
\title{Merge two SpaDES Plot objects}
\usage{
.updateSpadesPlot(newSP, curr, ...)

\S4method{.updateSpadesPlot}{.spadesPlot,ANY}(newSP, curr, ...)

\S4method{.updateSpadesPlot}{.spadesPlot,missing}(newSP, curr, ...)
}
\arguments{
\item{newSP}{The "new" \code{.spadesPlot} object.
I.e., the new merges and overwrites into current.}

\item{curr}{The "current" \code{.spadesPlot} object.
I.e., the one to be merged into.}

\item{...}{Additional arguments. Currently none implemented.}
}
\description{
Merges two \code{.spadesPlot} objects
}
\author{
Eliot McIntire
}

