% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-methods.R
\name{getPaths}
\alias{getPaths}
\alias{setPaths}
\title{Get and set default working directories}
\usage{
getPaths()

setPaths(cachePath = "~/SpaDES/cache", inputPath = "~/SpaDES/inputs",
  modulePath = "~/SpaDES/modules", outputPath = "~/SpaDES/outputs")
}
\arguments{
\item{cachePath}{The default local directory in which to cache simulation outputs.
If not specified, defaults to \code{~/SpaDES/cache}.}

\item{inputPath}{The default local directory in which to look for simulation inputs
If not specified, defaults to \code{~/SpaDES/inputs}.}

\item{modulePath}{The default local directory where modules and data will be downloaded and stored.
If not specified, defaults to \code{~/SpaDES/modules}.}

\item{outputPath}{The default local directory in which to save simulation outputs.
If not specified, defaults to \code{~/SpaDES/outputs}.}
}
\value{
Returns a named list of the user's default working directories.
\code{setPaths} is invoked for the side effect of setting these directories.
}
\description{
Wrapper functions to access the packages options for default working directories.
}
\author{
Alex Chubaty
}

