% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{valueAtClicks}
\alias{valueAtClicks}
\title{Retrieve values on a Plot at clicked points}
\usage{
valueAtClicks(X, n = 1, ...)
}
\arguments{
\item{X}{The plot name to get values on. This must be the correct name for the plot
on which the mouse clicks occur.}
}
\value{
The values on the X at the clicked points
}
\description{
Equivalent to running \code{X[SpatialPoints(locator(n))]} in base graphics. This
silently calls a \code{seekViewport(layerName(X))} before calling \code{grid.locator()}.
}

