% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{valuesAtClicks}
\alias{valuesAtClicks}
\title{Retrieve values on a Plot at clicked points}
\usage{
valuesAtClicks(n = 1, ...)
}
\arguments{
\item{n}{The number of mouse clicks to do}
}
\value{
The layer names and values at the clicked points
}
\description{
Equivalent to running \code{X[SpatialPoints(locator(n))]} in base graphics, but it
determines which grid location the click occurs in. This determines which place in the
grid.layout was clicked and makes all appropriate calculations to determine the value
on the raster(s) at that or those location(s)
}

