\docType{class}
\name{vectorAgent-class}
\alias{vectorAgent-class}
\title{The \code{vectorAgent} class.}
\description{
Extends the \code{agent} class by making it spatial.
Describes spatial agents whose spatial components are
characterized by points or a polygon (i.e., in contrast to
a raster). This is a virtual class and not intended to be
used directly.
}
\details{
\describe{ \item{ID}{A character}

\item{other}{A list for storing other (user-specific) agent
attributes.} }
}

