% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spread-process.R
\docType{methods}
\name{Ward}
\alias{Ward}
\title{Ward Seed Dispersal kernel}
\format{\preformatted{  expression(if (cellSize <= effDist) \{     ifelse(dis <= effDist, exp((dis - cellSize) * log(1 - k)/effDist) -          exp(dis * log(1 - k)/effDist), (1 - k) * exp((dis - cellSize -          effDist) * log(b)/maxDist) - (1 - k) * exp((dis - effDist) * log(b)/maxDist)) \} else \{     ifelse(dis == cellSize, exp((dis - cellSize) * log(1 - k)/effDist) -          (1 - k) * exp((dis - effDist) * log(b)/maxDist), (1 - k) * exp((dis -          cellSize - effDist) * log(b)/maxDist) - (1 - k) * exp((dis - effDist) *          log(b)/maxDist)) \})
}}
\usage{
Ward
}
\description{
A probability distribution used in Landis.
}
\author{
Eliot McIntire
}
\keyword{datasets}

