% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listWebData.R
\docType{methods}
\name{listWebData}
\alias{listWebData}
\alias{urls}
\title{List available data from a dataset  store in a http or ftp site.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 4 columns.}
\usage{
listWebData(urltble, datasetName)

urls
}
\arguments{
\item{urltble}{A data.table that stores available dataset name, url, password and filenames found within each dataset.
urltble is provided within the package as urls object.}

\item{datasetName}{A character string. Represent the dataset of interest for download. datasetName allow to derived
url and password from the urltble.}
}
\value{
List available data from a specific dataset

Object containing dataset available for download.
}
\description{
This function browse http and ftp site of interest using datasetName and store in a list available data.

R object that stores dataset with their respective url and username/password. This table makes the link between
dataset name, url and username/password. By only poviding the dataset name, listWebdata and dwdData functions extract url
username/password to access, extract or download data.
}
\examples{

dt<- data.table(dataset = c("NFDB"),
               url = c("http://cwfis.cfs.nrcan.gc.ca/downloads/nfdb/fire_poly/current_version/"),
               password= c(NA))

setwd(tempdir())
listWebData(dt, datasetName = "NFDB")

dt<- data.table(dataset = c("NFDB"),
               url = "http://cwfis.cfs.nrcan.gc.ca/downloads/nfdb/fire_poly/current_version/",
               password= NA,
               files = "NFDB_poly.zip")
}
\author{
Melina Houle

Melina Houle
}
\keyword{datasets}
