% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\docType{methods}
\name{.xyAxes}
\alias{.xyAxes}
\alias{.xyAxes,.spadesGrob,.arrangement-method}
\title{Internal function to determine which x and y axes to add to plots}
\usage{
.xyAxes(sGrob, arr, whPlotFrame)

\S4method{.xyAxes}{.spadesGrob,.arrangement}(sGrob, arr, whPlotFrame)
}
\arguments{
\item{sGrob}{spadesGrob object}

\item{arr}{An arrangement object}

\item{whPlotFrame}{Numeric. Which plot within the spadesGrobPlots object.}
}
\description{
Not intended for normal user.
}
\author{
Eliot McIntire
}

