% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module-dependencies.R
\docType{methods}
\name{addSimDep}
\alias{addSimDep}
\alias{addSimDep,moduleDeps-method}
\title{Add simulation dependencies}
\usage{
addSimDep(x)

\S4method{addSimDep}{moduleDeps}(x)
}
\arguments{
\item{x}{A \code{moduleDeps} object.}
}
\description{
Adds a \code{moduleDeps} object to the simulation dependency list stored in \code{.spadesEnv}.
}
\examples{
\dontrun{
  addSimDep(...)
}
}
\author{
Alex Chubaty
}

