\docType{class}
\name{agent-class}
\alias{agent-class}
\title{The \code{agent} class.}
\description{
Some details about this class and my plans for it in the
body.
}
\section{Slots}{

\describe{
\item{\code{ID}}{A character string that identifies the individual agent.
             By default, the numbers 1...n are used.}

\item{\code{other}}{A list for storing other (user-specific) agent attributes.}
}}
\note{
You can still add notes
}

