% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{placeOnRaster,SpatialPoints-method}
\alias{arrangePlots}
\alias{arrangePlots,Raster-method}
\alias{arrangePlots,SpatialPoints-method}
\alias{arrangePlots,list-method}
\alias{placeOnRaster,SpatialPoints-method}
\title{Determine optimal plotting arrangement of RasterStack}
\usage{
\S4method{placeOnRaster}{SpatialPoints}(obj, raster = NULL)

arrangePlots(toPlot, nam = NULL, which.to.plot = "all", axes = "L")

\S4method{arrangePlots}{Raster}(toPlot, nam = NULL, which.to.plot = "all",
  axes = "L")

\S4method{arrangePlots}{SpatialPoints}(toPlot,
  nam = deparse(substitute(toPlot)), axes = "L")

\S4method{arrangePlots}{list}(toPlot, nam, axes = "L")
}
\arguments{
\item{toPlot}{Raster* object}

\item{axes}{passed from Plot}
}
\description{
Hidden function.
}
\details{
This assesses the device geometry, the map geometry, and the number of rasters
to plot and builds an object that will be used by the Plot functions to plot
them efficiently
}

