% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{arrangeSimPlots}
\alias{arrangeSimPlots}
\title{Determine optimal plotting arrangement of RasterStack}
\usage{
arrangeSimPlots(rast, nam = NULL, which.to.plot = "all", axes = "L")
}
\arguments{
\item{rast}{Raster* object}

\item{axes}{passed from simPlot}
}
\description{
Hidden function.
}
\details{
This assesses the device geometry, the map geometry, and the number of rasters
to plot and builds an object that will be used by the simPlot functions to plot
them efficiently
}

