% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{arrangeSimPlots}
\alias{arrangeSimPlots}
\title{Determine optimal plotting arrangement of RasterStack}
\usage{
arrangeSimPlots(ext, dimx, nam, which.to.plot, axes = "L")
}
\arguments{
\item{ext}{extent object}

\item{dimx}{dimension of rasterStack}

\item{nam}{names in rasterStack}

\item{which.to.plot}{vector of numbers or names in rasterStack to plot}

\item{axes}{passed from simPlot}
}
\description{
Hidden function.
}
\details{
This assesses the device geometry, the map geometry, and the number of rasters
to plot and builds an object that will be used by the simPlot functions to plot
them efficiently
}

