% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\docType{methods}
\name{.arrangeViewports}
\alias{.arrangeViewports}
\alias{.arrangeViewports,.spadesPlot-method}
\title{Determine optimal plotting arrangement of plot objects}
\usage{
.arrangeViewports(sPlot, arr = NULL)

\S4method{.arrangeViewports}{.spadesPlot}(sPlot, arr = NULL)
}
\arguments{
\item{sPlot}{A \code{.spadesPlot} object.}

\item{arr}{A vector of length 2 indicating a desired arrangement of plot
areas indicating number of rows, number of columns.
Default NULL, meaning
let Plot function do it automatically.}
}
\description{
Internal function. Assesses the device geometry, the map geometry, and the
number of spatial objects to plot and builds an object that will be used by
the Plot functions to plot them efficiently.
}
\author{
Eliot McIntire
}

