% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{.arrangeViewports}
\alias{.arrangeViewports}
\alias{.arrangeViewports,.spadesPlot-method}
\title{Determine optimal plotting arrangement of plot objects}
\usage{
.arrangeViewports(sPlot)

\S4method{.arrangeViewports}{.spadesPlot}(sPlot)
}
\arguments{
\item{sPlot}{A \code{.spadesPlot} object.}
}
\description{
This assesses the device geometry, the map geometry, and the number of spatial
objects to plot and builds an object that will be used by the Plot functions to plot
them efficiently. This is meant to be used internally.
}
\author{
Eliot McIntire
}

