% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkpoint.R
\docType{methods}
\name{cache}
\alias{cache}
\title{Cache method for simList class objects}
\usage{
cache(cacheRepo = NULL, FUN, ..., notOlderThan = NULL)
}
\arguments{
\item{cacheRepo}{An object repository used for storing cached objects.}

\item{FUN}{A function to be called.}

\item{...}{Arguments for function \code{FUN}.}

\item{notOlderThan}{Restore an artifact from database only if it was created after notOlderThan.}
}
\value{
Identical to \code{\link[archivist]{cache}}
}
\description{
Because the \code{simList} has an environment as one of its slots, the caching mechanism
of the archivist package does not work. Here, we make a slight tweak to the
\code{cache} function to convert the \code{simList} to the \code{list} equivalent before
calling the \code{\link[digest]{digest}} function internally. Otherwise, this is exactly the
same as \code{\link{[archivist]{cache}}}.
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{[archivist]{cache}}}.
}

