% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{checkObject}
\alias{checkObject}
\alias{checkObject,ANY-method}
\alias{checkObject,Raster-method}
\title{Check for existence of a global object}
\usage{
checkObject(name, ...)

\S4method{checkObject}{Raster}(name, layer, ...)

\S4method{checkObject}{ANY}(name, ...)
}
\arguments{
\item{name}{A character string specifying
the name of an object to be checked}

\item{...}{Additional arguments. Not implemented.}

\item{layer}{Character string, specifying a layer name in a Raster}
}
\value{
Nothing is returned. Specified packages are loaded and attached using \code{library()}.
}
\description{
Check that a named object exists in the global environment, and optionally has
desired attributes.
}
\examples{
\dontrun{}
\dontrun{}
\dontrun{}
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{library}}.
}

