% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{checkObject}
\alias{checkObject}
\alias{checkObject,character,missing,character-method}
\alias{checkObject,character,missing,missing-method}
\alias{checkObject,missing,ANY,missing-method}
\alias{checkObject,missing,Raster,character-method}
\title{Check for existence of a global object}
\usage{
checkObject(name, object, layer, ...)

\S4method{checkObject}{missing,Raster,character}(name, object, layer, ...)

\S4method{checkObject}{missing,ANY,missing}(name, object, layer, ...)

\S4method{checkObject}{character,missing,missing}(name, object, layer, ...)

\S4method{checkObject}{character,missing,character}(name, object, layer, ...)
}
\arguments{
\item{name}{A character string specifying the name of an object to be checked.}

\item{object}{An object. This is mostly used internally, or with layer,
                 because it will fail if the object does not exist.}

\item{layer}{Character string, specifying a layer name in a Raster, if the
               \code{name} is a Raster* object.}

\item{...}{Additional arguments. Not implemented.}
}
\value{
Invisibly return \code{TRUE} indicating object exists; code{FALSE} if not.
}
\description{
Check that a named object exists in the global environment, and optionally has
desired attributes.
}
\author{
Alex Chubaty

Eliot McIntire
}
\seealso{
\code{\link{library}}.
}

