\docType{methods}
\name{cir}
\alias{cir}
\title{Circle around an agent.}
\usage{
cir(agent, radiuses, raster_world, scale_raster)
}
\arguments{
  \item{agent}{Description of this.}

  \item{radiuses}{Description of this, including why it
  isn't called radii ;p}

  \item{raster_world}{Description of this.}

  \item{scale_raster}{Description of this.}
}
\value{
A list of data.frames with x and y coordinates of each
unique pixel of the circle around each individual.

#
}
\description{
Identifies the xy coordinates of a circle around all live
agents.
}
\seealso{
\code{\link{print}} and \code{\link{cat}}
}

