% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{.cols}
\alias{.cols}
\title{A short selection of colour palettes that can be use}
\format{\preformatted{List of 6
 $ transparent.red: chr [1:9] "#00000000" "#25252566" "#52525266" "#73737366" ...
 $ grey           : chr [1:9] "#FFFFFF" "#F0F0F0" "#D9D9D9" "#BDBDBD" ...
 $ spectral       : chr [1:8] "#D53E4F" "#F46D43" "#FDAE61" "#FEE08B" ...
 $ terrain        : chr [1:100] "#F2F2F2FF" "#F2EDEDFF" "#F2E8E8FF" "#F1E4E3FF" ...
 $ heat           : chr [1:10] "#FF0000FF" "#FF2400FF" "#FF4900FF" "#FF6D00FF" ...
 $ topo           : chr [1:10] "#4C00FFFF" "#0019FFFF" "#0080FFFF" "#00E5FFFF" ...
}}
\usage{
.cols
}
\description{
Colour number 1 shows use of transparency
}
\keyword{datasets}

