\docType{methods}
\name{coordinates,pointAgent-method}
\alias{coordinates,pointAgent-method}
\title{get coordinates of a pointAgent}
\usage{
\S4method{coordinates}{pointAgent}(obj, ...)
}
\arguments{
  \item{obj}{ object deriving from class "Spatial" }

  \item{...}{additional arguments that may be used by
  particular instances}
}
\description{
get coordinates of a pointAgent
}

