% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\docType{methods}
\name{copy}
\alias{copy}
\alias{copy,simList,logical-method}
\alias{copy,simList,missing-method}
\title{Copy a simList object}
\usage{
copy(sim, objects = TRUE)

\S4method{copy}{simList,logical}(sim, objects = TRUE)

\S4method{copy}{simList,missing}(sim)
}
\arguments{
\item{sim}{A \code{simList} object.}

\item{objects}{Whether the objects contained within the simList environment
should be copied. Default = TRUE, which may be slow.}
}
\description{
Because a simList works with an environment to hold all objects,
all objects within that slot are pass-by-reference. That means
it is not possible to simply copy an object with an assignment operator:
the two objects will share the same objects. As one simList object changes
so will the other. when this is not the desired behaviour, use this function.
}

