% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{methods}
\name{.coreModules}
\alias{.coreModules}
\title{Blank (template) event list}
\usage{
.coreModules()
}
\value{
Returns a named list of the core modules.
}
\description{
Internal function called from \code{spades}, returning an empty event list.
}
\details{
Event lists are sorted (keyed) first by time, second by priority.
Each event is represented by a \code{\link{data.table}} row consisting of:
\tabular{ll}{
  \code{eventTime} \tab The time the event is to occur.\cr
  \code{moduleName} \tab The module from which the event is taken.\cr
  \code{eventType} \tab A character string for the programmer-defined event type.\cr
  \code{eventPriority} \tab The priority given to the event. \cr
}
}
\author{
Alex Chubaty
}
\keyword{internal}

