% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{createsOutput}
\alias{createsOutput}
\alias{createsOutput,character,character,character-method}
\title{Define an output object of a module}
\usage{
createsOutput(objectName, objectClass, desc, ...)

\S4method{createsOutput}{character,character,character}(objectName, objectClass,
  desc, ...)
}
\arguments{
\item{objectName}{Character string to define the output object's name.}

\item{objectClass}{Character string to specify the output object's class.}

\item{desc}{Text string providing a brief description of the output object.}

\item{...}{Other specifications of the output object.}
}
\value{
a data frame
}
\description{
Used to specify an output object's name, class, description and other specifications.
}
\examples{
library(dplyr)
outputObjets = bind_rows(
  createsOutput(objectName = "outputObject1", objectClass = "character",
                desc = "this is for example"),
  createsOutput(objectName = "outputObject2", objectClass = "numeric",
                desc = "this is for example",
                otherInformation = "I am the second output object")
)

}
\author{
Yong Luo
}

