% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{listOrEnvDigestRecursive}
\alias{listOrEnvDigestRecursive}
\alias{digestRaster}
\title{Custom tools for digesting objects}
\usage{
listOrEnvDigestRecursive(x)

digestRaster(object, compareRasterFileLength, algo)
}
\arguments{
\item{Any}{object}
}
\description{
For reproducibility, there are many features or attributes of objects that must
be removed e.g., environments have unique labels, rasters have several infrequently
used slots and elements that are not perfectly maintained with manipulation.
These customDigest functions attempt to deal with some of the types of problems.
In conjunction with \code{\link{makeDigestible}}, these are helpers to create
consisten cache results.
}
\seealso{
\code{\link{makeDigestible}}
}
\author{
Eliot McIntire
}
