% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module-dependencies.R
\docType{methods}
\name{defineModule}
\alias{defineModule}
\alias{defineModule,list-method}
\title{Define a new module}
\usage{
defineModule(x)

\S4method{defineModule}{list}(x)
}
\arguments{
\item{x}{A named list containing the parameters used to construct a new
           \code{moduleDeps} object.}
}
\description{
Specify a new module's metadata as well as object and package dependecies.
This is a constructor method for the \code{\link{moduleDeps}} class.
}
\examples{
\dontrun{
  moduleInfo <- list(...)
  defineModule(moduleInfo)
}
}
\author{
Alex Chubaty
}

