% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module-dependencies.R
\docType{methods}
\name{defineModule}
\alias{defineModule}
\alias{defineModule,character,character,character,person,Extent,POSIXt,character,list,list,list,data.frame,data.frame-method}
\title{Define a new module}
\usage{
defineModule(name, description, keywords, authors, spatialExtent, timeframe,
  timestep, translators, citation, reqdPkgs, inputObjects, outputObjects)


  \S4method{defineModule}{character,character,character,person,Extent,POSIXt,character,list,list,list,data.frame,data.frame}(name,
  description, keywords, authors, spatialExtent, timeframe, timestep,
  translators, citation, reqdPkgs, inputObjects, outputObjects)
}
\description{
Specify a new module's metadata as well as object and package dependecies.
This is a constructor method for the \code{\link{moduleDepends}} class.
}
\examples{
\dontrun{
  defineModule(...)
}
}
\author{
Alex Chubaty
}

