% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{defineModule}
\alias{defineModule}
\alias{defineModule,simList,list-method}
\title{Define a new module.}
\usage{
defineModule(sim, x)

\S4method{defineModule}{simList,list}(sim, x)
}
\arguments{
\item{sim}{A \code{simList} object.}

\item{x}{A named list containing the parameters used to construct a new
           \code{.moduleDeps} object.}
}
\value{
Updated \code{simList} object.
}
\description{
Specify a new module's metadata as well as object and package dependecies.
Packages are loaded during this call.
}
\examples{
\dontrun{
  moduleInfo <- list(...)
  defineModule(sim, moduleInfo)
}
}
\author{
Alex Chubaty
}

