% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{defineParameter}
\alias{defineParameter}
\alias{defineParameter,character,character,ANY,ANY,ANY-method}
\alias{defineParameter,character,character,ANY,missing,missing-method}
\title{Define a parameter used in a module}
\usage{
defineParameter(name, class, default, min, max)

\S4method{defineParameter}{character,character,ANY,ANY,ANY}(name, class,
  default, min, max)

\S4method{defineParameter}{character,character,ANY,missing,missing}(name, class,
  default)
}
\arguments{
\item{name}{Character string giving the parameter name.}

\item{class}{Character string giving the parameter class.}

\item{default}{The default value to use when none is specified by the user.
Non-standard evaluation is used for the expression.}

\item{min}{With \code{max}, used to define a suitable range of values.
Non-standard evaluation is used for the expression.}

\item{max}{With \code{min}, used to define a suitable range of values.
                 Non-standard evaluation is used for the expression.}
}
\value{
data.frame
}
\description{
Used to specify a parameter's name, value, and set a default.
}
\examples{
parameters = rbind(
  defineParameter("lambda", "numeric", 1.23),
  defineParameter("p", "numeric", 0.2, 0, 1)
)
}
\author{
Alex Chubaty
}

