% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POM.R
\docType{methods}
\name{DEoptim2}
\alias{DEoptim2}
\title{Replacement of DEoptim function, allowing custom parallel cluster}
\usage{
DEoptim2(fn, lower, upper, control = DEoptim.control(), ..., cl,
  fnMap = NULL)
}
\arguments{
\item{cl}{A cluster object. Optional. This would generally be created using
parallel::makeCluster or equivalent. This is an alternative way, instead
of beginCluster(), to use parallelism for this function, allowing for
more control over cluster use.}
}
\value{
See \code{\link[DEoptim]{DEoptim}}.
}
\description{
This is a verbatim copy of \code{\link[DEoptim]{DEoptim}}, but with
\code{parallelType == 3}, and a new argument, \code{cl}, allowing a custom
cluster object to be provided.
}
\author{
DEoptim authors with Eliot McIntire modification
}
\seealso{
\code{\link[parallel]{makeCluster}}
}

