% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modules-dependecies.R
\docType{class}
\name{dependsList-class}
\alias{dependsList-class}
\title{The \code{dependsList} class}
\description{
List of module object dependecies within a SpaDES simulation.
}
\section{Slots}{

\describe{
\item{\code{dependencies}}{List of \code{\link{moduleDeps}} dependency objects.}
}}
\author{
Alex Chubaty
}

