% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dev}
\alias{dev}
\title{Specify where to plot}
\usage{
dev(x, ...)
}
\arguments{
\item{x}{The number of a plot device.}

\item{...}{Additional arguments passed to \code{\link{quartz}},
           \code{\link{windows}}, or \code{\link{x11}}.}
}
\value{
Opens a new plot device on the screen.
}
\description{
Switch to an existing plot device, or if not already open,
launch a new graphics device based on operating system used.
}
\details{
For example, \code{dev(6)} switches the active plot device to device #6.
If it doesn't exist, it opens it. NOTE: if devices 1-5 don't exist
they will be opened too.
}

