\docType{methods}
\name{doEvent,simList,logical-method}
\alias{doEvent,simList,logical-method}
\alias{doEvent,simList,missing-method}
\title{doEvent}
\usage{
\S4method{doEvent}{simList,logical}(sim, debug)

\S4method{doEvent}{simList,missing}(sim)
}
\arguments{
  \item{sim}{A \code{simList} simulation object.}

  \item{debug}{Optional logical flag determines whether sim
  debug info will be printed (default is
  \code{debug=FALSE}).}
}
\description{
doEvent

doEvent
}

