% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{doSim}
\alias{doSim}
\alias{doSim,simList,logical-method}
\alias{doSim,simList,missing-method}
\title{Process a simulation event}
\usage{
doSim(sim, debug)

\S4method{doSim}{simList,logical}(sim, debug)

\S4method{doSim}{simList,missing}(sim)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}

\item{debug}{Optional logical flag determines whether sim debug info
             will be printed (default is \code{debug=FALSE}).}
}
\value{
Returns the modified \code{simList} object.
}
\description{
Internal function called from \code{doSim}.

doSim

doSim
}
\details{
Calls the module corresponding to the event call, and executes the event.

Based on code from Matloff (2011) ch. 7.8.3 on discrete event simulation:
- implemented using S4 classes and methods;
- uses `data.table` instead of `data.frame` (which is much faster);
- implemented in a more modular fashion so it's easier
  to add submodules to the simulation.
}
\note{
The debug option is primarily intended to facilitate building simulation
models by the user. Setting \code{debug=TRUE} allows the user to toggle debugging
statements in their own modules.
}
\examples{
\dontrun{mySim <- simInit(times=list(start=0.0, stop=10.0), params=list(Ncaribou=100),
modules=list("habitat", "caribou"), path="/path/to/my/modules/)}
\dontrun{doSim{mySim}}
}
\author{
Alex Chubaty
}
\references{
Matloff, N. (2011). The Art of R Programming (373 pp.). San Fransisco, CA: No Starch Press, Inc.. Retrieved from \url{http://www.nostarch.com/artofr.htm}
}
\seealso{
\code{\link{simInit}}.
}

