% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-repository.R
\docType{methods}
\name{downloadData}
\alias{downloadData}
\alias{downloadData,character,character-method}
\title{Download module data}
\usage{
downloadData(module, path)

\S4method{downloadData}{character,character}(module, path)
}
\arguments{
\item{module}{Character string giving the name of the module.}

\item{path}{Character string giving the path to the module directory.}
}
\value{
Invisibly, a character vector containing a list of downloaded files.
}
\description{
Download external data for a module if not already present in the module
directory or if there is a checksum mismatch indicating that the file is not
the correct one.
}
\author{
Alex Chubaty
}

