% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-repository.R
\docType{methods}
\name{downloadData}
\alias{downloadData}
\alias{downloadData,character,character,logical-method}
\alias{downloadData,character,missing,missing-method}
\alias{downloadData,character,missing,logical-method}
\alias{downloadData,character,character,missing-method}
\title{Download module data}
\usage{
downloadData(module, path, quiet)

\S4method{downloadData}{character,character,logical}(module, path, quiet)

\S4method{downloadData}{character,missing,missing}(module)

\S4method{downloadData}{character,missing,logical}(module, quiet)

\S4method{downloadData}{character,character,missing}(module, path)
}
\arguments{
\item{module}{Character string giving the name of the module.}

\item{path}{Character string giving the path to the module directory.}

\item{quiet}{Logical. This is passed to \code{download.file}. Default is FALSE.}
}
\value{
Invisibly, a list of downloaded files.
}
\description{
Download external data for a module if not already present in the module
directory or if there is a checksum mismatch indicating that the file is not
the correct one.
}
\author{
Alex Chubaty
}
