\docType{methods}
\name{drawArrows}
\alias{drawArrows}
\alias{drawArrows,mobileAgent-method}
\title{plot arrows showing direction of mobileAgent movement}
\usage{
drawArrows(agent, ...)

\S4method{drawArrows}{mobileAgent}(agent, ..., length = 0.1)
}
\arguments{
  \item{agent}{A \code{mobileAgent} object.}

  \item{...}{Additional plotting parameters.}

  \item{length}{The length of the arrows to draw (defaults
  to 0.1).}
}
\value{
Returns the modified \code{SimList} object.
}
\description{
Plots arrows showing direction of mobileAgent movement.

plot arrows showing direction of mobileAgent movement
}

