% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{drawArrows}
\alias{drawArrows}
\alias{drawArrows,SpatialPoints,SpatialPoints,character-method}
\alias{drawArrows,SpatialPoints,SpatialPoints,missing-method}
\title{Plots arrows showing direction of agent movement.}
\usage{
drawArrows(from, to, addTo, ...)

\S4method{drawArrows}{SpatialPoints,SpatialPoints,character}(from, to, addTo,
  ..., length = 0.1)

\S4method{drawArrows}{SpatialPoints,SpatialPoints,missing}(from, to, addTo, ...,
  length = 0.1)
}
\arguments{
\item{from}{Starting spatial coordinates (\code{SpatialPointsDataFrame}).}

\item{to}{Ending spatial coordinates (\code{SpatialPointsDataFrame})..}

\item{addTo}{Optional character string. The name of a map layer on which to draw the arrows.}

\item{...}{Additional plotting parameters passed to grid.polyline. Currently
does not appear to pass anything.}

\item{length}{The length of the arrows to draw (defaults to 0.1).}
}
\value{
Plots the vectors representing agent movement on the specified map.
}
\description{
Plots arrows showing direction of agent movement.
}
\examples{
# Make 2 objects
caribou1 <- SpatialPointsNamed(cbind(x=runif(10,-50,50),y=runif(10,-50,50)),name="caribou1")
caribou2 <- SpatialPointsNamed(cbind(x=runif(10,-50,50),y=runif(10,-50,50)),name="caribou2")

drawArrows(caribou1, caribou2)
seekViewport("caribou1")
grid.text("caribou1",0.5,1.05)
grid.xaxis()
grid.yaxis()

# or add to a previous Plot
fileList <-
   data.frame(files =
     dir(file.path(
                   find.package("SpaDES",
                                lib.loc=getOption("devtools.path"),
                                quiet=FALSE),
                  "maps"),
        full.names=TRUE, pattern= "tif"),
     functions="rasterToMemory",
     packages="SpaDES",
     stringsAsFactors=FALSE)

# Load files to memory (using rasterToMemory)
loadFiles(fileList=fileList)

Plot(DEM)
drawArrows(caribou1, caribou2, addTo="DEM")
}

