% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{equalExtent}
\alias{equalExtent}
\alias{equalExtent,list-method}
\title{Assess whether a list of extents are all equal}
\usage{
equalExtent(extents)

\S4method{equalExtent}{list}(extents)
}
\description{
Assess whether a list of extents are all equal
}

