% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-define.R
\name{.fillOutputRows}
\alias{.fillOutputRows}
\title{An internal function for coercing a data.frame to outputs()}
\usage{
.fillOutputRows(outputDF, endTime)
}
\arguments{
\item{outputDF}{A data.frame with partial columns to pass to outputs<-}

\item{endTime}{Numeric time. The end(sim).}
}
\description{
An internal function for coercing a data.frame to outputs()
}
\keyword{internal}
