% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{.findFactors}
\alias{.findFactors}
\title{Find factors}
\usage{
.findFactors(x)
}
\arguments{
\item{x}{An integer to factorize}
}
\value{
A vector of integer factors
}
\description{
Internal function (used in \code{link{gaussMap}}).
Finds the integer factors of an integer.
}
\keyword{internal}
