% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-diagrams.R
\docType{methods}
\name{ganttStatus}
\alias{ganttStatus}
\alias{ganttStatus,character-method}
\title{ganttStatus}
\usage{
ganttStatus(eventType)

\S4method{ganttStatus}{character}(eventType)
}
\arguments{
\item{eventType}{Character vector of events.}
}
\value{
A character vector.
}
\description{
Internal function assign the "status" of each event to be passed to
\code{\link[DiagrammeR]{mermaid}} to make a Gantt chart representing the
events in a completed simulation.
'init' events are set as "done"; 'plot' events as "critical"; and all others
as "active".
}
\author{
Alex Chubaty
}
\keyword{internal}
