% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getColors}
\alias{getColors}
\alias{getColors,Raster-method}
\title{Get colours for plotting Raster* objects.}
\usage{
getColors(object)

\S4method{getColors}{Raster}(object)
}
\arguments{
\item{x}{A \code{Raster*} object.}
}
\value{
Returns a named list of colors.
}
\description{
Get colours for plotting Raster* objects.
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{setColors}}, \code{\link{brewer.pal}}
}

