% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colours.R
\docType{methods}
\name{getColors}
\alias{getColors}
\alias{getColors,Raster-method}
\alias{getColours}
\title{Get colours for plotting Raster* objects.}
\usage{
getColors(object)

\S4method{getColors}{Raster}(object)
}
\arguments{
\item{object}{A \code{Raster*} object.}
}
\value{
Returns a named list of colors.
}
\description{
Get colours for plotting Raster* objects.
}
\examples{
map <- raster(system.file("maps/DEM.tif", package = "SpaDES"))
getColors(map) # none set
setColors(map, n=50) <- topo.colors(50)
getColors(map)
Plot(map, new=TRUE)
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{setColors<-}}, \code{\link[RColorBrewer]{brewer.pal}}
}

