% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-colours.R
\docType{methods}
\name{getColors}
\alias{getColors}
\alias{getColors,Raster-method}
\alias{getColours}
\title{Get colours for plotting Raster* objects.}
\usage{
getColors(object)

\S4method{getColors}{Raster}(object)
}
\arguments{
\item{object}{A \code{Raster*} object.}
}
\value{
Returns a named list of colors.
}
\description{
Get colours for plotting Raster* objects.
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{setColors<-}},
         \code{\link[RColorBrewer]{brewer.pal}}
}

