% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\docType{methods}
\name{.getSimEnvNames}
\alias{.getSimEnvNames}
\alias{.getSimEnvNames,environment-method}
\alias{.getSimEnvNames,missing-method}
\title{Get all simEnv names that exist in an environment}
\usage{
.getSimEnvNames(envir)

\S4method{.getSimEnvNames}{environment}(envir)

\S4method{.getSimEnvNames}{missing}()
}
\arguments{
\item{envir}{The environment to search for simEnvs within. Default is .GlobalEnv}
}
\description{
Internal function. Get names of simEnv environments. Primarily used within
Plot function.
}
\author{
Eliot McIntire
}

