% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{globals}
\alias{globals}
\alias{globals,.simList-method}
\alias{globals<-}
\alias{globals<-,.simList-method}
\title{Get and set simulation globals.}
\usage{
globals(object)

\S4method{globals}{.simList}(object)

globals(object) <- value

\S4method{globals}{.simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\description{
\code{globals} accesses or sets the "globals" slot in the \code{simList}.
}
\seealso{
\code{\link{SpaDES}}, specifically the section 1.2.1 on Simulation Parameters.
}

