% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{globals}
\alias{globals}
\alias{globals,.simList-method}
\alias{globals<-}
\alias{globals<-,.simList-method}
\title{Get and set simulation globals.}
\usage{
globals(sim)

\S4method{globals}{.simList}(sim)

globals(sim) <- value

\S4method{globals}{.simList}(sim) <- value
}
\arguments{
\item{sim}{A \code{simList} object from which to extract element(s) or
in which to replace element(s).}

\item{value}{The object to be stored at the slot.}
}
\description{
\code{globals} accesses or sets the "globals" slot in the \code{simList}.
}
\seealso{
\code{\link{SpaDES}}, specifically the section 1.2.1 on Simulation Parameters.

Other functions to access elements of a \code{simList} object: \code{\link{.addDepends}},
  \code{\link{doEvent.checkpoint}}, \code{\link{envir}},
  \code{\link{events}}, \code{\link{inputs}},
  \code{\link{ls.simList}}, \code{\link{ls.str.simList}},
  \code{\link{modules}}, \code{\link{objs}},
  \code{\link{packages}}, \code{\link{params}},
  \code{\link{paths}}, \code{\link{progressInterval}},
  \code{\link{times}}
}

