\docType{methods}
\name{heading}
\alias{heading}
\alias{heading,SpatialPoints,SpatialPoints-method}
\title{Heading between spatial points.}
\usage{
heading(from, to, ...)

\S4method{heading}{SpatialPoints,SpatialPoints}(from, to, ...)
}
\arguments{
  \item{from}{The starting position; an object of class
  SpatiolPoints.}

  \item{to}{The ending position; an object of class
  SpatiolPoints.}

  \item{...}{Additional arguments.}
}
\value{
The heading between the points, in RADIANS or DEGREES?.

#
}
\description{
Determines the heading between spatial points.
}
\examples{
# needs examples
}
\seealso{
\code{\link{print}} and \code{\link{cat}}
}

