% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{.identifyGrobToPlot}
\alias{.identifyGrobToPlot}
\alias{.identifyGrobToPlot,.spadesGrob,list,logical-method}
\alias{.identifyGrobToPlot,.spadesGrob,missing,logical-method}
\title{Identify where to get the grob from}
\usage{
.identifyGrobToPlot(grobNamesi, toPlot, takeFromPlotObj)

\S4method{.identifyGrobToPlot}{.spadesGrob,list,logical}(grobNamesi, toPlot,
  takeFromPlotObj)

\S4method{.identifyGrobToPlot}{.spadesGrob,missing,logical}(grobNamesi, toPlot,
  takeFromPlotObj)
}
\arguments{
\item{grobNamesi}{name of the object to plot}

\item{toPlot}{list containing the objects to plot, made as a call to the
\code{Plot} function}

\item{takeFromPlotObj}{logical. If \code{TRUE}, then take from the call to
\code{Plot}; if \code{FALSE} takes from global envir.}
}
\description{
Internal function.
}
\details{
Because the Plot function can use the global environment as a source of
objects to plot, not just the call itself, this function identifies where
the data for the grob should come from, the current call or the global
environment.
}
\author{
Eliot McIntire
}

