% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{.identifyGrobToPlot}
\alias{.identifyGrobToPlot}
\alias{.identifyGrobToPlot,spadesGrob,missing,logical-method}
\title{Identify where to get the grob from}
\usage{
.identifyGrobToPlot(grobNamesi, toPlot, takeFromPlotObj)

\S4method{.identifyGrobToPlot}{spadesGrob,missing,logical}(grobNamesi, toPlot,
  takeFromPlotObj)
}
\arguments{
\item{grobNamesi}{name of the object to plot}

\item{toPlot}{list containing the objects to plot, made as a call to the Plot function}

\item{takeFromPlotObj}{logical. If TRUE, then take from the call to Plot, FALSE takes
from global environment}
}
\description{
Because the Plot function can use the global environment as a source of
objects to plot, not just the call itself, this function identifies where
the data for the grob should come from, the current call or the global
environment. This should be called internally only.
}
\author{
Eliot McIntire
}

