% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{initialize,mobileAgent-method}
\alias{initialize,mobileAgent-method}
\title{initialize mobileAgent}
\usage{
\S4method{initialize}{mobileAgent}(.Object, ..., agentlocation = NULL,
  numagents = NULL, probinit = NULL)
}
\arguments{
\item{agentlocation}{The initial positions of the agents
                     (currently only \code{RasterLayer} or
                     \code{SpatialPolygonsDataFrame}) accepted.}

\item{numagents}{The number of agents to initialize.}

\item{probinit}{The probability of placing an agent at a given initial position.}

\item{.Object}{ An object:  see the Details section.}

\item{...}{data to include in the new object.  Named arguments
    correspond to slots in the class definition. Unnamed arguments must
    be objects from classes that this class extends.}
}
\description{
initialize mobileAgent
}

