% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SELES.R
\docType{methods}
\name{initiateAgents}
\alias{initiateAgents}
\title{SELES - Initiate agents}
\usage{
initiateAgents(map, numAgents, probInit, asSpatialPoints = TRUE, indices)

\S4method{initiateAgents}{Raster,missing,missing,ANY,missing}(map, numAgents,
  probInit, asSpatialPoints)

\S4method{initiateAgents}{Raster,missing,Raster,ANY,missing}(map, probInit,
  asSpatialPoints)

\S4method{initiateAgents}{Raster,numeric,missing,ANY,missing}(map, numAgents,
  probInit, asSpatialPoints = TRUE, indices)

\S4method{initiateAgents}{Raster,numeric,Raster,ANY,missing}(map, numAgents,
  probInit, asSpatialPoints)

\S4method{initiateAgents}{Raster,missing,missing,ANY,numeric}(map, numAgents,
  probInit, asSpatialPoints = TRUE, indices)
}
\arguments{
\item{map}{RasterLayer with extent and resolution of desired return object}

\item{numAgents}{numeric resulting from a call to \code{\link{numAgents}}}

\item{probInit}{RasterLayer resulting from a call to \code{\link{probInit}}}

\item{indices}{numeric. Indices of where agents should start}

\item{asRaster}{logical. Should returned object be raster or SpatialPointsDataFrame (default)}

\item{probInit}{a Raster resulting from a \code{\link{probInit}} call}
}
\value{
A SpatialPointsDataFrame, with each row representing an individual agent
}
\description{
A SELES-like function to maintain conceptual backwards compatability with that simulation
tool. Sets the the number of agents to initiate. THIS IS NOT FULLY IMPLEMENTED.
This is intended to ease transitinos from
\href{http://www.lfmi.uqam.ca/seles.htm}{SELES}.
You must know how to use SELES for these to be useful
}
\examples{
map <- raster(xmn=0, xmx=10, ymn=0, ymx=10, val=0, res=1)
map <- gaussMap(map, scale=1, var = 4, speedup=1)
pr <- probInit(map, p=(map/maxValue(map))^2)
agents <- initiateAgents(map, 100, pr)
Plot(map, new=T)
Plot(agents, addTo="map")

# If producing a Raster, then the number of points produced can't be more than
# the number of pixels:
agentsRas <- initiateAgents(map, 30, pr, asSpatialPoints=FALSE)
Plot(agentsRas)

# Check that the agents are more often at the higher probability areas based on pr
out <- data.frame(na.omit(crosstab(agentsRas, map)), table(round(map[]))) \%>\%
   mutate(selectionRatio=Freq/Freq.1) \%>\%
   select(-Var1, -Var1.1) \%>\%
   rename(Present=Freq, Avail=Freq.1, Type=Var2)
print(out)
}
\author{
Eliot McIntire
}

