% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{inRange}
\alias{inRange}
\title{Test whether a number lies within range \code{[a,b]}}
\usage{
inRange(x, a = 0, b = 1)
}
\arguments{
\item{x}{values to be tested}

\item{a}{lower bound (default 0)}

\item{b}{upper bound (default 1)}
}
\description{
Default values of \code{a=0; b=1} allow for quick test if
\code{x} is a probability.
}
\author{
Alex Chubaty
}

