% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{layerNames}
\alias{layerNames}
\alias{layerNames,Raster-method}
\alias{layerNames,SpatialLines-method}
\alias{layerNames,SpatialPoints-method}
\alias{layerNames,SpatialPolygons-method}
\alias{layerNames,gg-method}
\alias{layerNames,histogram-method}
\alias{layerNames,list-method}
\title{Extract the layer names of Spatial Objects}
\usage{
layerNames(object)

\S4method{layerNames}{list}(object)

\S4method{layerNames}{SpatialPoints}(object)

\S4method{layerNames}{SpatialPolygons}(object)

\S4method{layerNames}{SpatialLines}(object)

\S4method{layerNames}{Raster}(object)

\S4method{layerNames}{gg}(object)

\S4method{layerNames}{histogram}(object)
}
\arguments{
\item{object}{A \code{Raster*}, \code{SpatialPoints*}, \code{SpatialLines*}, or \code{SpatialPolygons*} object; or list of these.}
}
\description{
There are methods for \code{Raster*}, \code{SpatialPoints*}, \code{SpatialLines*}
and \code{SpatialPolygons*},
though the latter return an empty character vector of length 1.
}
\author{
Eliot McIntire
}

