% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{layerNames}
\alias{layerNames}
\alias{layerNames,Raster-method}
\alias{layerNames,SpatialPoints-method}
\alias{layerNames,SpatialPolygons-method}
\alias{layerNames,list-method}
\title{Extract the layer names of Spatial Objects}
\usage{
layerNames(object)

\S4method{layerNames}{list}(object)

\S4method{layerNames}{SpatialPoints}(object)

\S4method{layerNames}{SpatialPolygons}(object)

\S4method{layerNames}{Raster}(object)
}
\arguments{
\item{object}{a Raster*, SpatialPoints*, SpatialPolygons* or list of these.}
}
\description{
There are methods for Raster*, SpatialPoints*, and SpatialPolygons*, though the latter
return an empty character vector of length 1.
}

