% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{layerNames}
\alias{layerNames}
\alias{layerNames,Raster-method}
\alias{layerNames,SpatialPoints-method}
\alias{layerNames,SpatialPointsDataFrameNamed-method}
\alias{layerNames,SpatialPointsNamed-method}
\alias{layerNames,SpatialPolygons-method}
\alias{layerNames,SpatialPolygonsNamed-method}
\alias{layerNames,list-method}
\title{Extract the layer names of Spatial Objects}
\usage{
layerNames(object)

\S4method{layerNames}{list}(object)

\S4method{layerNames}{SpatialPointsNamed}(object)

\S4method{layerNames}{SpatialPoints}(object)

\S4method{layerNames}{SpatialPolygonsNamed}(object)

\S4method{layerNames}{SpatialPolygons}(object)

\S4method{layerNames}{SpatialPointsDataFrameNamed}(object)

\S4method{layerNames}{Raster}(object)
}
\description{
There are methods for Raster*, SpatialPoints*, and SpatialPolygons*, though the latter
return an empty character vector of length 1.
}

