% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{loadFiles}
\alias{.fileExtensions}
\alias{loadFiles}
\alias{loadFiles,missing,ANY-method}
\alias{loadFiles,missing,missing-method}
\alias{loadFiles,simList,missing-method}
\title{Load simulation objects according to fileList}
\usage{
loadFiles(sim, fileList, ...)

\S4method{loadFiles}{simList,missing}(sim, fileList, ...)

\S4method{loadFiles}{missing,ANY}(sim, fileList, ...)

\S4method{loadFiles}{missing,missing}(sim, fileList, ...)

.fileExtensions()
}
\arguments{
\item{sim}{A \code{simList} object}

\item{fileList}{List or data.frame to call loadFiles directly from the fileList as
described in Details}

\item{...}{Additional arguments.}
}
\description{
This function takes the fileList argument in the simList object and loads all the files
using the identified functions and arguments

How to load various types of files in R.
}
\details{
In the fileList object, either a list or a data.frame, there will be minimally a column called "files".
All other columns are optional.

Other optional columns are:

- \code{objectNames}: a character string indicating the name of the object once the
file is loaded. Default is to use the file names, with file extension removed.

- \code{functions}: a character string indicating the function to be used to load the file.
Default is to use the mapping between file extensions in the \code{.fileExtensions} function
and the actual file extensions.

- \code{intervals}: a numeric indicating the interval between repeated loading of the same
file. This should be NA or the column absent if the file is only loaded once. Default is
absent, so files are loaded only at \code{start} in the \code{simList}.

- \code{loadTime}: a numeric indicating when the file should be loaded. Defaults to
simTime=0,but this can be any time. The loading will be scheduled to occur
at the "loadTime", whatever that is. If the same file is to loaded many times,
but not at a regular interval, then there should be separate line, with a unique
loadTime for each.

- \code{.stackName}: a character (vector) indicating the name(s) of the RasterStack(s)
to stack all the RasterLayers into. If left blank, then the individual RasterLayers will
not be automatically stacked. If this is specified, the individual RasterLayers loaded
will not actually be created as individual R objects; they will be loaded directly
into the stack. Finally, the behaviour here is

- \code{arguments}: is a list of lists of named arguments, one list for each loading function.
For example, if raster is a loading function, arguments = list(native = TRUE). If there is only
one list, then it is assumed to apply to all load attempts and will be repeated
for each load function.
}
\examples{
#load random maps included with package
fileList = data.frame(files = dir(file.path(find.package("SpaDES", quiet = FALSE), "maps"),
   full.names=TRUE, pattern= "tif"), functions="rasterToMemory", package="SpaDES",
   stringsAsFactors=FALSE)

loadFiles(fileList=fileList)
Plot(DEM)

# Second, more sophisticated. All maps loaded at time = 0, and the last one is reloaded
#  at time = 10 (via "intervals"). Also, pass the single argument as a list to all functions...
#  specifically, when add "native = TRUE" as an argument to the raster function
arguments = list(native=TRUE)
files = dir(file.path(find.package("SpaDES", quiet = FALSE), "maps"),
     full.names=TRUE, pattern= "tif")
fileList = data.frame(
   files = files,
   functions="rasterToMemory",
   packages="SpaDES",
   objectNames = NA,
   arguments = arguments,
   loadTimes = 0,
   intervals = c(rep(NA, length(files)-1), 10),
   stringsAsFactors=FALSE)

loadFiles(fileList=fileList)
}
\author{
Eliot McIntire

Alex Chubaty
}

