% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{reload.module.later}
\alias{reload.module.later}
\alias{reload.module.later,ANY,character-method}
\title{Load modules for simulation.}
\usage{
reload.module.later(sim, depends)

\S4method{reload.module.later}{ANY,character}(sim, depends)
}
\arguments{
\item{sim}{An object of class \code{SimList}.}

\item{depends}{A list of character strings specifying
the names of modules upon which the current module depends.}
}
\value{
\code{TRUE}/\code{FALSE}.
}
\description{
Checks the dependencies of the current module on other modules.
These dependencies need to be loaded first, so if they are not
already loaded, hold off loading the cuurent module until after
dependencies are loaded.
}
\seealso{
\code{\link{library}}.
}

