% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{ls.str.simList}
\alias{ls.str.simList}
\alias{ls.str,missing,simList-method}
\alias{ls.str,simList,missing-method}
\title{List simulation objects and their structure}
\usage{
ls.str.simList(name)

\S4method{ls.str}{missing,simList}(name)

\S4method{ls.str}{simList,missing}(pos)
}
\arguments{
\item{name}{A \code{simList} object.}

\item{pos}{A \code{simList} object, used only if \code{name} not provided.}
}
\description{
A variation of applying \code{\link{str}} to each matched name.
Can be used with a \code{simList} object, because the method for this class
is simply a wrapper for calling \code{ls} on the simulation environment
stored in the \code{simList} object.

export
}
\seealso{
Other functions to access elements of a \code{simList} object: \code{\link{.addDepends}},
  \code{\link{doEvent.checkpoint}}, \code{\link{envir}},
  \code{\link{events}}, \code{\link{globals}},
  \code{\link{inputs}}, \code{\link{ls.simList}},
  \code{\link{modules}}, \code{\link{objs}},
  \code{\link{packages}}, \code{\link{params}},
  \code{\link{paths}}, \code{\link{progressInterval}},
  \code{\link{times}}
}
