% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{.makeColorMatrix}
\alias{.makeColorMatrix}
\title{Convert Raster to color matrix useable by raster function for plotting}
\usage{
.makeColorMatrix(grobToPlot, zoomExtent, maxpixels, legendRange, cols = NULL,
  na.color = "#FFFFFF00", zero.color = NULL, skipSample = TRUE)
}
\arguments{
\item{grobToPlot}{a SpatialObject}

\item{zoomExtent}{an extent object for zooming to. Defaults to whole extent of grobToPlot}

\item{maxpixels}{numeric. Number of cells to subsample the complete grobToPlot}

\item{legendRange}{numeric vector giving values that, representing the lower
and upper bounds of a legend (i.e., \code{1:10} or \code{c(1,10)} will give same result)
that will override the data bounds contained within the \code{grobToPlot}.}

\item{cols}{colours specified in a way that can be understood directly or by
 \code{\link{colorRampPalette}}.}

\item{na.color}{string indicating the color for \code{NA} values. Default transparent.}

\item{zero.color}{string indicating the color for zero values, when zero is
the minimum value. Otherwise, it is treated as any other color. Default transparent.
Use \code{NULL} if zero should be the value given to it by the colortable associated with the Raster.}

\item{skipSample}{logical. If no downsampling is necessary, skip. Default \code{TRUE}.}
}
\description{
Convert Raster to color matrix useable by raster function for plotting
}
\author{
Eliot McIntire
}

