% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{makeDigestible}
\alias{makeDigestible}
\title{makeDigestible for simList class objects}
\usage{
makeDigestible(object, objects, compareRasterFileLength = 1e+06,
  algo = "xxhash64")
}
\description{
See \code{\link[reproducible]{makeDigestible}}. This strips out stuff
from a simList class object that would make it otherwise not
reproducibly digestible between sessions, operating systems,
or machines. This will likely still not allow identical digest
results across R versions.
}
