% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{methods}
\name{makeDigestible}
\alias{makeDigestible}
\alias{makeDigestible,simList-method}
\title{makeDigestible for simList class objects}
\usage{
makeDigestible(object, objects, compareRasterFileLength = 1e+06,
  algo = "xxhash64")

\S4method{makeDigestible}{simList}(object, objects,
  compareRasterFileLength = 1e+06, algo = "xxhash64")
}
\arguments{
\item{object}{an object to convert to a 'digestible' state}

\item{objects}{Optional character vector indicating which objects are to
be considered while making digestible.}

\item{compareRasterFileLength}{Numeric. Optional. When there are Rasters, that
have file-backed storage, this is passed to the length arg in \code{digest}
when determining if the Raster file is already in the database.
Note: uses \code{\link[digest]{digest}} for file-backed Raster.
Default 1e6. Passed to \code{prepareFileBackedRaster}.}

\item{algo}{The algorithms to be used; currently available choices are
    \code{md5}, which is also the default, \code{sha1}, \code{crc32},
    \code{sha256}, \code{sha512}, \code{xxhash32}, \code{xxhash64} and
    \code{murmur32}.}
}
\description{
See \code{\link[reproducible]{makeDigestible}}. This strips out stuff
from a simList class object that would make it otherwise not
reproducibly digestible between sessions, operating systems,
or machines. This will likely still not allow identical digest
results across R versions.
}
\seealso{
\code{\link[reproducible]{makeDigestible}}
}
