% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{makeList}
\alias{makeList}
\alias{makeList,list-method}
\title{Convert plotArgs to list of lists}
\usage{
makeList(plotArgs, numSpadesPlotObjects)

\S4method{makeList}{list}(plotArgs, numSpadesPlotObjects)
}
\arguments{
\item{plotArgs}{}
}
\description{
Take the inputs as plotArgs to the Plot function, and make them list entries
of lists
}
\author{
Eliot McIntire
}

