% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{.makeList}
\alias{.makeList}
\alias{.makeList,list-method}
\title{Convert plotArgs to list of lists}
\usage{
.makeList(plotArgs, numSpadesPlotObjects)

\S4method{.makeList}{list}(plotArgs, numSpadesPlotObjects)
}
\arguments{
\item{plotArgs}{list. The arguments passed to Plot}

\item{numSpadesPlotObjects}{numeric. The number of .spadesPlotObjects. This can't
easily be deduced from the plotArgs because of the RasterStacks. So passed manually.}
}
\description{
Take the inputs as plotArgs to the Plot function, and make them a list of length
\code{numSpadesPlotObjects} entries of lists. To be called internally only.
}
\author{
Eliot McIntire
}

