% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{makeSpadesPlot}
\alias{makeSpadesPlot}
\alias{makeSpadesPlot,list,list-method}
\alias{makeSpadesPlot,list,missing-method}
\alias{makeSpadesPlot,missing,missing-method}
\title{Make SpaDES Plot object}
\usage{
makeSpadesPlot(plotObjects, plotArgs, ...)

\S4method{makeSpadesPlot}{list,list}(plotObjects, plotArgs, ...)

\S4method{makeSpadesPlot}{list,missing}(plotObjects, plotArgs, ...)

\S4method{makeSpadesPlot}{missing,missing}(plotObjects, plotArgs, ...)
}
\arguments{
\item{plotObjects}{list. Any plot objects, whether a legend should be drawn. Default \code{TRUE}.}

\item{plotArgs}{list. Any arguments that the the grid package can accept for the specific
grob types, e.g., rasterGrob, polygonGrob, etc.}

\item{...}{additional arguments. Currently nothing.}
}
\description{
Builds a SpaDES plot object from a list of objects
}

