% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{makeSpadesPlot}
\alias{Plot,missing-method}
\alias{makeSpadesPlot}
\alias{makeSpadesPlot,list,list-method}
\alias{makeSpadesPlot,list,missing-method}
\alias{makeSpadesPlot,missing,missing-method}
\title{Make SpaDES Plot object}
\usage{
makeSpadesPlot(plotObjects, plotArgs, ...)

\S4method{makeSpadesPlot}{list,list}(plotObjects, plotArgs, ...)

\S4method{makeSpadesPlot}{list,missing}(plotObjects, plotArgs, ...)

\S4method{makeSpadesPlot}{missing,missing}(plotObjects, plotArgs, ...)

\S4method{Plot}{missing}(..., new = FALSE, addTo = NULL, gp = gpar(),
  gpText = gpar(), gpAxis = gpar(), axes = "L", speedup = 1, size = 5,
  cols = NULL, zoomExtent = NULL, visualSqueeze = 0.75, legend = TRUE,
  legendRange = NULL, legendText = NULL, draw = TRUE, pch = 19,
  title = TRUE, na.color = "#FFFFFF00", zero.color = "#FFFFFF00")
}
\arguments{
\item{plotObjects}{list. Any plot objects, whether a legend should be drawn. Default \code{TRUE}.}

\item{plotArgs}{list. Any arguments that the the grid package can accept for the specific
grob types, e.g., rasterGrob, polygonGrob, etc.}

\item{...}{additional arguments. Currently nothing.}
}
\description{
Builds a SpaDES plot object from a list of objects
}

