% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{.makeViewports}
\alias{.makeViewports}
\title{Make viewports}
\usage{
.makeViewports(spadesPlot, zoom, newArr = FALSE)
}
\arguments{
\item{spadesPlot}{An object of class \code{spadesPlot}}

\item{zoom}{[description needed]}

\item{newArr}{Logical. Whether this function will create a completely new viewport.
               Default \code{FALSE}.}
}
\description{
Given a set of extents, and a layout for these extents, this function will output
a viewport tree to allow plotting.
}
\details{
This function will either create a totally new set of viewports, or simply add
some nested viewports to an existing arrangement, i.e., is there still white
space availabe to plot.
}
\author{
Eliot McIntire
}

