% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{.makeViewports}
\alias{.makeViewports}
\title{Make viewports}
\usage{
.makeViewports(sPlot, newArr = FALSE)
}
\arguments{
\item{sPlot}{An object of class \code{.spadesPlot}.}

\item{newArr}{Logical indicating whether this function will create a
               completely new viewport. Default \code{FALSE}.

# @importFrom NetLogoRClasses extent}
}
\description{
Given a set of extents, and a layout for these extents, this function will
output a viewport tree to allow plotting.
}
\details{
This function will either create a totally new set of viewports, or simply
add some nested viewports to an existing arrangement, i.e., is there still
white space available to plot.
}
\author{
Eliot McIntire
}
\keyword{internal}
