% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread-process.R
\name{\%fin\%}
\alias{\%fin\%}
\title{A faster '%in%' based on fastmatch package}
\usage{
x \%fin\% table
}
\arguments{
\item{x}{See \code{fmatch} in fastmatch package}

\item{table}{See \code{fmatch} in fastmatch package}
}
\description{
A faster '%in%', directly pulled from fastmatch match, based on
http://stackoverflow.com/questions/32934933/faster-in-operator
}

