% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeRaster.R
\docType{methods}
\name{mergeRaster}
\alias{mergeRaster}
\alias{mergeRaster,list-method}
\title{Merge the splitted tiles to one raster layer.}
\usage{
mergeRaster(x)

\S4method{mergeRaster}{list}(x)
}
\arguments{
\item{x}{A list of splitted raster tiles with or without buffer.}
}
\value{
A original raster layer before it is splitted.
}
\description{
Recombine the splitted tiles from \code{splitRaster} to
the original raster layer.
}
\examples{
require(raster)
# an example with dimensions:
# nrow = 77
# ncol = 101
# nlayers = 3
b <- brick(system.file("external/rlogo.grd", package = "raster"))
r <- b[[1]] # use first layer only
nx <- 3
ny <- 4
y0 <- splitRaster(r, nx, ny) # without buffer
y1 <- splitRaster(r, nx, ny, c(3, 4)) # with 3 pixels buffer at horizontal scale
                                      # and 4 pixels buffer at vertical scale


# can be recombined using `raster::merge`
r0 <- do.call(merge, y0) # regular merge call from raster package
all.equal(r0, r) # TRUE

r1 <- mergeRaster(y0) # mergeRaster function for splitted tiles without buffer
all.equal(r1, r) # TRUE

r2 <- mergeRaster(y1) # mergeRaster function for splitted tiles with buffer
all.equal(r2, r) # TRUE

}
\author{
Yong Luo
}
\seealso{
\code{\link{merge}}, \code{\link{splitRaster}}.
}

