% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/times.R
\docType{methods}
\name{minTimeunit}
\alias{minTimeunit}
\alias{minTimeunit,simList-method}
\title{Determine what the smallest timestepUnit in a simObject}
\usage{
minTimeunit(sim)

\S4method{minTimeunit}{simList}(sim)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}
}
\value{
The timestepUnit as a character string. This defaults to "second" if
none of the modules has explicit units.
}
\description{
When modules have different timestepUnit, SpaDES automatically takes the
largest (e.g., "year") as the unit for a simulation. This function determines which
is the largest unit.
}
\author{
Eliot McIntire
}

