% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\docType{methods}
\name{minTimeunit}
\alias{minTimeunit}
\alias{minTimeunit,simList-method}
\alias{minTimeunit,list-method}
\title{Determine the smallest timeunit in a simulation}
\usage{
minTimeunit(sim)

\S4method{minTimeunit}{simList}(sim)

\S4method{minTimeunit}{list}(sim)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}
}
\value{
The timeunit as a character string. This defaults to "second" if
none of the modules has explicit units.
}
\description{
When modules have different timeunit, SpaDES automatically takes the smallest
(e.g., "second") as the unit for a simulation.
}
\author{
Eliot McIntire
}
