% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting-diagrams.R
\docType{methods}
\name{moduleDiagram}
\alias{moduleDiagram}
\alias{moduleDiagram,simList,ANY-method}
\alias{moduleDiagram,simList,missing-method}
\title{Simulation module dependency diagram}
\usage{
moduleDiagram(sim, type, ...)

\S4method{moduleDiagram}{simList,ANY}(sim, type, ...)

\S4method{moduleDiagram}{simList,missing}(sim, type, ...)
}
\arguments{
\item{sim}{A \code{simList} object (typically corresponding to a
            completed simulation).}

\item{type}{Any or missing, but only responds to a character string, either \code{rgl}
or \code{tk} for rgl plot or tk plot, each calls igraphh::rglplot or igraphh::tkplot,
respectively. Default missing.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\value{
Plots module dependency diagram.
}
\description{
Create a network diagram illustrating the simplified module dependencies of a
simulation. Offers a less detailed view of specific objects than does
plotting the \code{depsEdgeList} directly with \code{\link{objectDiagram}}.
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{igraph}}.
}

