% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{name,SpatialPointsNamed-method}
\alias{name,SpatialPointsDataFrameNamed-method}
\alias{name,SpatialPointsNamed-method}
\alias{name<-}
\title{set name of SpatialPoints and SpatialPointsDataFrames}
\usage{
\S4method{name}{SpatialPointsNamed}(object)

\S4method{name}{SpatialPointsDataFrameNamed}(object)

name(object) <- value
}
\description{
set name of SpatialPoints and SpatialPointsDataFrames
}

