\docType{methods}
\name{name,SpatialPointsNamed-method}
\alias{name,RasterStackNamed-method}
\alias{name,SpatialPointsDataFrameNamed-method}
\alias{name,SpatialPointsNamed-method}
\alias{name<-}
\title{set name of SpatialPoints and SpatialPointsDataFrames}
\usage{
\S4method{name}{SpatialPointsNamed}(object)

\S4method{name}{SpatialPointsDataFrameNamed}(object)

\S4method{name}{RasterStackNamed}(object)

name(object) <- value
}
\description{
set name of SpatialPoints and SpatialPointsDataFrames
}

