% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{newPlot}
\alias{newPlot}
\title{Open a new plotting window}
\usage{
newPlot(noRStudioGD = TRUE, ...)
}
\arguments{
\item{noRStudioGD}{Logical Passed to dev.new. Default is TRUE to avoid using
RStudio graphics device, which is slow.}

\item{...}{Additional arguments.}
}
\description{
Open a new plotting window
}
\note{
\code{\link{dev.new}} is supposed to be the correct way to open a new
window in a platform-generic way, however, this doesn't work in RStudio.
}
\author{
Eliot McIntire and Alex Chubaty
}
\seealso{
\code{\link{dev}}.
}

