% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{newPlot}
\alias{newPlot}
\title{Open a new plotting window}
\usage{
newPlot(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\description{
Launch a new graphics device based on operating system used.
Mac OS: open device with \code{quartz()}.
Linux: open device with \code{x11()}.
Windows: open device with \code{windows()}.
}
\note{
\code{\link{dev.new}} is supposed to be the correct way to open a new
window in a platform-generic way, however, this doesn't work in RStudio.
}
\seealso{
\code{\link{quartz}}, \code{\link{windows}}, \code{\link{x11}}.
}

