% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/progress.R
\docType{methods}
\name{newProgressBar}
\alias{newProgressBar}
\title{Progress bar}
\usage{
newProgressBar(sim)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}
}
\description{
Shows a progress bar that is scaled to simulation stop time.
}
\details{
The progress bar object is stored in a separate environment,
\code{.spadesEnv}.
}
\author{
Alex Chubaty

Eliot McIntire
}

