% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{nlayers,list-method}
\alias{nlayers,SpatialPoints-method}
\alias{nlayers,SpatialPolygons-method}
\alias{nlayers,list-method}
\title{Find the number of layers in a Spatial Object}
\usage{
\S4method{nlayers}{list}(x)

\S4method{nlayers}{SpatialPolygons}(x)

\S4method{nlayers}{SpatialPoints}(x)
}
\arguments{
\item{x}{A spatialObjects object or list of these.}
}
\description{
There are already methods for Raster* in the raster package. Adding methods
for list and SpatialPolygons and SpatialPoints. These latter classes return 1.
}

