% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{nlayers}
\alias{nlayers}
\alias{nlayers,SpatialPolygons-method}
\title{Find the number of layers in a Spatial Object}
\usage{
\S4method{nlayers}{list}(x)

\S4method{nlayers}{SpatialPolygons}(x)
}
\arguments{
\item{x}{Raster* object}
}
\description{
There are methods for Raster*. All other object classes return 1.
}

