% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{.objectNames}
\alias{.objectNames}
\title{Extracts the object names}
\usage{
.objectNames(calledFrom = "Plot", argClass = ".spadesPlotObjects",
  argName = "")
}
\arguments{
\item{calledFrom}{character vector, length 1, indicating which function call is
desired. Defaults to \code{Plot}}

\item{argClass}{character vector, length 1, indicating which class is being
searched for among the arguments. Defaults to \code{.spadesPlotObjects}}

\item{argName}{character vector, length 1, or \code{NULL}, indicating if the
arguments to select have a name, no name (empty string), or do not use name (NULL).}
}
\value{
\code{NULL}. This function is invoked for its side effects.
}
\description{
This is primarily used from Plot.
}
\author{
Eliot McIntire
}

