% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{packages}
\alias{packages}
\alias{packages,.simList-method}
\alias{packages,missing-method}
\title{Get module or simulation package dependencies}
\usage{
packages(sim, ...)

\S4method{packages}{.simList}(sim, ...)

\S4method{packages}{missing}(sim, ...)
}
\arguments{
\item{sim}{A \code{simList} object.}

\item{...}{Additional arguments.
Currently only \code{module}, specifying the name of a module,
and \code{filename}, specifying a module filename, are supported.}
}
\value{
A sorted character vector of package names.
}
\description{
Get module or simulation package dependencies
}
\author{
Alex Chubaty
}
\seealso{
Other functions to access elements of a \code{simList} object: \code{\link{.addDepends}},
  \code{\link{doEvent.checkpoint}}, \code{\link{envir}},
  \code{\link{events}}, \code{\link{globals}},
  \code{\link{inputs}}, \code{\link{ls.simList}},
  \code{\link{ls.str.simList}}, \code{\link{modules}},
  \code{\link{objs}}, \code{\link{params}},
  \code{\link{paths}}, \code{\link{progressInterval}},
  \code{\link{times}}
}

