% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{.parseModule}
\alias{.parseModule}
\alias{.parseModule,simList,list-method}
\title{Parse and initialize a module}
\usage{
.parseModule(sim, modules)

\S4method{.parseModule}{simList,list}(sim, modules)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}

\item{modules}{A list of modules with a logical attribute "parsed".}
}
\value{
A \code{simList} simulation object.
}
\description{
Internal function, used during \code{\link{simInit}}.
}
\author{
Alex Chubaty
}

