% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{.parseTimeunit}
\alias{.parseModule}
\alias{.parseModule,simList,list-method}
\alias{.parseTimeunit}
\alias{.parseTimeunit,simList,list-method}
\title{Parse and initialize a module}
\usage{
.parseTimeunit(sim, modules)

\S4method{.parseTimeunit}{simList,list}(sim, modules)

.parseModule(sim, modules)

\S4method{.parseModule}{simList,list}(sim, modules)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}

\item{modules}{A list of modules with a logical attribute "parsed".}
}
\value{
.parseTimeunit just returns the time units of each module

A \code{simList} simulation object.
}
\description{
Internal function, used during \code{\link{simInit}}.
}
\author{
Alex Chubaty

Alex Chubaty
}

