\docType{methods}
\name{placeOnRaster}
\alias{placeOnRaster}
\alias{placeOnRaster,NamedSpatialPoints-method}
\title{Makes a raster around a SpatialPoints object}
\usage{
placeOnRaster(obj, raster = NULL)

\S4method{placeOnRaster}{NamedSpatialPoints}(obj, raster)
}
\description{
Makes a raster around a SpatialPoints object
}

