% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plotGrob}
\alias{plotGrob}
\alias{plotGrob,SpatialPoints-method}
\alias{plotGrob,SpatialPolygons-method}
\alias{plotGrob,matrix-method}
\title{Plot spatial grobs (using grid package)}
\usage{
plotGrob(grobToPlot, col = NULL, size = unit(5, "points"), minv, maxv,
  legend = TRUE, draw = TRUE, gp = gpar(), pch = 19, speedup = 1, ...)

\S4method{plotGrob}{matrix}(grobToPlot, col = NULL, size = unit(5,
  "points"), minv, maxv, legend = TRUE, draw = TRUE, gp = gpar(),
  pch = 19, speedup = 1, ...)

\S4method{plotGrob}{SpatialPoints}(grobToPlot, col = NULL, size = unit(5,
  "points"), minv, maxv, legend = TRUE, draw = TRUE, gp = gpar(),
  pch = 19, speedup = 1, ...)

\S4method{plotGrob}{SpatialPolygons}(grobToPlot, col = NULL, size = unit(5,
  "points"), minv, maxv, legend = TRUE, draw = TRUE, gp = gpar(),
  pch = 19, speedup = 1, ...)
}
\arguments{
\item{grobToPlot}{Raster*, SpatialPoints*, SpatialPolygons* object}

\item{col}{Currently only used for the legend of a Raster* object.}

\item{size}{The size of the SpatialPoints}

\item{minv}{The minimum value on a Raster*. Required because not all Rasters
have this defined internally}

\item{maxv}{The maximum value on a Raster*. Required because not all Rasters
have this defined internally}

\item{legend}{logical, whether a legend should be drawn. Default TRUE.}

\item{draw}{logical, whether the grob, after being created should be drawn. Default TRUE.}

\item{gp}{grid parameters, usually the output of a call to \code{\link{gpar}}}

\item{pch}{for SpatialPoints, as \code{par}}

\item{speedup}{Numeric. The factor by which the number of vertices in SpatialPolygons will be subsampled.
The vertices are already subsampled by default to make plotting faster.}

\item{...}{additional arguments. Currently nothing.}
}
\description{
Plot a raster Grob, a points Grob, polygon Grob. This should mostly be called internally.
}
\details{
\code{speedup} is only used for SpatialPolygons in this function. Attempts have been made
to subsample at a good level that optimizes speed of plotting, without losing visible
quality. From a speed perspective, there appears to be an optimal subsampling
when using \code{thin} from fastshp. Presumably too much thinning requires
large distance matrices to be calculated, slowing plotting down. Too little thinning
causes an overabundance of points to be plotted, slowing plotting down.
}

