% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{.plotGrob}
\alias{.plotGrob}
\alias{.plotGrob,SpatialLines-method}
\alias{.plotGrob,SpatialPoints-method}
\alias{.plotGrob,SpatialPolygons-method}
\alias{.plotGrob,matrix-method}
\title{Plot spatial grobs (using grid package)}
\usage{
.plotGrob(grobToPlot, col = NULL, real = FALSE, size = unit(5, "points"),
  minv, maxv, legend = TRUE, legendText = NULL, gp = gpar(),
  gpText = gpar(), pch = 19, speedup = 1, ...)

\S4method{.plotGrob}{matrix}(grobToPlot, col = NULL, real = FALSE,
  size = unit(5, "points"), minv, maxv, legend = TRUE, legendText = NULL,
  gp = gpar(), gpText = gpar(), pch = 19, speedup = 1, ...)

\S4method{.plotGrob}{SpatialPoints}(grobToPlot, col = NULL, real = FALSE,
  size = unit(5, "points"), minv, maxv, legend = TRUE, legendText = NULL,
  gp = gpar(), gpText = gpar(), pch = 19, speedup = 1, ...)

\S4method{.plotGrob}{SpatialPolygons}(grobToPlot, col = NULL, real = FALSE,
  size = unit(5, "points"), minv, maxv, legend = TRUE, legendText = NULL,
  gp = gpar(), gpText = gpar(), pch = 19, speedup = 1, ...)

\S4method{.plotGrob}{SpatialLines}(grobToPlot, col, size, legend, gp = gpar(),
  pch, length, ...)
}
\arguments{
\item{grobToPlot}{\code{Raster*}, \code{SpatialPoints*}, \code{SpatialPolygons*}, or
\code{SpatialLines*} object}

\item{col}{Currently only used for the legend of a \code{Raster*} object.}

\item{real}{logical. Whether the data are real numbers (vs. integer or factor).}

\item{size}{The size of the SpatialPoints}

\item{minv}{The minimum value on a Raster*. Required because not all Rasters
have this defined internally}

\item{maxv}{The maximum value on a Raster*. Required because not all Rasters
have this defined internally}

\item{legend}{logical, whether a legend should be drawn. Default \code{TRUE}.}

\item{legendText}{Vector of values to use for legend value labels. Defaults to \code{NULL} which results
in a pretty numeric representation. If \code{Raster*} has a Raster Attribute Table (rat, see raster
package), this will be used by default. Currently, only a single vector is accepted.}

\item{gp}{grid parameters, usually the output of a call to \code{\link{gpar}}}

\item{gpText}{gpar object for legend label text}

\item{pch}{for \code{SpatialPoints}, as \code{par}}

\item{speedup}{Numeric. The factor by which the number of vertices in \code{SpatialPolygons}
and \code{SpatialLines*} will be subsampled. The vertices are already subsampled
by default to make plotting faster.}

\item{...}{additional arguments. Currently nothing.}

\item{length}{Numeric.}
}
\description{
Plot a raster Grob, a points Grob, polygon Grob. This should mostly be called internally.
}
\details{
\code{speedup} is only used for SpatialPolygons in this function. Attempts have been made
to subsample at a good level that optimizes speed of plotting, without losing visible
quality. From a speed perspective, there appears to be an optimal subsampling
when using \code{thin} from the \code{fastshp} package. Presumably too much thinning requires
large distance matrices to be calculated, slowing plotting down. Too little thinning
causes an overabundance of points to be plotted, slowing plotting down.

The suggested package \code{fastshp} can be installed with:
  \code{devtools::install_github("s-u/fastshp")}

NOTE: you may get errors relating to not having installed the software tools
required for building R packages on your system. For development purposes on
a Windows machine, you'll need to install Rtools from http://cran.r-project.org/bin/windows/Rtools/.
}
\author{
Eliot McIntire
}

