% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plotGrob}
\alias{plotGrob}
\alias{plotGrob,SpatialPoints-method}
\alias{plotGrob,matrix-method}
\title{Plot either a raster Grob or a points Grob}
\usage{
plotGrob(grobToPlot, col = NULL, size = unit(5, "points"), name = "plot",
  minv, maxv, legend = TRUE, addedNACol = addedNACol, draw = TRUE,
  gp = gpar(), vp = NULL, pch = 19, childrenvp = NULL, ...)

\S4method{plotGrob}{matrix}(grobToPlot, col = NULL, size = unit(5,
  "points"), name = "plot", minv, maxv, legend = TRUE,
  addedNACol = addedNACol, draw = TRUE, gp = gpar(), vp = NULL,
  pch = 19, childrenvp = NULL, ...)

\S4method{plotGrob}{SpatialPoints}(grobToPlot, col = NULL, size = unit(5,
  "points"), name = "plot", minv, maxv, legend = TRUE,
  addedNACol = addedNACol, draw = TRUE, gp = gpar(), vp = NULL,
  pch = 19, childrenvp = NULL, ...)
}
\arguments{
\item{grobToPlot}{Raster* or SpatialPoints* object}
}
\description{
Plot either a raster Grob or a points Grob
}

