% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{pointAgent}
\alias{pointAgent}
\alias{pointAgent-class}
\title{The \code{pointAgent} class.}
\description{
Extends the \code{vectorAgent} class. Describes spatial agents whose spatial
components are characterized by points (i.e., in contrast to a polygon).
}
\section{Slots}{

\describe{
\item{\code{ID}}{A character string that identifies the individual agent.
             By default, the numbers 1...n are used.}

\item{\code{spatial}}{A \code{SpatialPoints} object.}

\item{\code{other}}{A list for storing other (user-specific) agent attributes.}
}}
\note{
add additional notes here.
}

