% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POM.R
\docType{methods}
\name{POM}
\alias{POM}
\alias{POM,simList,character-method}
\title{Use Pattern Oriented Modeling to fit unknown parameters}
\usage{
POM(sim, params, objects, objFn, cl, optimizer = "DEoptim", ...,
  objFnCompare = "MAD")

\S4method{POM}{simList,character}(sim, params, objects, objFn, cl,
  optimizer = "DEoptim", ..., objFnCompare = "MAD")
}
\arguments{
\item{sim}{A \code{simList} simulation object, generally produced by \code{simInit}.}

\item{objFn}{An objective function to be passed into
\code{optimizer}}

\item{cl}{A cluster object. Optional. This would generally be created using
parallel::makeCluster or equivalent. This is an alternative way, instead
of beginCluster(), to use parallelism for this function, allowing for
more control over cluster use.}

\item{optimizer}{The function to use to optimize. Default is
DEoptim}

\item{...}{All objects needed in objFn}

\item{objFnCompare}{Character string. Either, "MAD" or "RMSE" indicating that inside the objective
function, data and prediction will be compared by Mean Absolute Deviation or
Root Mean Squared Error. Default is "MAD".}
}
\value{
The values for parameters used in objFn that minimize
the objFn.
}
\description{
Use Pattern Oriented Modeling to fit unknown parameters
}
\examples{
\dontrun{
 set.seed(52461)
 library(parallel)
 mySim <- simInit(
   times = list(start = 0.0, end = 2.0, timeunit = "year"),
   params = list(
     .globals = list(stackName = "landscape", burnStats = "nPixelsBurned"),
     fireSpread = list(nFires = 1),
     randomLandscapes = list(nx = 200, ny = 200)
   ),
   modules = list("randomLandscapes", "fireSpread", "caribouMovement"),
   paths = list(modulePath = system.file("sampleModules", package = "SpaDES"))
 )
 out <- spades(copy(mySim), .plotInitialTime = NA)
 fireData <- out$landscape$Fires
 Plot(fireData, new=TRUE)

 cl <- makeCluster(6)
 POM(mySim, "spreadprob", list(fireData = "landscape$Fires"),
     cl = cl)

 N <- length(out$caribou)
 POM(mySim, c("spreadprob", "N"),
    list(fireData = "landscape$Fires",
         N = function(caribou) length(caribou)),
     cl = cl)
 stopCluster(cl)

 )
 }
}
\author{
Eliot McIntire
}
\references{
Matloff, N. (2011). The Art of R Programming (ch. 7.8.3). San Fransisco, CA: No Starch Press, Inc.. Retrieved from \url{https://www.nostarch.com/artofr.htm}
}
\seealso{
\code{\link{spades}}, \code{\link[parallel]{makeCluster}},
\code{\link{simInit}}
}

