% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{.prepareRaster}
\alias{.prepareRaster}
\title{Prepare raster for plotting}
\usage{
.prepareRaster(grobToPlot, zoomExtent, legendRange, takeFromPlotObj, arr,
  speedup, newArr)
}
\arguments{
\item{zoomExtent}{[description needed]}

\item{legendRange}{[description needed]}

\item{takeFromPlotObj}{[description needed]}

\item{arr}{[description needed]}

\item{speedup}{[description needed]}

\item{newArr}{[description needed]}

\item{gropToPlot}{[description needed]}
}
\description{
Prepare raster for plotting
}
\author{
Eliot McIntire
}

