% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{.prepareRaster}
\alias{.prepareRaster}
\title{Prepare raster for plotting}
\usage{
.prepareRaster(grobToPlot, zoomExtent, legendRange, takeFromPlotObj, arr,
  speedup, newArr)
}
\arguments{
\item{grobToPlot}{.spadesGrob}

\item{zoomExtent}{an extent object}

\item{legendRange}{a numeric vector of length >=2 indicating the desired legend range.}

\item{takeFromPlotObj}{logical. Should the object be found in the Plot call or .GlobalEnv}

\item{arr}{an \code{.arrangement} object}

\item{speedup}{numeric, greater than 1 will usually speed up plotting at the expense of resolution}

\item{newArr}{logical, whether this is a new arrangement or just adding to a previous one}
}
\description{
This is to be used internally. This takes a raster .spadesGrob, and converts zoomExtent
into a zoom, legendRange into a legend, and calculates the maxpixels to plot for speed.
}
\author{
Eliot McIntire
}

