% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialize.R
\docType{methods}
\name{randomPolygons}
\alias{randomPolygons}
\title{randomPolygons}
\usage{
randomPolygons(ras = raster(extent(0, 100, 0, 100), res = 1), p = 0.1,
  A = 0.3, speedup = 1, numTypes = 1, minpatch = 10, ...)
}
\arguments{
\item{ras}{A raster that whose extent will be used for the randomPolygons}

\item{p}{numeric vector. Parameter to control fragmentation. If this is a vector,
then there will be a polygon map produced with length(p) unique levels. See \code{randomLandscape}}

\item{A}{numeric vector. Parameter for expected proportion of habitat. If this is a vector,
then there will be a polygon map produced with length(A) unique levels. See \code{randomLandscape}}

\item{speedup}{An index of how much faster than normal to generate maps. This is achieved
by aggregating then disagregating at the end, so that the resulting raster is the
same extent as \code{ras}}

\item{numTypes}{numeric value. The number of unique polygon types to use. This will
be overridden by p, A or minpatch, if any of these are vectors.}

\item{minpatch}{numeric vector. Integer minimum size of patch. If this is a vector,
then there will be a polygon map produced with length(A) unique levels. See \code{randomLandscape}}

\item{...}{Additional arguments to \code{randomHabitat}.}
}
\value{
A map of extent \code{ext} with a random polygons.
}
\description{
Produces a raster of with random polygons of varying parameters, using the
The Modified Random Cluster algorithm of Saura and Martinez-Millan (2000).
}
\details{
This is a wrapper for the \code{secr::randomHabitat} function in the secr
package. The two main additions are the \code{speedup} argument which allows
for faster map generation for large rasters and addition of multiple unique polygon
values, using code drawn from http://www.guru-gis.net/generate-a-random-landscape/.
}
\examples{
r1 <- randomPolygons(p=c(0.1, 0.3, 0.5), A=0.3, minpatch=2)
Plot(r1, cols=c("white","dark green","blue","dark red"), new=TRUE)
}
\seealso{
\code{\link{randomHabitat}} and \code{\link{raster}}
}

