\docType{class}
\name{rasterAgent}
\alias{rasterAgent}
\alias{rasterAgent-class}
\title{The \code{rasterAgent} class.}
\description{
Extends the \code{agent} class by making it spatial.
Describes spatial agents whose spatial components are
characterized by a raster (i.e., in contrast to points or a
polygon).
}
\details{
This is a virtual class and not intended to be used
directly.
}
\section{Slots}{

\describe{
\item{\code{ID}}{A character string that identifies the individual agent.
             By default, the numbers 1...n are used.}

\item{\code{other}}{A list for storing other (user-specific) agent attributes.}
}}
\note{
This is a virtual class and not intended to be used
directly.
}

