% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sampleV}
\alias{sampleV}
\title{Vectorized version of \code{sample} using \code{Vectorize}}
\usage{
sampleV(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{Either a vector of one or more elements from which to choose,
    or a positive integer.  See \sQuote{Details.}}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{Should sampling be with replacement?}

\item{prob}{A vector of probability weights for obtaining the elements
    of the vector being sampled.}
}
\value{
A random perumtation, as in \code{sample}, but with \code{size} vectorized.
}
\description{
Intended for internal use only. \code{size} is vectorized.
}
\keyword{internal}
