% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\docType{methods}
\name{saveFiles}
\alias{saveFiles}
\title{Save simulation objects according to params}
\usage{
saveFiles(sim)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}
}
\description{
If there is a list entry with \code{.saveObjects} as a character string vector of
object names to save, then these objects will be saved with a call to saveFiles.
The file names will be equal to the object name plus
\code{time(sim)} is appended at the end. The files are saved as \code{.rds} files,
meaning, only one object gets saved per file.
}
\examples{
\dontrun{
 sim <- saveFiles(mySim)
}
}
\author{
Eliot McIntire

Alex Chubaty
}

