% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{saveFiles}
\alias{saveFiles}
\title{Save simulation objects according to simParams}
\usage{
saveFiles(sim)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}
}
\description{
Save simulation objects according to simParams
}
\author{
Eliot McIntire

Alex Chubaty
}

