% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{scheduleEvent}
\alias{scheduleEvent}
\alias{scheduleEvent,simList,NULL,character,character-method}
\alias{scheduleEvent,simList,numeric,character,character-method}
\title{Schedule a simulation event}
\usage{
scheduleEvent(sim, eventTime, moduleName, eventType)

\S4method{scheduleEvent}{simList,numeric,character,character}(sim, eventTime,
  moduleName, eventType)

\S4method{scheduleEvent}{simList,`NULL`,character,character}(sim, eventTime,
  moduleName, eventType)
}
\arguments{
\item{sim}{A \code{simList} simulation object.}

\item{eventTime}{A numeric specifying the time of the next event.}

\item{moduleName}{A character string specifying the module from which to call the event.}

\item{eventType}{A character string specifying the type of event from within the module.}
}
\value{
Returns the modified \code{simList} object.
}
\description{
Adds a new event to the simulation's event queue, updating the simulation object.
}
\details{
Based on code from chapter 7.8.3 of Matloff (2011): "Discrete event simulation".
Here, we implement a simulation in a more modular fashion so it's easier to add
submodules to the simulation. We use S4 classes and methods, and use `data.table`
instead of `data.frame` to implement the event queue (because it is much faster).
}
\examples{
\dontrun{scheduleEvent(x, 10.5, "firemodule", "burn")}
}
\author{
Alex Chubaty
}
\references{
Matloff, N. (2011). The Art of R Programming (ch. 7.8.3). San Fransisco, CA: No Starch Press, Inc.. Retrieved from \url{http://www.nostarch.com/artofr.htm}
}

