% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SELES.R
\docType{methods}
\name{agentLocation}
\alias{agentLocation}
\title{\code{SELES} - Agent Location at initiation}
\usage{
agentLocation(map)
}
\arguments{
\item{map}{A \code{SpatialPoints*}, \code{SpatialPolygons*}, or \code{Raster*} object.}
}
\value{
Object of same class as provided as input.
         If a \code{Raster*}, then zeros are converted to \code{NA}.
}
\description{
Sets the the location of the intiating agents. NOT YET FULLY IMPLEMENTED.

A \code{SELES}-like function to maintain conceptual backwards compatibility
with that simulation tool. This is intended to ease transitions from
\href{http://www.gowlland.ca/}{SELES}.

You must know how to use SELES for these to be useful.
}
\author{
Eliot McIntire
}
