% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SELES.R
\docType{methods}
\name{numAgents}
\alias{numAgents}
\title{SELES - Number of Agents to initiate}
\usage{
numAgents(N, probInit)
}
\arguments{
\item{N}{Number of agents to intitate (integer scalar).}

\item{probInit}{Probability of initalizing an agent at the location.}
}
\value{
A numeric, indicating number of agents to start
}
\description{
Sets the the number of agents to initiate. THIS IS NOT YET FULLY IMPLEMENTED.

A \code{SELES}-like function to maintain conceptual backwards compatability
with that simulation tool. This is intended to ease transitions from SELES.

You must know how to use SELES for these to be useful.
}
\author{
Eliot McIntire
}

