% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SELES.R
\docType{methods}
\name{numAgents}
\alias{numAgents}
\title{SELES - Number of Agents to initiate}
\usage{
numAgents(N)
}
\arguments{
\item{N}{Number of agents to intitate}
}
\value{
A numeric, indicating number of agents to start
}
\description{
A SELES-like function to maintain conceptual backwards compatability with that simulation
tool. Sets the the number of agents to initiate. THIS IS NOT FULLY IMPLEMENTED.
This is intended to ease transitinos from
\href{http://www.lfmi.uqam.ca/seles.htm}{SELES}.
You must know how to use SELES for these to be useful
}
\author{
Eliot McIntire
}

