% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SELES.R
\docType{methods}
\name{numAgents}
\alias{numAgents}
\title{SELES - Number of Agents to initiate}
\usage{
numAgents(N, probInit)
}
\arguments{
\item{N}{Number of agents to intitate (integer scalar).}

\item{probInit}{Probability of initalizing an agent at the location.}
}
\value{
A numeric, indicating number of agents to start
}
\description{
Sets the the number of agents to initiate. THIS IS NOT YET FULLY IMPLEMENTED.

A \code{SELES}-like function to maintain conceptual backwards compatibility
with that simulation tool. This is intended to ease transitions from
\href{http://www.gowlland.ca/}{SELES}.

You must know how to use SELES for these to be useful.
}
\author{
Eliot McIntire
}
