% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SELES.R
\docType{methods}
\name{transitions}
\alias{transitions}
\title{\code{SELES} - Transitioning to next time step}
\usage{
transitions(p, agent)
}
\arguments{
\item{p}{realized probability of persisting (i.e., either 0 or 1).}

\item{agent}{\code{SpatialPoints*} object.}
}
\value{
Returns new \code{SpatialPoints*} object with potentially fewer agents.
}
\description{
Describes the probability of an agent successfully persisting until next
time step. THIS IS NOT YET FULLY IMPLEMENTED.

A \code{SELES}-like function to maintain conceptual backwards compatibility
with that simulation tool. This is intended to ease transitions from
\href{http://www.gowlland.ca/}{SELES}.

You must know how to use SELES for these to be useful.
}
\author{
Eliot McIntire
}
