% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setColors<-}
\alias{setColors}
\alias{setColors<-}
\alias{setColors<-Raster,missing,list-method}
\alias{setColors<-Raster,numeric,list-method}
\alias{setColors<-RasterLayer,missing,character-method}
\alias{setColors<-RasterLayer,numeric,character-method}
\title{Set colours for plotting Raster* objects.}
\usage{
setColors(object, ..., n) <- value
}
\arguments{
\item{object}{A \code{Raster*} object.}

\item{...}{Additional arguments to \code{colorRampPalette}.}

\item{n}{An optional vector of values specifiying the number
             of levels from which to interpolate the color palette.}

\item{value}{Named list of hex color codes (e.g., from \code{RColorBrewer::brewer.pal}),
             corresponding to the names of RasterLayers in \code{x}.}
}
\value{
Returns a Raster with the colortable slot set to values.
}
\description{
Set colours for plotting Raster* objects.
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{brewer.pal}}, \code{\link{colorRampPalette}}.
}

