% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setColors<-}
\alias{setColors<-}
\alias{setColors<-Raster,list,missing-method}
\alias{setColors<-Raster,list,numeric-method}
\alias{setColors<-RasterLayer,character,missing-method}
\alias{setColors<-RasterLayer,character,numeric-method}
\title{Set colours for plotting Raster* objects.}
\usage{
setColors(object, ..., n) <- value
}
\arguments{
\item{x}{A \code{Raster*} object.}

\item{...}{Additional arguments to \code{colorRampPalette}.}

\item{value}{Named list of hex color codes (e.g., from \code{RColorBrewer::brewer.pal}),
             corresponding to the names of RasterLayers in \code{x}.}

\item{n}{An optional vector of values specifiying the number
             of levels from which to interpolate the color palette.}
}
\value{
Returns a Raster with the colortable slot set to values.
}
\description{
Set colours for plotting Raster* objects.

set colortable of a raster object

set colortable of a raster object

set colortable of a raster object

set colortable of a raster object
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{brewer.pal}}, \code{\link{colorRampPalette}}.
}

