% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{show,simList-method}
\alias{show,simList-method}
\title{Show an Object}
\usage{
\S4method{show}{simList}(object)
}
\arguments{
\item{object}{\code{simList}}
}
\description{
Show an Object
}
