% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,SpatialPointsDataFrameNamed-method}
\alias{show,SpatialPointsDataFrameNamed-method}
\title{show SpatialPointsDataFrameNamed}
\usage{
\S4method{show}{SpatialPointsDataFrameNamed}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
show SpatialPointsDataFrameNamed
}

