% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,SpatialPolygonsDataFrameNamed-method}
\alias{show,SpatialPolygonsDataFrameNamed-method}
\title{show SpatialPolygonsDataFrameNamed}
\usage{
\S4method{show}{SpatialPolygonsDataFrameNamed}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
show SpatialPolygonsDataFrameNamed
}

