% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-diagrams.R
\docType{methods}
\name{.sim2gantt}
\alias{.sim2gantt}
\alias{.sim2gantt,simList,numeric,character,numeric-method}
\title{sim2gantt}
\usage{
.sim2gantt(sim, n, startDate, width)

\S4method{.sim2gantt}{simList,numeric,character,numeric}(sim, n, startDate,
  width)
}
\arguments{
\item{sim}{A \code{simList} object (typically corresponding to a
completed simulation).}

\item{n}{The number of most recently completed events to plot.}

\item{startDate}{A character representation of date in \code{YYYY-MM-DD} format.}

\item{width}{Numeric. Passed to determine scale of vertical bars.}
}
\value{
A list of data.frames
}
\description{
Internal function to convert the completed events list of a \code{simList}
object to a list of \code{data.frame}s suitable to pass to a call to
\code{\link[DiagrammeR]{mermaid}} to make a Gannt chart representing the
events in a completed simulation.
}
\author{
Alex Chubaty
}
\keyword{internal}
