% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-dependencies-class.R
\docType{class}
\name{.simDeps-class}
\alias{.simDeps-class}
\alias{.simDeps}
\title{The \code{.simDeps} class}
\description{
Defines all simulation dependencies for all modules within a SpaDES simulation.
}
\section{Slots}{

\describe{
\item{\code{dependencies}}{List of \code{\link{.moduleDeps}} dependency objects.}
}}

\seealso{
\code{\link{.moduleDeps}}, \code{\link{spadesClasses}}
}
\author{
Alex Chubaty
}
\keyword{internal}
