% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-dependencies-class.R
\docType{class}
\name{.simDeps-class}
\alias{.simDeps}
\alias{.simDeps-class}
\title{The \code{.simDeps} class}
\description{
Defines all simulation dependencies for all modules within a SpaDES simulation.
}
\section{Slots}{

\describe{
\item{\code{dependencies}}{List of \code{\link{.moduleDeps}} dependency objects.}

\item{\code{.allObjNames}}{A named list in which to store the names of all input/output
objects, for fast lookup during the simulation.}
}}
\author{
Alex Chubaty
}
\seealso{
\code{\link{.moduleDeps}}, \code{\link{spadesClasses}}
}
\keyword{internal}

