% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{simInit}
\alias{simInit}
\alias{simInit,list,list,list,character,character-method}
\alias{simInit,list,list,list,character,missing-method}
\alias{simInit,list,list,list,missing,character-method}
\alias{simInit,list,list,list,missing,missing-method}
\title{Initialize a new simulation}
\usage{
simInit(times, params, modules, path, loadOrder)

\S4method{simInit}{list,list,list,character,character}(times, params, modules,
  path, loadOrder)

\S4method{simInit}{list,list,list,missing,character}(times, params, modules,
  loadOrder)

\S4method{simInit}{list,list,list,character,missing}(times, params, modules,
  path)

\S4method{simInit}{list,list,list,missing,missing}(times, params, modules)
}
\arguments{
\item{times}{A named list of numeric simulation start and stop times
       (e.g., \code{times=list(start=0.0, stop=10.0)}).}

\item{params}{A named list of simulation parameters and their values.}

\item{modules}{A named list of character strings specfying the names
of modules to be loaded for the simulation. **Note:** the module name
should correspond to the R source file from which the module is loaded.
Example: a module named "caribou" will be sourced form the file \code{caribou.R},
located at the specified \code{path} (see below).}

\item{path}{An optional character string specifying the location of the module source files.
             If no path is specified, it defaults to the current working directory.}

\item{.loadOrder}{An optional list of module names specfiying the order in
                   which to load the modules. If not specified, the module
                   load order will be determined automatically.}
}
\value{
A \code{simList} simulation object, pre-initialized from values specified
in the arguments supplied.
}
\description{
Create a new simulation object, preloaded with parameters, modules, times, etc.
}
\details{
We implement a discrete event simulation in a more modular fashion so it's
easier to add submodules to the simulation. We use S4 classes and methods,
and use \code{data.table} instead of \code{data.frame} to implement the event
queue (because it is much faster).
}
\examples{
\dontrun{
 mySim <- simInit(times=list(start=0.0, stop=10.0), params=list(Ncaribou=100),
 modules=list("habitat", "caribou"), path="/path/to/my/modules/")
 mySim
}
}
\author{
Alex Chubaty
}
\references{
Matloff, N. (2011). The Art of R Programming (ch. 7.8.3). San Fransisco, CA: No Starch Press, Inc.. Retrieved from \url{http://www.nostarch.com/artofr.htm}
}
\seealso{
\code{\link{spades}}.
}

