% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{simInit}
\alias{simInit}
\alias{simInit,list,list,list,list,character,ANY,ANY,character-method}
\alias{simInit,list,list,list,list,character,ANY,ANY,missing-method}
\alias{simInit,list,list,list,list,character,ANY,missing,missing-method}
\alias{simInit,list,list,list,list,character,missing,ANY,missing-method}
\alias{simInit,list,list,list,list,character,missing,missing,missing-method}
\alias{simInit,list,list,list,list,missing,ANY,ANY,character-method}
\alias{simInit,list,list,list,list,missing,ANY,ANY,missing-method}
\alias{simInit,list,list,list,list,missing,ANY,missing,character-method}
\alias{simInit,list,list,list,list,missing,ANY,missing,missing-method}
\alias{simInit,list,list,list,list,missing,missing,ANY,character-method}
\alias{simInit,list,list,list,list,missing,missing,ANY,missing-method}
\alias{simInit,list,list,list,list,missing,missing,missing,character-method}
\alias{simInit,list,list,list,list,missing,missing,missing,missing-method}
\alias{simInit,list,list,list,missing,character,ANY,ANY,character-method}
\alias{simInit,list,list,list,missing,character,ANY,ANY,missing-method}
\alias{simInit,list,list,list,missing,character,ANY,missing,character-method}
\alias{simInit,list,list,list,missing,character,ANY,missing,missing-method}
\alias{simInit,list,list,list,missing,character,missing,ANY,character-method}
\alias{simInit,list,list,list,missing,character,missing,ANY,missing-method}
\alias{simInit,list,list,list,missing,character,missing,missing,character-method}
\alias{simInit,list,list,list,missing,character,missing,missing,missing-method}
\alias{simInit,list,list,list,missing,missing,ANY,ANY,character-method}
\alias{simInit,list,list,list,missing,missing,ANY,ANY,missing-method}
\alias{simInit,list,list,list,missing,missing,ANY,missing,character-method}
\alias{simInit,list,list,list,missing,missing,ANY,missing,missing-method}
\alias{simInit,list,list,list,missing,missing,missing,ANY,character-method}
\alias{simInit,list,list,list,missing,missing,missing,ANY,missing-method}
\alias{simInit,list,list,list,missing,missing,missing,missing,character-method}
\alias{simInit,list,list,list,missing,missing,missing,missing,missing-method}
\alias{simInit,missing,missing,missing,missing,missing,missing,missing,missing-method}
\title{Initialize a new simulation}
\usage{
simInit(times, params, modules, objects, path, inputs, outputs, loadOrder)

\S4method{simInit}{list,list,list,list,character,ANY,ANY,character}(times,
  params, modules, objects, path, inputs, outputs, loadOrder)

\S4method{simInit}{list,list,list,list,character,ANY,missing,missing}(times,
  params, modules, objects, path, inputs)

\S4method{simInit}{list,list,list,list,character,ANY,missing,missing}(times,
  params, modules, objects, path, inputs)

\S4method{simInit}{list,list,list,list,missing,ANY,missing,character}(times,
  params, modules, objects, inputs, loadOrder)


  \S4method{simInit}{list,list,list,missing,character,ANY,missing,character}(times,
  params, modules, path, inputs, loadOrder)

\S4method{simInit}{list,list,list,list,missing,ANY,missing,missing}(times,
  params, modules, objects, inputs)

\S4method{simInit}{list,list,list,missing,missing,ANY,missing,character}(times,
  params, modules, inputs, loadOrder)

\S4method{simInit}{list,list,list,missing,character,ANY,missing,missing}(times,
  params, modules, path, inputs)

\S4method{simInit}{list,list,list,missing,missing,ANY,missing,missing}(times,
  params, modules, inputs)


  \S4method{simInit}{list,list,list,list,character,missing,missing,missing}(times,
  params, modules, objects, path, inputs)


  \S4method{simInit}{list,list,list,list,character,missing,missing,missing}(times,
  params, modules, objects, path, inputs)


  \S4method{simInit}{list,list,list,list,missing,missing,missing,character}(times,
  params, modules, objects, inputs, loadOrder)


  \S4method{simInit}{list,list,list,missing,character,missing,missing,character}(times,
  params, modules, path, inputs, loadOrder)

\S4method{simInit}{list,list,list,list,missing,missing,missing,missing}(times,
  params, modules, objects, inputs)


  \S4method{simInit}{list,list,list,missing,missing,missing,missing,character}(times,
  params, modules, inputs, loadOrder)


  \S4method{simInit}{list,list,list,missing,character,missing,missing,missing}(times,
  params, modules, path, inputs)


  \S4method{simInit}{list,list,list,missing,missing,missing,missing,missing}(times,
  params, modules, inputs)

\S4method{simInit}{list,list,list,list,character,missing,ANY,missing}(times,
  params, modules, objects, path, outputs)

\S4method{simInit}{list,list,list,list,character,missing,ANY,missing}(times,
  params, modules, objects, path, outputs)

\S4method{simInit}{list,list,list,list,missing,missing,ANY,character}(times,
  params, modules, objects, outputs, loadOrder)


  \S4method{simInit}{list,list,list,missing,character,missing,ANY,character}(times,
  params, modules, path, outputs, loadOrder)

\S4method{simInit}{list,list,list,list,missing,missing,ANY,missing}(times,
  params, modules, objects, outputs)

\S4method{simInit}{list,list,list,missing,missing,missing,ANY,character}(times,
  params, modules, outputs, loadOrder)

\S4method{simInit}{list,list,list,missing,character,missing,ANY,missing}(times,
  params, modules, path, outputs)

\S4method{simInit}{list,list,list,missing,missing,missing,ANY,missing}(times,
  params, modules, outputs)

\S4method{simInit}{list,list,list,list,character,ANY,ANY,missing}(times, params,
  modules, objects, path, inputs, outputs)

\S4method{simInit}{list,list,list,list,character,ANY,ANY,missing}(times, params,
  modules, objects, path, inputs, outputs)

\S4method{simInit}{list,list,list,list,missing,ANY,ANY,character}(times, params,
  modules, objects, inputs, outputs, loadOrder)

\S4method{simInit}{list,list,list,missing,character,ANY,ANY,character}(times,
  params, modules, path, inputs, outputs, loadOrder)

\S4method{simInit}{list,list,list,list,missing,ANY,ANY,missing}(times, params,
  modules, objects, inputs, outputs)

\S4method{simInit}{list,list,list,missing,missing,ANY,ANY,character}(times,
  params, modules, inputs, outputs, loadOrder)

\S4method{simInit}{list,list,list,missing,character,ANY,ANY,missing}(times,
  params, modules, path, inputs, outputs)

\S4method{simInit}{list,list,list,missing,missing,ANY,ANY,missing}(times,
  params, modules, inputs, outputs)


  \S4method{simInit}{missing,missing,missing,missing,missing,missing,missing,missing}(inputs,
  outputs)
}
\arguments{
\item{times}{A named list of numeric simulation start and stop times
       (e.g., \code{times=list(start=0.0, stop=10.0)}).}

\item{params}{A named list of simulation parameters and their values.}

\item{modules}{A named list of character strings specfying the names
of modules to be loaded for the simulation. Note: the module name
should correspond to the R source file from which the module is loaded.
Example: a module named "caribou" will be sourced form the file
\code{caribou.R}, located at the specified \code{path} (see below).}

\item{objects}{A list of data objects to be used in the simulation.}

\item{path}{An optional character string specifying the location of the module source files.
             If no path is specified, it defaults to the current working directory.}

\item{inputs}{A list of length 2, named "path" and "filelist".}

\item{outputs}{A list of length 2, named "path" and "filelist".}

\item{loadOrder}{An optional list of module names specfiying the order in
                  which to load the modules. If not specified, the module
                  load order will be determined automatically.}
}
\value{
A \code{simList} simulation object, pre-initialized from values
specified in the arguments supplied.
}
\description{
Create a new simulation object, preloaded with parameters, modules, times, etc.
}
\details{
We implement a discrete event simulation in a more modular fashion so it's
easier to add submodules to the simulation. We use S4 classes and methods,
and use \code{data.table} instead of \code{data.frame} to implement the event
queue (because it is much faster).
}
\examples{
\dontrun{
 mySim <- simInit(times=list(start=0.0, stop=10.0), params=list(Ncaribou=100),
 modules=list("habitat", "caribou"), path="/path/to/my/modules/")
 mySim
}
}
\author{
Alex Chubaty and Eliot McIntire
}
\references{
Matloff, N. (2011). The Art of R Programming (ch. 7.8.3). San Fransisco, CA: No Starch Press, Inc.. Retrieved from \url{http://www.nostarch.com/artofr.htm}
}
\seealso{
\code{\link{spades}}.
}

