% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{envir}
\alias{envir}
\alias{envir,simList-method}
\alias{envir<-}
\alias{envir<-,simList-method}
\alias{simList-accessors-envir}
\title{Simulation environment}
\usage{
envir(object)

\S4method{envir}{simList}(object)

envir(object) <- value

\S4method{envir}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Accessor functions for the \code{.envir} slot in a \code{simList} object.
These are included for advanced users.
}
\details{
Currently, only get and set methods are defined. Subset methods are not.
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{simList-class}},
         \code{\link{simList-accessors-events}},
         \code{\link{simList-accessors-inout}},
         \code{\link{simList-accessors-modules}},
         \code{\link{simList-accessors-objects}},
         \code{\link{simList-accessors-params}},
         \code{\link{simList-accessors-paths}},
         \code{\link{simList-accessors-times}}.
}

