% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{simEvents}
\alias{simCompleted}
\alias{simCompleted,simList-method}
\alias{simCompleted<-}
\alias{simCompleted<-,simList-method}
\alias{simEvents}
\alias{simEvents,simList-method}
\alias{simEvents<-}
\alias{simEvents<-,simList-method}
\alias{simList-accessors-events}
\title{Simulation event lists}
\usage{
simEvents(object)

\S4method{simEvents}{simList}(object)

simEvents(object) <- value

\S4method{simEvents}{simList}(object) <- value

simCompleted(object)

\S4method{simCompleted}{simList}(object)

simCompleted(object) <- value

\S4method{simCompleted}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Accessor functions for the \code{events} and \code{completed} slots of a
\code{simList} object.
By default, the event lists are shown when the \code{simList} object is printed,
thus most users will not require direct use of these methods.
\tabular{ll}{
   \code{simEvents} \tab Scheduled simulation events (the event queue).\cr
   \code{simCompleted} \tab Completed simulation events.\cr
}
}
\details{
Currently, only get and set methods are defined. Subset methods are not.
}
\note{
Each event is represented by a \code{\link{data.table}} row consisting of:
       \tabular{ll}{
         \code{eventTime} \tab The time the event is to occur.\cr
         \code{moduleName} \tab The module from which the event is taken.\cr
         \code{eventType} \tab A character string for the programmer-defined event type.\cr
       }
}
\seealso{
\code{\link{simList-class}},
         \code{\link{simList-accessors-params}},
         \code{\link{simList-accessors-modules}},
         \code{\link{simList-accessors-times}}.
}

