% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{simObjects}
\alias{simDepends}
\alias{simDepends,simList-method}
\alias{simDepends<-}
\alias{simDepends<-,simList-method}
\alias{simList-accessors-modules}
\alias{simModules}
\alias{simModules,simList-method}
\alias{simModules<-}
\alias{simModules<-,simList-method}
\alias{simModulesLoadOrder}
\alias{simModulesLoadOrder,simList-method}
\alias{simModulesLoadOrder<-}
\alias{simModulesLoadOrder<-,simList-method}
\alias{simModulesLoaded}
\alias{simModulesLoaded,simList-method}
\alias{simModulesLoaded<-}
\alias{simModulesLoaded<-,simList-method}
\alias{simObjects}
\alias{simObjectsLoaded}
\alias{simObjectsLoaded,simList-method}
\alias{simObjectsLoaded<-}
\alias{simObjectsLoaded<-,simList-method}
\title{Simulation objects}
\usage{
simObjects(object, ...)

simModules(object)

\S4method{simModules}{simList}(object)

simModules(object) <- value

\S4method{simModules}{simList}(object) <- value

simModulesLoaded(object)

\S4method{simModulesLoaded}{simList}(object)

simModulesLoaded(object) <- value

\S4method{simModulesLoaded}{simList}(object) <- value

simModulesLoadOrder(object)

\S4method{simModulesLoadOrder}{simList}(object)

simModulesLoadOrder(object) <- value

\S4method{simModulesLoadOrder}{simList}(object) <- value

simDepends(object)

\S4method{simDepends}{simList}(object)

simDepends(object) <- value

\S4method{simDepends}{simList}(object) <- value

simObjectsLoaded(object)

\S4method{simObjectsLoaded}{simList}(object)

simObjectsLoaded(object) <- value

\S4method{simObjectsLoaded}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{...}{Additional arguments passed to \code{ls.str}.}

\item{value}{The object to be stored at the slot.}

\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.

Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Accessor functions for the \code{objects} slot in a \code{simList} object.
These are included for advanced users.

Accessor functions for the \code{depends}, \code{modules}, \code{.loaded},
and \code{.loadOrder} slots in a \code{simList} object.
These are included for advanced users.
\tabular{ll}{
   \code{\link{simDepends}} \tab List of simulation module dependencies. (advanced) \cr
   \code{\link{simModules}} \tab List of simulation modules to be loaded. (advanced) \cr
   \code{\link{simModulesLoaded}} \tab List of loaded simulation modules. (advanced) \cr
   \code{\link{simModulesLoadOrder}} \tab List specifying the order in which to load modules. (advanced) \cr
   \code{\link{simObjectsLoaded}} \tab List of loaded objects used in simulation. (advanced) \cr
}
}
\details{
Currently, only get and set methods are defined. Subset methods are not.

Currently, only get and set methods are defined. Subset methods are not.
}
\author{
Alex Chubaty

Alex Chubaty
}
\seealso{
\code{\link{simList-class}},
         \code{\link{simList-accessors-modules}},
         \code{\link{simList-accessors-params}},
         \code{\link{simList-accessors-events}},
         \code{\link{simList-accessors-times}}.

\code{\link{simList-class}},
         \code{\link{simList-accessors-params}},
         \code{\link{simList-accessors-events}},
         \code{\link{simList-accessors-times}}.
}

