% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{simParams}
\alias{simCheckpointFile}
\alias{simCheckpointFile,simList-method}
\alias{simCheckpointFile<-}
\alias{simCheckpointFile<-,simList-method}
\alias{simCheckpointInterval}
\alias{simCheckpointInterval,simList-method}
\alias{simCheckpointInterval<-}
\alias{simCheckpointInterval<-,simList-method}
\alias{simFileList}
\alias{simFileList,simList-method}
\alias{simFileList<-}
\alias{simFileList<-,simList-method}
\alias{simGlobals}
\alias{simGlobals,simList-method}
\alias{simGlobals<-}
\alias{simGlobals<-,simList-method}
\alias{simGlobalsOutputPath}
\alias{simGlobalsOutputPath,simList-method}
\alias{simGlobalsOutputPath<-}
\alias{simGlobalsOutputPath<-,simList-method}
\alias{simList-accessors-params}
\alias{simParams}
\alias{simParams,simList-method}
\alias{simParams<-}
\alias{simParams<-,simList-method}
\alias{simProgressGraphical}
\alias{simProgressGraphical,simList-method}
\alias{simProgressGraphical<-}
\alias{simProgressGraphical<-,simList-method}
\alias{simProgressInterval}
\alias{simProgressInterval,simList-method}
\alias{simProgressInterval<-}
\alias{simProgressInterval<-,simList-method}
\title{Get and set simulation parameters.}
\usage{
simParams(object)

\S4method{simParams}{simList}(object)

simParams(object) <- value

\S4method{simParams}{simList}(object) <- value

simCheckpointFile(object)

\S4method{simCheckpointFile}{simList}(object)

simCheckpointFile(object) <- value

\S4method{simCheckpointFile}{simList}(object) <- value

simCheckpointInterval(object)

\S4method{simCheckpointInterval}{simList}(object)

simCheckpointInterval(object) <- value

\S4method{simCheckpointInterval}{simList}(object) <- value

simFileList(object)

\S4method{simFileList}{simList}(object)

simFileList(object) <- value

\S4method{simFileList}{simList}(object) <- value

simProgressGraphical(object)

\S4method{simProgressGraphical}{simList}(object)

simProgressGraphical(object) <- value

\S4method{simProgressGraphical}{simList}(object) <- value

simProgressInterval(object)

\S4method{simProgressInterval}{simList}(object)

simProgressInterval(object) <- value

\S4method{simProgressInterval}{simList}(object) <- value

simGlobals(object)

\S4method{simGlobals}{simList}(object)

simGlobals(object) <- value

\S4method{simGlobals}{simList}(object) <- value

simGlobalsOutputPath(object)

\S4method{simGlobalsOutputPath}{simList}(object)

simGlobalsOutputPath(object) <- value

\S4method{simGlobalsOutputPath}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Accessor functions for the \code{params} slot of a \code{simList} object
and its elements.
Additonal methods are provided to access core module and global parameters:
\tabular{ll}{
   Global parameters \tab \cr
   \code{simGlobals} \tab List of global simulation parameters.\cr
   \code{simGlobalsOutputPath} \tab List of global simulation output path.\cr
}
\tabular{ll}{
   User module parameters \tab \cr
   \code{simParams} \tab List of simulation parameters.\cr
}
\tabular{ll}{
   Core module parameters \tab \cr
   \code{simCheckpointFile} \tab Name of the checkpoint file. (advanced)\cr
   \code{simCheckpointInterval} \tab The simulation checkpoint interval. (advanced)\cr
   \code{simObjectsLoaded} \tab List of loaded simulation objects. (advanced)\cr
   \code{simProgressGraphical} \tab Type of graphical progress bar used. (advanced)\cr
   \code{simProgressInterval} \tab Interval for the progress bar. (advanced)\cr
}
}
\details{
Currently, only get and set methods are defined. Subset methods are not.
}
\seealso{
\code{\link{simList-class}},
         \code{\link{simList-accessors-modules}},
         \code{\link{simList-accessors-events}},
         \code{\link{simList-accessors-times}}.
}

