% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{params}
\alias{globals}
\alias{globals,simList-method}
\alias{globals<-}
\alias{globals<-,simList-method}
\alias{outputPath}
\alias{outputPath,simList-method}
\alias{outputPath<-}
\alias{outputPath<-,simList-method}
\alias{params}
\alias{params,simList-method}
\alias{params<-}
\alias{params<-,simList-method}
\alias{simCheckpointFile}
\alias{simCheckpointFile,simList-method}
\alias{simCheckpointFile<-}
\alias{simCheckpointFile<-,simList-method}
\alias{simCheckpointInterval}
\alias{simCheckpointInterval,simList-method}
\alias{simCheckpointInterval<-}
\alias{simCheckpointInterval<-,simList-method}
\alias{simFileList}
\alias{simFileList,simList-method}
\alias{simFileList<-}
\alias{simFileList<-,simList-method}
\alias{simList-accessors-params}
\alias{simProgressGraphical}
\alias{simProgressGraphical,simList-method}
\alias{simProgressGraphical<-}
\alias{simProgressGraphical<-,simList-method}
\alias{simProgressInterval}
\alias{simProgressInterval,simList-method}
\alias{simProgressInterval<-}
\alias{simProgressInterval<-,simList-method}
\title{Get and set simulation parameters.}
\usage{
params(object)

\S4method{params}{simList}(object)

params(object) <- value

\S4method{params}{simList}(object) <- value

simCheckpointFile(object)

\S4method{simCheckpointFile}{simList}(object)

simCheckpointFile(object) <- value

\S4method{simCheckpointFile}{simList}(object) <- value

simCheckpointInterval(object)

\S4method{simCheckpointInterval}{simList}(object)

simCheckpointInterval(object) <- value

\S4method{simCheckpointInterval}{simList}(object) <- value

simFileList(object)

\S4method{simFileList}{simList}(object)

simFileList(object) <- value

\S4method{simFileList}{simList}(object) <- value

simProgressGraphical(object)

\S4method{simProgressGraphical}{simList}(object)

simProgressGraphical(object) <- value

\S4method{simProgressGraphical}{simList}(object) <- value

simProgressInterval(object)

\S4method{simProgressInterval}{simList}(object)

simProgressInterval(object) <- value

\S4method{simProgressInterval}{simList}(object) <- value

globals(object)

\S4method{globals}{simList}(object)

globals(object) <- value

\S4method{globals}{simList}(object) <- value

outputPath(object)

\S4method{outputPath}{simList}(object)

outputPath(object) <- value

\S4method{outputPath}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Accessor functions for the \code{params} slot of a \code{simList} object
and its elements.
Additonal methods are provided to access core module and global parameters:
Commonly used
\tabular{ll}{
   \code{globals} \tab List of global simulation parameters.\cr
   \code{params} \tab Nested list of all simulation parameter.\cr
}
Advanced use
\tabular{lll}{
   Accessor method \tab Module \tab Description \cr
   \code{simCheckpointFile} \tab \code{.checkpoint} \tab Name of the checkpoint file. (advanced)\cr
   \code{simCheckpointInterval} \tab \code{.checkpoint} \tab The simulation checkpoint interval. (advanced)\cr
   \code{outputPath} \tab \code{NA} \tab Global simulation output path. (advanced)\cr
   \code{simFileList} \tab \code{.load} \tab List of files to load for the simulation. (advanced)\cr
   \code{simObjectsLoaded} \tab \code{.load} \tab List of loaded simulation objects. (advanced)\cr
   \code{simProgressGraphical} \tab \code{.progress} \tab Type of graphical progress bar used. (advanced)\cr
   \code{simProgressInterval} \tab \code{.progress} \tab Interval for the progress bar. (advanced)\cr
}
}
\details{
Currently, only get and set methods are defined. Subset methods are not.
}
\seealso{
\code{\link{simList-class}},
         \code{\link{simList-accessors-modules}},
         \code{\link{simList-accessors-events}},
         \code{\link{simList-accessors-times}}.
}

