% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{params}
\alias{checkpointFile}
\alias{checkpointFile,simList-method}
\alias{checkpointFile<-}
\alias{checkpointFile<-,simList-method}
\alias{checkpointInterval}
\alias{checkpointInterval,simList-method}
\alias{checkpointInterval<-}
\alias{checkpointInterval<-,simList-method}
\alias{globals}
\alias{globals,simList-method}
\alias{globals<-}
\alias{globals<-,simList-method}
\alias{inputArgs}
\alias{inputArgs,simList-method}
\alias{inputArgs<-}
\alias{inputArgs<-,simList-method}
\alias{inputPath}
\alias{inputPath,simList-method}
\alias{inputPath<-}
\alias{inputPath<-,simList-method}
\alias{inputs}
\alias{inputs,simList-method}
\alias{inputs<-}
\alias{inputs<-,simList-method}
\alias{modulePath}
\alias{modulePath,simList-method}
\alias{modulePath<-}
\alias{modulePath<-,simList-method}
\alias{outputArgs}
\alias{outputArgs,simList-method}
\alias{outputArgs<-}
\alias{outputArgs<-,simList-method}
\alias{outputPath}
\alias{outputPath,simList-method}
\alias{outputPath<-}
\alias{outputPath<-,simList-method}
\alias{outputs}
\alias{outputs,simList-method}
\alias{outputs<-}
\alias{outputs<-,simList-method}
\alias{params}
\alias{params,simList-method}
\alias{params<-}
\alias{params<-,simList-method}
\alias{paths}
\alias{paths,simList-method}
\alias{paths<-}
\alias{paths<-,simList-method}
\alias{progressInterval}
\alias{progressInterval,simList-method}
\alias{progressInterval<-}
\alias{progressInterval<-,simList-method}
\alias{progressType}
\alias{progressType,simList-method}
\alias{progressType<-}
\alias{progressType<-,simList-method}
\alias{simList-accessors-params}
\title{Get and set simulation parameters.}
\usage{
params(object)

\S4method{params}{simList}(object)

params(object) <- value

\S4method{params}{simList}(object) <- value

checkpointFile(object)

\S4method{checkpointFile}{simList}(object)

checkpointFile(object) <- value

\S4method{checkpointFile}{simList}(object) <- value

checkpointInterval(object)

\S4method{checkpointInterval}{simList}(object)

checkpointInterval(object) <- value

\S4method{checkpointInterval}{simList}(object) <- value

inputs(object)

\S4method{inputs}{simList}(object)

inputs(object) <- value

\S4method{inputs}{simList}(object) <- value

outputs(object)

\S4method{outputs}{simList}(object)

outputs(object) <- value

\S4method{outputs}{simList}(object) <- value

inputPath(object)

\S4method{inputPath}{simList}(object)

inputPath(object) <- value

\S4method{inputPath}{simList}(object) <- value

modulePath(object)

\S4method{modulePath}{simList}(object)

modulePath(object) <- value

\S4method{modulePath}{simList}(object) <- value

paths(object)

\S4method{paths}{simList}(object)

paths(object) <- value

\S4method{paths}{simList}(object) <- value

inputArgs(object)

\S4method{inputArgs}{simList}(object)

inputArgs(object) <- value

\S4method{inputArgs}{simList}(object) <- value

outputArgs(object)

\S4method{outputArgs}{simList}(object)

outputArgs(object) <- value

\S4method{outputArgs}{simList}(object) <- value

progressType(object)

\S4method{progressType}{simList}(object)

progressType(object) <- value

\S4method{progressType}{simList}(object) <- value

progressInterval(object)

\S4method{progressInterval}{simList}(object)

progressInterval(object) <- value

\S4method{progressInterval}{simList}(object) <- value

globals(object)

\S4method{globals}{simList}(object)

globals(object) <- value

\S4method{globals}{simList}(object) <- value

outputPath(object)

\S4method{outputPath}{simList}(object)

outputPath(object) <- value

\S4method{outputPath}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Accessor functions for the \code{params} slot of a \code{simList} object
and its elements.
Additonal methods are provided to access core module and global parameters:
Commonly used
\tabular{ll}{
   \code{globals} \tab List of global simulation parameters.\cr
   \code{params} \tab Nested list of all simulation parameter.\cr
}
Advanced use
\tabular{lll}{
   Accessor method \tab Module \tab Description \cr
   \code{checkpointFile} \tab \code{.checkpoint} \tab Name of the checkpoint file. (advanced)\cr
   \code{checkpointInterval} \tab \code{.checkpoint} \tab The simulation checkpoint interval. (advanced)\cr
   \code{modulePath} \tab \code{NA} \tab Global simulation module path. (advanced)\cr
   \code{outputPath} \tab \code{NA} \tab Global simulation output path. (advanced)\cr
   \code{inputPath} \tab \code{NA} \tab Global simulation input path. (advanced)\cr
   \code{paths} \tab \code{NA} \tab Global simulation paths (modules, inputs, outputs). (advanced)\cr
   \code{inputs} \tab \code{inputs} \tab Data.table showing objects, files etc.. (advanced)\cr
   \code{progressType} \tab \code{.progress} \tab Type of graphical progress bar used. (advanced)\cr
   \code{progressInterval} \tab \code{.progress} \tab Interval for the progress bar. (advanced)\cr
}
}
\details{
Currently, only get and set methods are defined. Subset methods are not.

\code{outputs} accepts a data.frame or data.table, with 6 columns. Currently, only one is required.
Columns are \code{file}, \code{fun},
}
\seealso{
\code{\link{simList-class}},
         \code{\link{simList-accessors-modules}},
         \code{\link{simList-accessors-events}},
         \code{\link{simList-accessors-times}}.
}

