% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{simTimes}
\alias{simCurrentTime}
\alias{simCurrentTime,simList-method}
\alias{simCurrentTime<-}
\alias{simCurrentTime<-,simList-method}
\alias{simList-accessors-times}
\alias{simStartTime}
\alias{simStartTime,simList-method}
\alias{simStartTime<-}
\alias{simStartTime<-,simList-method}
\alias{simStopTime}
\alias{simStopTime,simList-method}
\alias{simStopTime<-}
\alias{simStopTime<-,simList-method}
\alias{simTimes}
\alias{simTimes,simList-method}
\alias{simTimes<-}
\alias{simTimes<-,simList-method}
\alias{simTimestepUnit}
\alias{simTimestepUnit,simList-method}
\alias{simTimestepUnit<-}
\alias{simTimestepUnit<-,simList-method}
\title{Get and set simulation times.}
\usage{
simTimes(object)

\S4method{simTimes}{simList}(object)

simTimes(object) <- value

\S4method{simTimes}{simList}(object) <- value

simCurrentTime(object)

\S4method{simCurrentTime}{simList}(object)

simCurrentTime(object) <- value

\S4method{simCurrentTime}{simList}(object) <- value

simStartTime(object)

\S4method{simStartTime}{simList}(object)

simStartTime(object) <- value

\S4method{simStartTime}{simList}(object) <- value

simStopTime(object)

\S4method{simStopTime}{simList}(object)

simStopTime(object) <- value

\S4method{simStopTime}{simList}(object) <- value

simTimestepUnit(object)

\S4method{simTimestepUnit}{simList}(object)

simTimestepUnit(object) <- value

\S4method{simTimestepUnit}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Accessor functions for the \code{simtimes} slot of a \code{simList} object
and its elements.
Additonal methods are provided to access the current, start, and stop times
of the simulation.
\tabular{ll}{
   \code{simCurrentTime} \tab Current simulation time.\cr
   \code{simStartTime} \tab Simulation start time.\cr
   \code{simStopTime} \tab Simulation stop time.\cr
   \code{simTimes} \tab List of all simulation times (current, start, stop).\cr
}
}
\details{
Currently, only get and set methods are defined. Subset methods are not.

\code{simTimestepUnit} will extract the current units of the time used in a spades call. If
it is set within a \code{simInit} as say, \code{times=list(start=0, stop=52, timestep="week")}
it will set the units for that simulation. But default, a simInit call will use the largest
units contained within the meta data for the modules being used. NA for timestep defaults to
"year"
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{simList-class}},
         \code{\link{simList-accessors-envir}},
         \code{\link{simList-accessors-modules}},
         \code{\link{simList-accessors-params}},
         \code{\link{simList-accessors-events}}.
}

