% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R, R/times.R
\docType{methods}
\name{simTimes}
\alias{.convUnits,numeric,character-method}
\alias{moduleTimestepUnit}
\alias{moduleTimestepUnit,NULL-method}
\alias{moduleTimestepUnit,simList-method}
\alias{simCurrentTime}
\alias{simCurrentTime,simList,character-method}
\alias{simCurrentTime,simList,missing-method}
\alias{simCurrentTime<-}
\alias{simCurrentTime<-,simList-method}
\alias{simList-accessors-times}
\alias{simModuleTimestepUnits}
\alias{simModuleTimestepUnits,simList-method}
\alias{simStartTime}
\alias{simStartTime,simList,character-method}
\alias{simStartTime,simList,missing-method}
\alias{simStartTime<-}
\alias{simStartTime<-,simList-method}
\alias{simStopTime}
\alias{simStopTime,simList,character-method}
\alias{simStopTime,simList,missing-method}
\alias{simStopTime<-}
\alias{simStopTime<-,simList-method}
\alias{simTimes}
\alias{simTimes,simList-method}
\alias{simTimes<-}
\alias{simTimes<-,simList-method}
\alias{simTimestepUnit}
\alias{simTimestepUnit,simList-method}
\alias{simTimestepUnit<-}
\alias{simTimestepUnit<-,simList-method}
\title{Get and set simulation times.}
\usage{
simTimes(object)

\S4method{simTimes}{simList}(object)

simTimes(object) <- value

\S4method{simTimes}{simList}(object) <- value

simCurrentTime(object, unit)

\S4method{simCurrentTime}{simList,missing}(object)

\S4method{simCurrentTime}{simList,character}(object, unit)

simCurrentTime(object) <- value

\S4method{simCurrentTime}{simList}(object) <- value

simStopTime(object, unit)

\S4method{simStopTime}{simList,missing}(object)

\S4method{simStopTime}{simList,character}(object, unit)

simStopTime(object) <- value

\S4method{simStopTime}{simList}(object) <- value

simStartTime(object, unit)

\S4method{simStartTime}{simList,missing}(object)

\S4method{simStartTime}{simList,character}(object, unit)

simStartTime(object) <- value

\S4method{simStartTime}{simList}(object) <- value

moduleTimestepUnit(object)

\S4method{moduleTimestepUnit}{simList}(object)

\S4method{moduleTimestepUnit}{`NULL`}(object)

simTimestepUnit(object)

\S4method{simTimestepUnit}{simList}(object)

simTimestepUnit(object) <- value

\S4method{simTimestepUnit}{simList}(object) <- value

simModuleTimestepUnits(object)

\S4method{simModuleTimestepUnits}{simList}(object)

\S4method{.convUnits}{numeric,character}(time, unit)
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}

\item{unit}{character. One of the time units used in SpaDES.}

\item{unit}{character. One of the time units used in SpaDES.}

\item{unit}{character. One of the time units used in SpaDES.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Accessor functions for the \code{simtimes} slot of a \code{simList} object
and its elements.
Additonal methods are provided to access the current, start, and stop times
of the simulation.
\tabular{ll}{
   \code{simCurrentTime} \tab Current simulation time.\cr
   \code{simStartTime} \tab Simulation start time.\cr
   \code{simStopTime} \tab Simulation stop time.\cr
   \code{simTimes} \tab List of all simulation times (current, start, stop).\cr
}
}
\details{
Currently, only get and set methods are defined. Subset methods are not.

\code{simTimestepUnit} will extract the current units of the time used in a spades call. If
it is set within a \code{simInit} as say, \code{times=list(start=0, stop=52, timestepUnit="week")}
it will set the units for that simulation. But default, a simInit call will use the largest
units contained within the meta data for the modules being used. NA for timestepUnit defaults to
"year"

\code{simModuleTimestepUnits} will extract the current units of the time of all
modules used in a simObject. This is different from \code{simTimestepUnits} because it
is not necessarily associated with a spades call
}
\author{
Alex Chubaty and Eliot McIntire

Eliot McIntire
}
\seealso{
\code{\link{simList-class}},
         \code{\link{simList-accessors-envir}},
         \code{\link{simList-accessors-modules}},
         \code{\link{simList-accessors-params}},
         \code{\link{simList-accessors-events}}.
}

