% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simList.R
\docType{methods}
\name{simModules}
\alias{simCheckpointFile}
\alias{simCheckpointFile,simList-method}
\alias{simCheckpointFile<-}
\alias{simCheckpointFile<-,simList-method}
\alias{simCheckpointInterval}
\alias{simCheckpointInterval,simList-method}
\alias{simCheckpointInterval<-}
\alias{simCheckpointInterval<-,simList-method}
\alias{simCompleted}
\alias{simCompleted,simList-method}
\alias{simCompleted<-}
\alias{simCompleted<-,simList-method}
\alias{simDepends}
\alias{simDepends,simList-method}
\alias{simDepends<-}
\alias{simDepends<-,simList-method}
\alias{simEvents}
\alias{simEvents,simList-method}
\alias{simEvents<-}
\alias{simEvents<-,simList-method}
\alias{simFileList}
\alias{simFileList,simList-method}
\alias{simFileList<-}
\alias{simFileList<-,simList-method}
\alias{simGlobals}
\alias{simGlobals,simList-method}
\alias{simGlobals<-}
\alias{simGlobals<-,simList-method}
\alias{simGlobalsOutputPath}
\alias{simGlobalsOutputPath,simList-method}
\alias{simGlobalsOutputPath<-}
\alias{simGlobalsOutputPath<-,simList-method}
\alias{simModules}
\alias{simModules,simList-method}
\alias{simModules<-}
\alias{simModules<-,simList-method}
\alias{simModulesLoadOrder}
\alias{simModulesLoadOrder,simList-method}
\alias{simModulesLoadOrder<-}
\alias{simModulesLoadOrder<-,simList-method}
\alias{simModulesLoaded}
\alias{simModulesLoaded,simList-method}
\alias{simModulesLoaded<-}
\alias{simModulesLoaded<-,simList-method}
\alias{simObjectsLoaded}
\alias{simObjectsLoaded,simList-method}
\alias{simObjectsLoaded<-}
\alias{simObjectsLoaded<-,simList-method}
\alias{simParams}
\alias{simParams,simList-method}
\alias{simParams<-}
\alias{simParams<-,simList-method}
\alias{simProgressGraphical}
\alias{simProgressGraphical,simList-method}
\alias{simProgressGraphical<-}
\alias{simProgressGraphical<-,simList-method}
\alias{simProgressInterval}
\alias{simProgressInterval,simList-method}
\alias{simProgressInterval<-}
\alias{simProgressInterval<-,simList-method}
\title{Get and set slots in a \code{simList} object}
\usage{
simModules(object)

\S4method{simModules}{simList}(object)

simModules(object) <- value

\S4method{simModules}{simList}(object) <- value

simModulesLoadOrder(object)

\S4method{simModulesLoadOrder}{simList}(object)

simModulesLoadOrder(object) <- value

\S4method{simModulesLoadOrder}{simList}(object) <- value

simModulesLoaded(object)

\S4method{simModulesLoaded}{simList}(object)

simModulesLoaded(object) <- value

\S4method{simModulesLoaded}{simList}(object) <- value

simObjectsLoaded(object)

\S4method{simObjectsLoaded}{simList}(object)

simObjectsLoaded(object) <- value

\S4method{simObjectsLoaded}{simList}(object) <- value

simParams(object)

\S4method{simParams}{simList}(object)

simParams(object) <- value

\S4method{simParams}{simList}(object) <- value

simCheckpointFile(object)

\S4method{simCheckpointFile}{simList}(object)

simCheckpointFile(object) <- value

\S4method{simCheckpointFile}{simList}(object) <- value

simCheckpointInterval(object)

\S4method{simCheckpointInterval}{simList}(object)

simCheckpointInterval(object) <- value

\S4method{simCheckpointInterval}{simList}(object) <- value

simFileList(object)

\S4method{simFileList}{simList}(object)

simFileList(object) <- value

\S4method{simFileList}{simList}(object) <- value

simProgressGraphical(object)

\S4method{simProgressGraphical}{simList}(object)

simProgressGraphical(object) <- value

\S4method{simProgressGraphical}{simList}(object) <- value

simProgressInterval(object)

\S4method{simProgressInterval}{simList}(object)

simProgressInterval(object) <- value

\S4method{simProgressInterval}{simList}(object) <- value

simGlobals(object)

\S4method{simGlobals}{simList}(object)

simGlobals(object) <- value

\S4method{simGlobals}{simList}(object) <- value

simGlobalsOutputPath(object)

\S4method{simGlobalsOutputPath}{simList}(object)

simGlobalsOutputPath(object) <- value

\S4method{simGlobalsOutputPath}{simList}(object) <- value

simEvents(object)

\S4method{simEvents}{simList}(object)

simEvents(object) <- value

\S4method{simEvents}{simList}(object) <- value

simCompleted(object)

\S4method{simCompleted}{simList}(object)

simCompleted(object) <- value

\S4method{simCompleted}{simList}(object) <- value

simDepends(object)

\S4method{simDepends}{simList}(object)

simDepends(object) <- value

\S4method{simDepends}{simList}(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Collection of accessor functions for the slots (and their elements) of a
\code{\link{simList}} object.
Currently, only get and set methods are defined. Subset methods are not.
See details below.
}
\section{1 Simulation parameters}{


Accessor functions for the \code{params} slot and its elements.
}

\section{1.1 Global parameters}{


\code{\link{simGlobals}} List of global simulation parameters.

\code{\link{simGlobalsOutputPath}} List of global simulation output path.
}

\section{1.2 Core module parameters}{


\code{\link{simCheckpointFile}} Name of the checkpoint file.

\code{\link{simCheckpointInterval}} The simulation checkpoint interval.

\code{\link{simObjectsLoaded}} List of loaded simulation objects.

\code{\link{simProgressGraphical}} Type of graphical progress bar used.

\code{\link{simProgressInterval}} Interval for the progress bar.
}

\section{1.3 User module parameters}{


\code{\link{simParams}} List of simulation parameters.
}

\section{2 Simulation modules and dependencies}{


Accessor functions for the \code{depends}, \code{modules}, \code{.loaded},
and \code{.loadOrder} slots.

\code{\link{simDepends}} List of simulation module dependencies.

\code{\link{simModules}} List of simulation modules to be loaded.

\code{\link{simModulesLoaded}} List of loaded simulation modules.

\code{\link{simModulesLoadOrder}} List specifying the order in which to load modules.
}

\section{3 Simulation event queues}{


Accessor functions for the \code{events} and \code{completed} slots.

\code{\link{simEvents}} Scheduled simulation events (the event queue).

\code{\link{simCompleted}} Completed simulation events.
}

\section{4 Simulation times}{


Accessor functions for the \code{simtimes} slot and its elements.

\code{\link{simTimes}} List of simulation times (current, start, stop).

\code{\link{simCurrentTime}} Current simulation time.

\code{\link{simStartTime}} Simulation start time.

\code{\link{simStopTime}} Simulation stop time.
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{simList-class}}
}

