% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList.R
\docType{class}
\name{simList-class}
\alias{initialize,simList-method}
\alias{ls.simList}
\alias{ls.str.simList}
\alias{show,simList-method}
\alias{simList}
\alias{simList-class}
\title{The \code{simList} class}
\usage{
\S4method{initialize}{simList}(.Object)

\S4method{show}{simList}(object)

ls.simList(x)

ls.str.simList(name)
}
\arguments{
\item{.Object}{An object.}

\item{object}{Any R object.}

\item{x}{A \code{simList} object.}

\item{x}{A \code{simList} object.}
}
\description{
Contains the minimum components of a SpaDES simulation.
}
\details{
Based on code from chapter 7.8.3 of Matloff (2011): "Discrete event simulation".
Here, we implement a discrete event simulation in a more modular fashion so it's
easier to add simulation components (i.e., "simulation modules").

We use S4 classes and methods, and use \code{\link{data.table}} instead of
\code{\link{data.frame}} to implement the event queue (because it is much faster).
}
\section{Slots}{

\describe{
\item{\code{.envir}}{Environment referencing the objects used in the simulation.}

\item{\code{.loadOrder}}{Character vector of names specifying the order in which modules are to be loaded.}

\item{\code{.loaded}}{List of character names specifying which modules and objects are currently loaded.}

\item{\code{modules}}{List of character names specifying which modules to load.}

\item{\code{params}}{Named list of potentially other lists specifying simulation parameters.}

\item{\code{events}}{The list of scheduled events (i.e., event queue), as a \code{data.table}.
                 This event queue is always sorted (keyed) by time,
                 making it easy to insert new events into the queue.}

\item{\code{completed}}{The list of completed events, as a \code{data.table}.}

\item{\code{depends}}{A \code{.simDeps} list of \code{.moduleDeps} objects containing
                 module object dependency information.}

\item{\code{simtimes}}{List of numerical values describing the simulation start and stop times;
                 as well as the current simulation time.}
}}
\note{
Each event is represented by a \code{\link{data.table}} row consisting of:
       \tabular{ll}{
         \code{eventTime} \tab The time the event is to occur.\cr
         \code{moduleName} \tab The module from which the event is taken.\cr
         \code{eventType} \tab A character string for the programmer-defined event type.\cr
       }
}
\author{
Alex Chubaty
}
\references{
Matloff, N. (2011). The Art of R Programming (ch. 7.8.3). San Fransisco, CA: No Starch Press, Inc.. Retrieved from \url{http://www.nostarch.com/artofr.htm}
}
\seealso{
\code{\link{data.table}}.
}

