% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{simList}
\alias{simList}
\alias{simList-class}
\title{The \code{simList} class}
\description{
This class contains the minimum components of a simulation.
}
\details{
Based on code from Matloff (2011) ch. 7.8.3 on discrete event simulation:
- implemented using S4 classes and methods;
- uses `data.table` instead of `data.frame` (which is much faster);
- implemented in a more modular fashion so it's easier
  to add submodules to the simulation.
}
\section{Slots}{

\describe{
\item{\code{.loaded}}{List of character names specifying which modules are currently loaded.}

\item{\code{modules}}{List of character names specifying which modules to load.}

\item{\code{params}}{Named list of potentially other lists specifying simulation parameters.}

\item{\code{events}}{The list of scheduled events, as a data.table class. This is implemented
                 such that the data.table is always sorted (keyed) by time, making it easy
                 to insert new events into the table.}

\item{\code{completed}}{The list of completed events, as a data.table class.}

\item{\code{simtimes}}{List of numerical values describing the simulation start and stop timos,
                 and the current simulation time.}

\item{\code{debug}}{Logical value specifying whether to run simulation in debugging mode.}
}}
\note{
Each event is represented by a data.table row consisting of:
         eventTime: the time the event is to occur;
         moduleName: the module from which the event is taken;
         eventType: a character string for the programmer-defined event type.
}
\author{
Alex Chubaty
}
\references{
Matloff, N. (2011). The Art of R Programming (373 pp.). San Fransisco, CA: No Starch Press, Inc.. Retrieved from \url{http://www.nostarch.com/artofr.htm}
}
\seealso{
\code{\link{data.table}}
}

