% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{simList-class}
\alias{simList-class}
\title{The \code{simList} class}
\description{
This class contains the minimum components of a simulation.
}
\details{
Based on code from chapter 7.8.3 of Matloff (2011): "Discrete event simulation".
Here, we implement a simulation in a more modular fashion so it's easier to add
submodules to the simulation. We use S4 classes and methods, and use `data.table`
instead of `data.frame` to implement the event queue (because it is much faster).
}
\section{Slots}{

\describe{
\item{\code{.loaded}}{List of character names specifying which modules and objects are currently loaded.}

\item{\code{modules}}{List of character names specifying which modules to load.}

\item{\code{params}}{Named list of potentially other lists specifying simulation parameters.}

\item{\code{events}}{The list of scheduled events, as a data.table class. This is implemented
                 such that the data.table is always sorted (keyed) by time, making it easy
                 to insert new events into the table.}

\item{\code{completed}}{The list of completed events, as a data.table class.}

\item{\code{simtimes}}{List of numerical values describing the simulation start and stop timos,
                 and the current simulation time.}
}}
\note{
Each event is represented by a data.table row consisting of:
         eventTime: the time the event is to occur;
         moduleName: the module from which the event is taken;
         eventType: a character string for the programmer-defined event type.
}
\author{
Alex Chubaty
}
\references{
Matloff, N. (2011). The Art of R Programming (ch. 7.8.3). San Fransisco, CA: No Starch Press, Inc.. Retrieved from \url{http://www.nostarch.com/artofr.htm}
}
\seealso{
\code{\link{data.table}}
}

