% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simList-accessors.R
\docType{methods}
\name{[[}
\alias{$}
\alias{$,simList-method}
\alias{$<-}
\alias{$<-,simList-method}
\alias{[[}
\alias{[[,simList,ANY,ANY-method}
\alias{[[<-}
\alias{[[<-,simList,ANY,ANY,ANY-method}
\title{Extract or replace parts of an object from the simulation environment}
\usage{
\S4method{[[}{simList,ANY,ANY}(x, i, j, ..., drop)

\S4method{[[}{simList,ANY,ANY,ANY}(x, i) <- value

\S4method{$}{simList}(x, name)

\S4method{$}{simList}(x, name) <- value
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s).}

\item{i}{indices specifying elements to extract or replace.}

\item{j}{see \code{i}.}

\item{...}{see \code{i}.}

\item{drop}{not implemented.}

\item{value}{Any R object.}

\item{name}{A literal character string or a \code{\link{name}}.}
}
\description{
Extract or replace parts of an object from the simulation environment
}

