% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{simModulesLoaded}
\alias{simModulesLoaded}
\alias{simModulesLoaded,simList-method}
\alias{simModulesLoaded<-}
\alias{simModulesLoaded<-,simList-method}
\title{Get and set the modules loaded for the simulation.}
\usage{
simModulesLoaded(object)

\S4method{simModulesLoaded}{simList}(object)

simModulesLoaded(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Currently, only get and set methods are defined. Subset methods are not.

get list of loaded simulation modules

set list of loaded simulation modules

set list of loaded simulation modules
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{simCheckpointFile}} for accessing the name of the checkpoint file;
         \code{\link{simCheckpointInterval}} for accessing the simulation checkpoint interval;
         \code{\link{simObjectsLoaded}} for accessing the list of loaded simulation objects;
         \code{\link{simGlobals}} for accessing the global simulation parameters;
         \code{\link{simGlobalsOutputPath}} for accessing the global simulation output path;
         \code{\link{simModules}} for accessing the list of simulation modules to be loaded;
         \code{\link{simParams}} for accessing the list of simulation parameters;
         \code{\link{simEvents}} for accessing the scheduled simulation event queue;
         \code{\link{simCompleted}} for accessing the list of completed simulation events;
         \code{\link{simTimes}} for accessing the list of simulation times
         (\code{simCurrentTime}, \code{simStartTime}, and \code{simStopTime}
         to access the simulation current, start, and stop times individually).
}

