% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{simPlot}
\alias{simPlot}
\alias{simPlot,RasterLayer-method}
\alias{simPlot,RasterStack-method}
\alias{simPlot,SpatialPoints-method}
\title{Plotting methods}
\usage{
simPlot(x, on.which.to.plot = 1, which.to.plot = "all",
  col = rev(terrain.colors(255)), visualSqueeze = 0.75, ..., add = FALSE,
  speedup = 1, axes = "L", add.legend = TRUE)

\S4method{simPlot}{RasterStack}(x, on.which.to.plot = 1,
  which.to.plot = "all", col = rev(terrain.colors(255)),
  visualSqueeze = 0.75, ..., add = FALSE, speedup = 1, axes = "L",
  add.legend = TRUE)

\S4method{simPlot}{RasterLayer}(x, on.which.to.plot, which.to.plot, col,
  visualSqueeze, delete.previous = TRUE, ..., add, speedup, axes,
  add.legend = TRUE)

\S4method{simPlot}{SpatialPoints}(x, on.which.to.plot = 1, which.to.plot,
  visualSqueeze = 0.75, delete.previous = TRUE, max.agents = 10000, ...,
  add = TRUE, speedup, axes, add.legend)
}
\arguments{
\item{x}{rasterStack, rasterLayer, or list of named rasters (not implemented yet), SpatialPoints* object.}

\item{on.which.to.plot}{when add=TRUE, numeric or character string identifying on which raster in existing plot window to plot. Used on \code{RasterLayer} and \code{pointAgent} and \code{mobileAgent}. Defaults to 1.}

\item{which.to.plot}{Numeric or character vector identifying which rasters in \code{rasterStack} to plot.}

\item{col}{a colour palette vector, possibly from RColorBrewer. Defaults to rev(terrain.colors(255))}

\item{visualSqueeze}{numeric index (from 0 to 1) that indicates how tightly the rasters should be plotted next
to each other. Default is 0.75, which allows for legends.}

\item{...}{Additional plotting functions passed to grid.raster (if rasterStack) or grid.points (if pointAgent)}

\item{add}{Logical indicating whether to plot new maps (\code{FALSE}) or update exising maps (\code{TRUE}).
Default is \code{FALSE} for rasters and \code{TRUE} for agents.}

\item{speedup}{Scalar indicating how much faster than normal to make plots (see Details). Defaults to 1.}

\item{axes}{String either "all", "L", or "none" (see Details). Default is "L".}

\item{ext}{an extent object to describe the size of the map that is being plotted on}

\item{delete.previous}{should the immediately previously simPlotted object be removed before adding current simPlot call}

\item{max.agents}{is the maximum number of agents to plot. \code{speedup} reduces the number plotted from this max.agents}
}
\value{
Creates a plot within the active plotting device.
}
\description{
Plotting for RasterStack, Agents and other classes used in simulations.
}
\details{
For class \code{RasterStack}, this method plots many rasters. This method automatically plots individual
rasters in with an arrangement on the plotting window that is optimal for the size and shape of
the window. \code{Speedup} will make downsample the number of pixels, allowing for greater plotting
speed, at a cost of more fuzzy plots.
}
\seealso{
\code{\link{grid.raster}}
}

