% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{simStartTime}
\alias{simStartTime}
\alias{simStartTime,simList-method}
\alias{simStartTime<-}
\alias{simStartTime<-,simList-method}
\title{Get and set the simulation start time.}
\usage{
simStartTime(object)

\S4method{simStartTime}{simList}(object)

simStartTime(object) <- value
}
\arguments{
\item{object}{A \code{simList} simulation object.}

\item{value}{The object to be stored at the slot.}
}
\value{
Returns or sets the value of the slot from the \code{simList} object.
}
\description{
Get and set the simulation start time.

get the simulation start time

set the simulation start time

set the simulation start time
}
\author{
Alex Chubaty
}
\seealso{
\code{\link{simCheckpointFile}} for accessing the name of the checkpoint file;
         \code{\link{simCheckpointInterval}} for accessing the simulation checkpoint interval;
         \code{\link{simGlobals}} for accessing the global simulation parameters;
         \code{\link{simGlobalsOutputPath}} for accessing the global simulation output path;
         \code{\link{simModulesLoaded}} for accessing the list of loaded simulation modules;
         \code{\link{simObjectsLoaded}} for accessing the list of loaded simulation objects;
         \code{\link{simModules}} for accessing the list of simulation modules to be loaded;
         \code{\link{simParams}} for accessing the list of simulation parameters;
         \code{\link{simEvents}} for accessing the scheduled simulation event queue;
         \code{\link{simCompleted}} for accessing the list of completed simulation events;
         \code{\link{simTimes}} for accessing the list of simulation times
         (\code{simCurrentTime} and \code{simStopTime}
         to access the simulation current and stop times individually).
}

