% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{simTimes,simList-method}
\alias{<-}
\alias{simTimes,simList-method}
\alias{simTimes<-}
\title{get list of simulation times}
\usage{
\S4method{simTimes}{simList}(object)

simTimes(object) <- value
}
\description{
get list of simulation times

set list of simulation times

set list of simulation times
}

