% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades-classes.R
\name{spadesClasses}
\alias{spadesClasses}
\title{Classes defined in SpaDES}
\description{
These S4 classes are defined within \code{SpaDES}. "dot" classes are not exported and
are therefore intended for internal use only.
}
\section{Simulation classes}{

\tabular{ll}{
  \code{\link{simList}} \tab The 'simList' class\cr
  \code{\link{.moduleDeps}} \tab Descriptor object for specifying \code{SpaDES} module dependencies\cr
  \code{\link{.simDeps}} \tab Defines all simulation dependencies for all modules within a \code{SpaDES} simulation\cr
  --------------------------- \tab ------------------------------------------------------------------------------------------ \cr
}
}

\seealso{
\code{\link{simInit}}
}
\author{
Eliot McIntire and Alex Chubaty
}
