% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spades-package.R
\docType{data}
\name{SpaDES-maps}
\alias{SpaDES-maps}
\alias{spades-maps}
\title{Dummy maps included with SpaDES}
\format{raster}
\description{
All maps included with SpaDES are randomly generated maps created by \code{gaussMap()}.
These are located within the \code{maps} folder of the package, and are used in the vignettes.
Use \code{system.file(package="SpaDES", "maps")} to locate the \code{maps} directory on your system.
}
\details{
\code{DEM.tif}: converted to a a small number of discrete levels (in 100m hypothetical units)

\code{habitatQuality.tif}: made to look like a continuous habitat surface, rescaled to 0 to 1

\code{forestAge.tif}: rescaled to possible forest ages in a boreal forest setting

\code{forestCover.tif}: rescaled to possible forest cover in a boreal forest setting

\code{percentPine.tif}: rescaled to percentages.
}
\keyword{maps}

