% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spades-package.R
\docType{package}
\name{SpaDES-package}
\alias{SpaDES}
\alias{SpaDES-package}
\alias{spades-package}
\title{Develop and run spatially explicit discrete event simulation models}
\description{
Easily implement a variety of simulation models, with a focus on spatially explicit agent based models. The core simulation components are built upon a discrete event simulation framework that facilitates modularity, and easily enables the user to include additional functionality by running user-built simulation modules. Included are numerous tools to visualize raster and other maps.

Bug reports: https://github.com/achubaty/SpaDES/issues
}
\author{
Alex M. Chubaty \email{Alexander.Chubaty@NRCan.gc.ca} and Eliot J. B. McIntire \email{Eliot.McIntire@NRCan.gc.ca}.
}
\keyword{package}

