% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\docType{data}
\name{.spadesEnv}
\alias{.spadesEnv}
\title{The SpaDES environment}
\format{\preformatted{<environment: 0x000000000a40ead8> 
}}
\usage{
.spadesEnv
}
\description{
Environment used internally to store internal package objects and methods.
}
\keyword{datasets}

