% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\docType{data}
\name{.spadesEnv}
\alias{.spadesEnv}
\title{The SpaDES environment}
<<<<<<< HEAD
\format{\preformatted{<environment: 0x7fa23e35e980> 
=======
\format{\preformatted{<environment: 0x000000001f350238> 
>>>>>>> origin/development
}}
\usage{
.spadesEnv
}
\description{
Environment used internally to store internal package objects and methods.
}
\keyword{datasets}

