% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/environment.R
\docType{data}
\name{.spadesEnv}
\alias{.spadesEnv}
\title{The SpaDES environment}
\format{\preformatted{<environment: 0x0000000014d85410> 
}}
\usage{
.spadesEnv
}
\description{
needs description
}
\keyword{datasets}

