% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{.spadesPlotObjects}
\alias{.spadesPlotObjects}
\title{The \code{.spadesPlotObjects} class}
\description{
This class contains the union of .spatialObjects and several other plot-type objects.
These are the object classes that the \code{\link{Plot}} function can handle.
}
\section{Slots}{

\describe{
\item{\code{members}}{SpatialPoints*, SpatialPolygons*, SpatialLines*, RasterLayer, RasterStack}
}}
\author{
Eliot McIntire
}
\seealso{
\code{\link{spadesClasses}}
}

