% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{spadesPlotObjects-class}
\alias{spadesPlotObjects}
\alias{spadesPlotObjects-class}
\title{The \code{spadesPlotObjects} class}
\description{
This class contains the plotting arrangement information.

This class contains the plotting arrangement information.
}
\section{Slots}{

\describe{
\item{\code{members}}{SpatialPoints*, SpatialPolygons*, SpatialLines*, RasterLayer, RasterStack}

\item{\code{members}}{SpatialPoints*, SpatialPolygons*, SpatialLines*, RasterLayer, RasterStack}
}}
\author{
Eliot McIntire

Eliot McIntire
}

