% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-classes.R
\docType{class}
\name{.spadesPlottables-class}
\alias{.spadesPlottables-class}
\alias{.spadesPlottables}
\title{The \code{.spadesPlottables} class}
\description{
This class is the union of all .spadesPlotObjects (e.g., \code{RasterLayer*},
\code{SpatialPoints*}, \code{SpatialPolygons*}, \code{ggplot}, \code{hist}, etc.)
and \code{\link{.spadesPlot}} class objects.
This allows replotting of a \code{\link{.spadesPlot}} object.
}
\section{Slots}{

\describe{
\item{\code{members}}{\code{\link{.spadesPlotObjects}} and \code{\link{.spadesPlot}}}
}}

\seealso{
\code{\link{spadesClasses}}
}
\author{
Eliot McIntire
}
\keyword{internal}
