% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{.spadesPlottables}
\alias{.spadesPlottables}
\title{The \code{.spadesPlottables} class}
\description{
This class is the union of all .spadesPlotObjects (e.g., RasterLayer*,
SpatialPoints*, SpatialPolygons*, ggplot, hist etc.) and \code{\link{.spadesPlot}}
class objects.  This allows replotting of a \code{\link{.spadesPlot}} object
}
\section{Slots}{

\describe{
\item{\code{members}}{\code{\link{.spadesPlotObjects}} and \code{\link{.spadesPlot}}}
}}
\section{All non-exported object classes used with \code{\link{Plot}}}{


\tabular{ll}{
  New classes\tab \cr
  \code{\link{.spadesPlot}} \tab Main class for \code{\link{Plot}} - contains .spadesGrob and .arrangement objects\cr
  \code{\link{.spadesGrob}} \tab GRaphical OBject used by SpaDES - smallest unit\cr
  \code{\link{.arrangement}} \tab The layout or "arrangement" of plot objects\cr
  --------------------------- \tab ------------------------------------------------------------------------------------------ \cr
}
\tabular{ll}{
  Unions of existing classes\tab \cr
  \code{\link{.spadesPlottables}} \tab The union of all object classes Plot can accept\cr
  \code{\link{.spadesPlotObjects}} \tab The union of .spatialObjects and several others\cr
  \code{\link{.spatialObjects}} \tab The union of several spatial classes\cr
  --------------------------- \tab ------------------------------------------------------------------------------------------ \cr
}
}
\author{
Eliot McIntire
}

