% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{dyears}
\alias{dNA}
\alias{dNA,ANY-method}
<<<<<<< HEAD
\alias{dday}
\alias{dhour}
\alias{dmonth}
\alias{dmonths}
\alias{dmonths,numeric-method}
\alias{dsecond}
\alias{dweek}
\alias{dweeks}
\alias{dweeks,numeric-method}
\alias{dyear}
=======
\alias{dmonths}
\alias{dmonths,numeric-method}
\alias{dweek}
\alias{dweeks}
\alias{dweeks,numeric-method}
>>>>>>> origin/development
\alias{dyears}
\alias{dyears,numeric-method}
\title{New time units}
\usage{
dyears(x)

\S4method{dyears}{numeric}(x)

dmonths(x)

\S4method{dmonths}{numeric}(x)

dweeks(x)

\S4method{dweeks}{numeric}(x)

<<<<<<< HEAD
dweek(x)

dmonth(x)

dyear(x)

dsecond(x)

dday(x)

dhour(x)

=======
>>>>>>> origin/development
dNA(x)

\S4method{dNA}{ANY}(x)
}
\arguments{
\item{x}{numeric. Number of the desired units}
}
\value{
Number of seconds within each unit
}
\description{
SpaDES commonly needs generic durations, like "year" which will round neatly over
a century (i.e., leap years are added within each year with an extra 1/4 day,
i.e., year=365.25 days), months are defined as year/12, weeks as year/52. This
year is also known as the "astronomical" or "Julian" year.
}
\author{
Eliot McIntire
}

