% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SpatialObjects-class}
\alias{SpatialObjects-class}
\title{The \code{SpatialObjects} class}
\description{
This class contains the plotting arrangement information.
}
\section{Slots}{

\describe{
\item{\code{members}}{SpatialPoints*, SpatialPolygons*, RasterLayer, and RasterStack}
}}
\author{
Eliot McIntire
}

