% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-classes.R
\docType{class}
\name{spatialObjects-class}
\alias{spatialObjects-class}
\alias{spatialObjects}
\title{The \code{spatialObjects} class}
\description{
This class is the union of several spatial objects from \code{raster} and
\code{sp} packages:
}
\details{
\itemize{
  \item \code{RasterLayer}, \code{RasterLayerSparse}, \code{RasterStack};
  \item \code{SpatialLines}, \code{SpatialLinesDataFrame};
  \item \code{SpatialPixels}, \code{SpatialPixelsDataFrame};
  \item \code{SpatialPoints}, \code{SpatialPointsDataFrame};
  \item \code{SpatialPolygons}, \code{SpatialPolygonsDataFrame}.
}

Notably missing is \code{RasterBrick}, for now.
}
\seealso{
\code{\link{spadesClasses}}
}
\author{
Eliot McIntire
}
