% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spatialObjects-class}
\alias{spatialObjects-class}
\title{The \code{spatialObjects} class}
\description{
This class contains the plotting arrangement information.
}
\section{Slots}{

\describe{
\item{\code{members}}{SpatialPoints*, SpatialPolygons*, RasterLayer, and RasterStack}
}}
\author{
Eliot McIntire
}

