% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{spatialObjects-class}
\alias{spatialObjects-class}
\title{The \code{spatialObjects} class}
\description{
This class contains the plotting arrangement information.
}
\section{Slots}{

\describe{
\item{\code{members}}{SpatialPoints*, SpatialPolygons*, SpatialLines*, RasterLayer, RasterStack}
}}
\author{
Eliot McIntire
}

