% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{.spatialObjects-class}
\alias{.spatialObjects}
\alias{.spatialObjects-class}
\title{The \code{.spatialObjects} class}
\description{
This class is the union of several spatial objects from raster and sp packages. Notably
missing is RasterBrick, for now.
}
\section{Slots}{

\describe{
\item{\code{members}}{SpatialPoints*, SpatialPolygons*, SpatialLines*,
               RasterLayer, RasterStack}
}}
\author{
Eliot McIntire
}
\seealso{
\code{\link{spadesClasses}}
}

