% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-classes.R
\name{spatialObjects-class}
\alias{spatialObjects}
\alias{spatialObjects-class}
\title{The \code{spatialObjects} class}
\description{
This class is the union of several spatial objects from raster and sp packages.
Notably missing is \code{RasterBrick}, for now.
}
\section{Slots}{

\describe{
\item{\code{members}}{SpatialPoints*, SpatialPolygons*, SpatialLines*,
               RasterLayer, RasterStack

#@importClassesFrom NetLogoRClasses agentMatrix
#@importClassesFrom NetLogoRClasses NLworldMatrix}
}}
\author{
Eliot McIntire
}
\seealso{
\code{\link{spadesClasses}}
}

