% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{SpatialPointsDataFrameNamed}
\alias{SpatialPointsDataFrameNamed}
\alias{SpatialPointsDataFrameNamed,character-method}
\title{Create a new \code{SpatialPointsDataFrameNamed} object.}
\usage{
SpatialPointsDataFrameNamed(..., name)

\S4method{SpatialPointsDataFrameNamed}{character}(..., name)
}
\arguments{
\item{...}{Additional arguments to \code{SpatialPointsDataFrame}}

\item{name}{The name of the object.}
}
\value{
Returns an object of class \code{SpatialPointsDataFrameNamed}.
}
\description{
Create a new \code{SpatialPointsDataFrameNamed} object.
}
\author{
Eliot McIntire

Alex Chubaty
}
\seealso{
\code{\link{SpatialPointsDataFrame}}
}

