% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{spreadAgent}
\alias{spreadAgent}
\alias{spreadAgent-class}
\title{The \code{spreadAgent} class.}
\description{
Extends the \code{rasterAgent} class. Describes spatial agents whose
spatial component ("area") is characterized by a raster and are capable
of spreading (growing/shrinking).
}
\section{Slots}{

\describe{
\item{\code{ID}}{A character string that identifies the individual agent.
             By default, the numbers 1...n are used.}

\item{\code{NumPix}}{A numeric object describing the number of raster pixels occupied.}

\item{\code{other}}{A list for storing other (user-specific) agent attributes.}
}}
\note{
add additional notes here.
}

