% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/times.R
\docType{methods}
\name{inSeconds}
\alias{inSeconds}
\alias{inSeconds,NULL-method}
\alias{inSeconds,character-method}
\title{Convert time units}
\usage{
inSeconds(unit)

\S4method{inSeconds}{character}(unit)

\S4method{inSeconds}{`NULL`}(unit)
}
\arguments{
\item{x}{character indicating time units to be returned as seconds}
}
\description{
In addition to using the \code{lubridate} package, some additional functions and
to work with times are provided.
}
\author{
Alex Chubaty & Eliot McIntire
}

