% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/times.R
\docType{methods}
\name{inSecs}
\alias{inSecs}
\alias{inSecs,character-method}
\title{Convert time units}
\usage{
inSecs(unit)

\S4method{inSecs}{character}(unit)
}
\arguments{
\item{x}{character indicating time units to be returned as seconds}
}
\description{
In addition to using the \code{lubridate} package, some additional functions and
to work with times are provided.
}

