% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{.unittrim}
\alias{.unittrim}
\title{Convert grid.locator units}
\usage{
.unittrim(grid.locator)
}
\arguments{
\item{grid.locator}{an object that was output by a call to grid.locator and mouse click(s)}
}
\description{
Converts them to meaningful units. Used within \code{.clickCoord}
}
\author{
Eliot McIntire
}

