% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{updateSpadesPlot}
\alias{updateSpadesPlot}
\alias{updateSpadesPlot,spadesPlot,spadesPlot-method}
\title{Merge two SpaDES Plot objects}
\usage{
updateSpadesPlot(newSpadesPlot, existingSpadesPlot, ...)

\S4method{updateSpadesPlot}{spadesPlot,spadesPlot}(newSpadesPlot,
  existingSpadesPlot, ...)
}
\arguments{
\item{newSpadesPlot}{spadesPlot object.}

\item{existingSpadesPlot}{spadesPlot object.}

\item{...}{additional arguments. Currently nothing.}
}
\description{
Merges two \code{spadesPlot} objects
}

