% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{.updateSpadesPlot}
\alias{.updateSpadesPlot}
\alias{.updateSpadesPlot,.spadesPlot,.spadesPlot-method}
\alias{.updateSpadesPlot,.spadesPlot,ANY-method}
\title{Merge two SpaDES Plot objects}
\usage{
.updateSpadesPlot(newSP, curr, ...)

\S4method{.updateSpadesPlot}{.spadesPlot,.spadesPlot}(newSP, curr, ...)

\S4method{.updateSpadesPlot}{.spadesPlot,ANY}(newSP, curr, ...)
}
\arguments{
\item{newSP}{"new" .spadesPlot object. i.e., the new merges and overwrites into current}

\item{curr}{"current" .spadesPlot object. i.e., the one to be merged into.}

\item{...}{additional arguments. Currently nothing.}
}
\description{
Merges two \code{.spadesPlot} objects
}
\author{
Eliot McIntire
}

