% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{vectorAgent}
\alias{vectorAgent}
\alias{vectorAgent-class}
\title{The \code{vectorAgent} class.}
\description{
Extends the \code{agent} class by making it spatial.
Describes spatial agents whose  spatial components are characterized by
points or a polygon (i.e., in contrast to a raster).
}
\details{
This is a placeholder class and not intended to be used directly.
}
\section{Slots}{

\describe{
\item{\code{ID}}{A character string that identifies the individual agent.
             By default, the numbers 1...n are used.}

\item{\code{other}}{A list for storing other (user-specific) agent attributes.}
}}
\note{
This is a virtual class and not intended to be used directly.
}

