% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module-template.R
\docType{methods}
\name{zipModule}
\alias{zipModule}
\alias{zipModule,character,character,character-method}
\alias{zipModule,character,missing,character-method}
\alias{zipModule,character,missing,missing-method}
\title{Create a zip archive of a module subdirectory}
\usage{
zipModule(name, path, version)

\S4method{zipModule}{character,character,character}(name, path, version)

\S4method{zipModule}{character,missing,character}(name, version)

\S4method{zipModule}{character,missing,missing}(name)
}
\arguments{
\item{name}{Character string giving the module name.}

\item{path}{A file path to a directory containing the module subdirectory.}

\item{version}{The module version}
}
\description{
The most common use of this would be from a "modules" directory, rather than
inside a given module.
}
\author{
Eliot McIntire and Alex Chubaty
}

